@@ -1,11 +1,10 @@
 name: flutter_icons
 description: Customizable Icons for Flutter，you can use with over 3K+ icons in your flutter project
-version: 1.1.0
-author: flutter-studio<2534290808@qq.com>
-homepage: https://github.com/flutter-studio/flutter-icons.git
+version: 2.0.0
+homepage: https://github.com/flutter-studio/flutter-icons
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -14,6 +13,8 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
+
+  lints: ^1.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
