@@ -6,463 +6,427 @@ import 'flutter_icon_data.dart';
 /// http://entypo.com
 class Entypo {
   Entypo._();
-  static const IconData $500px = const FlutterIconData.entypo(61696);
-  static const IconData $500px_with_circle =
-      const FlutterIconData.entypo(61697);
-  static const IconData add_to_list = const FlutterIconData.entypo(61698);
-  static const IconData add_user = const FlutterIconData.entypo(61699);
-  static const IconData address = const FlutterIconData.entypo(61700);
-  static const IconData adjust = const FlutterIconData.entypo(61701);
-  static const IconData air = const FlutterIconData.entypo(61702);
-  static const IconData aircraft = const FlutterIconData.entypo(61703);
-  static const IconData aircraft_landing = const FlutterIconData.entypo(61704);
-  static const IconData aircraft_take_off = const FlutterIconData.entypo(61705);
-  static const IconData align_bottom = const FlutterIconData.entypo(61706);
-  static const IconData align_horizontal_middle =
-      const FlutterIconData.entypo(61707);
-  static const IconData align_left = const FlutterIconData.entypo(61708);
-  static const IconData align_right = const FlutterIconData.entypo(61709);
-  static const IconData align_top = const FlutterIconData.entypo(61710);
-  static const IconData align_vertical_middle =
-      const FlutterIconData.entypo(61711);
-  static const IconData app_store = const FlutterIconData.entypo(61712);
-  static const IconData archive = const FlutterIconData.entypo(61713);
-  static const IconData area_graph = const FlutterIconData.entypo(61714);
-  static const IconData arrow_bold_down = const FlutterIconData.entypo(61715);
-  static const IconData arrow_bold_left = const FlutterIconData.entypo(61716);
-  static const IconData arrow_bold_right = const FlutterIconData.entypo(61717);
-  static const IconData arrow_bold_up = const FlutterIconData.entypo(61718);
-  static const IconData arrow_down = const FlutterIconData.entypo(61719);
-  static const IconData arrow_left = const FlutterIconData.entypo(61720);
-  static const IconData arrow_long_down = const FlutterIconData.entypo(61721);
-  static const IconData arrow_long_left = const FlutterIconData.entypo(61722);
-  static const IconData arrow_long_right = const FlutterIconData.entypo(61723);
-  static const IconData arrow_long_up = const FlutterIconData.entypo(61724);
-  static const IconData arrow_right = const FlutterIconData.entypo(61725);
-  static const IconData arrow_up = const FlutterIconData.entypo(61726);
-  static const IconData arrow_with_circle_down =
-      const FlutterIconData.entypo(61727);
-  static const IconData arrow_with_circle_left =
-      const FlutterIconData.entypo(61728);
-  static const IconData arrow_with_circle_right =
-      const FlutterIconData.entypo(61729);
-  static const IconData arrow_with_circle_up =
-      const FlutterIconData.entypo(61730);
-  static const IconData attachment = const FlutterIconData.entypo(61731);
-  static const IconData awareness_ribbon = const FlutterIconData.entypo(61732);
-  static const IconData back = const FlutterIconData.entypo(61733);
-  static const IconData back_in_time = const FlutterIconData.entypo(61734);
-  static const IconData baidu = const FlutterIconData.entypo(61735);
-  static const IconData bar_graph = const FlutterIconData.entypo(61736);
-  static const IconData basecamp = const FlutterIconData.entypo(61737);
-  static const IconData battery = const FlutterIconData.entypo(61738);
-  static const IconData beamed_note = const FlutterIconData.entypo(61739);
-  static const IconData behance = const FlutterIconData.entypo(61740);
-  static const IconData bell = const FlutterIconData.entypo(61741);
-  static const IconData blackboard = const FlutterIconData.entypo(61742);
-  static const IconData block = const FlutterIconData.entypo(61743);
-  static const IconData book = const FlutterIconData.entypo(61744);
-  static const IconData bookmark = const FlutterIconData.entypo(61745);
-  static const IconData bookmarks = const FlutterIconData.entypo(61746);
-  static const IconData bowl = const FlutterIconData.entypo(61747);
-  static const IconData box = const FlutterIconData.entypo(61748);
-  static const IconData briefcase = const FlutterIconData.entypo(61749);
-  static const IconData browser = const FlutterIconData.entypo(61750);
-  static const IconData brush = const FlutterIconData.entypo(61751);
-  static const IconData bucket = const FlutterIconData.entypo(61752);
-  static const IconData bug = const FlutterIconData.entypo(61753);
-  static const IconData cake = const FlutterIconData.entypo(61754);
-  static const IconData calculator = const FlutterIconData.entypo(61755);
-  static const IconData calendar = const FlutterIconData.entypo(61756);
-  static const IconData camera = const FlutterIconData.entypo(61757);
-  static const IconData ccw = const FlutterIconData.entypo(61758);
-  static const IconData chat = const FlutterIconData.entypo(61759);
-  static const IconData check = const FlutterIconData.entypo(61760);
-  static const IconData chevron_down = const FlutterIconData.entypo(61761);
-  static const IconData chevron_left = const FlutterIconData.entypo(61762);
-  static const IconData chevron_right = const FlutterIconData.entypo(61763);
-  static const IconData chevron_small_down =
-      const FlutterIconData.entypo(61764);
-  static const IconData chevron_small_left =
-      const FlutterIconData.entypo(61765);
-  static const IconData chevron_small_right =
-      const FlutterIconData.entypo(61766);
-  static const IconData chevron_small_up = const FlutterIconData.entypo(61767);
-  static const IconData chevron_thin_down = const FlutterIconData.entypo(61768);
-  static const IconData chevron_thin_left = const FlutterIconData.entypo(61769);
-  static const IconData chevron_thin_right =
-      const FlutterIconData.entypo(61770);
-  static const IconData chevron_thin_up = const FlutterIconData.entypo(61771);
-  static const IconData chevron_up = const FlutterIconData.entypo(61772);
+
+  static const IconData $500px = FlutterIconData.entypo(61696);
+  static const IconData $500px_with_circle = FlutterIconData.entypo(61697);
+  static const IconData add_to_list = FlutterIconData.entypo(61698);
+  static const IconData add_user = FlutterIconData.entypo(61699);
+  static const IconData address = FlutterIconData.entypo(61700);
+  static const IconData adjust = FlutterIconData.entypo(61701);
+  static const IconData air = FlutterIconData.entypo(61702);
+  static const IconData aircraft = FlutterIconData.entypo(61703);
+  static const IconData aircraft_landing = FlutterIconData.entypo(61704);
+  static const IconData aircraft_take_off = FlutterIconData.entypo(61705);
+  static const IconData align_bottom = FlutterIconData.entypo(61706);
+  static const IconData align_horizontal_middle = FlutterIconData.entypo(61707);
+  static const IconData align_left = FlutterIconData.entypo(61708);
+  static const IconData align_right = FlutterIconData.entypo(61709);
+  static const IconData align_top = FlutterIconData.entypo(61710);
+  static const IconData align_vertical_middle = FlutterIconData.entypo(61711);
+  static const IconData app_store = FlutterIconData.entypo(61712);
+  static const IconData archive = FlutterIconData.entypo(61713);
+  static const IconData area_graph = FlutterIconData.entypo(61714);
+  static const IconData arrow_bold_down = FlutterIconData.entypo(61715);
+  static const IconData arrow_bold_left = FlutterIconData.entypo(61716);
+  static const IconData arrow_bold_right = FlutterIconData.entypo(61717);
+  static const IconData arrow_bold_up = FlutterIconData.entypo(61718);
+  static const IconData arrow_down = FlutterIconData.entypo(61719);
+  static const IconData arrow_left = FlutterIconData.entypo(61720);
+  static const IconData arrow_long_down = FlutterIconData.entypo(61721);
+  static const IconData arrow_long_left = FlutterIconData.entypo(61722);
+  static const IconData arrow_long_right = FlutterIconData.entypo(61723);
+  static const IconData arrow_long_up = FlutterIconData.entypo(61724);
+  static const IconData arrow_right = FlutterIconData.entypo(61725);
+  static const IconData arrow_up = FlutterIconData.entypo(61726);
+  static const IconData arrow_with_circle_down = FlutterIconData.entypo(61727);
+  static const IconData arrow_with_circle_left = FlutterIconData.entypo(61728);
+  static const IconData arrow_with_circle_right = FlutterIconData.entypo(61729);
+  static const IconData arrow_with_circle_up = FlutterIconData.entypo(61730);
+  static const IconData attachment = FlutterIconData.entypo(61731);
+  static const IconData awareness_ribbon = FlutterIconData.entypo(61732);
+  static const IconData back = FlutterIconData.entypo(61733);
+  static const IconData back_in_time = FlutterIconData.entypo(61734);
+  static const IconData baidu = FlutterIconData.entypo(61735);
+  static const IconData bar_graph = FlutterIconData.entypo(61736);
+  static const IconData basecamp = FlutterIconData.entypo(61737);
+  static const IconData battery = FlutterIconData.entypo(61738);
+  static const IconData beamed_note = FlutterIconData.entypo(61739);
+  static const IconData behance = FlutterIconData.entypo(61740);
+  static const IconData bell = FlutterIconData.entypo(61741);
+  static const IconData blackboard = FlutterIconData.entypo(61742);
+  static const IconData block = FlutterIconData.entypo(61743);
+  static const IconData book = FlutterIconData.entypo(61744);
+  static const IconData bookmark = FlutterIconData.entypo(61745);
+  static const IconData bookmarks = FlutterIconData.entypo(61746);
+  static const IconData bowl = FlutterIconData.entypo(61747);
+  static const IconData box = FlutterIconData.entypo(61748);
+  static const IconData briefcase = FlutterIconData.entypo(61749);
+  static const IconData browser = FlutterIconData.entypo(61750);
+  static const IconData brush = FlutterIconData.entypo(61751);
+  static const IconData bucket = FlutterIconData.entypo(61752);
+  static const IconData bug = FlutterIconData.entypo(61753);
+  static const IconData cake = FlutterIconData.entypo(61754);
+  static const IconData calculator = FlutterIconData.entypo(61755);
+  static const IconData calendar = FlutterIconData.entypo(61756);
+  static const IconData camera = FlutterIconData.entypo(61757);
+  static const IconData ccw = FlutterIconData.entypo(61758);
+  static const IconData chat = FlutterIconData.entypo(61759);
+  static const IconData check = FlutterIconData.entypo(61760);
+  static const IconData chevron_down = FlutterIconData.entypo(61761);
+  static const IconData chevron_left = FlutterIconData.entypo(61762);
+  static const IconData chevron_right = FlutterIconData.entypo(61763);
+  static const IconData chevron_small_down = FlutterIconData.entypo(61764);
+  static const IconData chevron_small_left = FlutterIconData.entypo(61765);
+  static const IconData chevron_small_right = FlutterIconData.entypo(61766);
+  static const IconData chevron_small_up = FlutterIconData.entypo(61767);
+  static const IconData chevron_thin_down = FlutterIconData.entypo(61768);
+  static const IconData chevron_thin_left = FlutterIconData.entypo(61769);
+  static const IconData chevron_thin_right = FlutterIconData.entypo(61770);
+  static const IconData chevron_thin_up = FlutterIconData.entypo(61771);
+  static const IconData chevron_up = FlutterIconData.entypo(61772);
   static const IconData chevron_with_circle_down =
-      const FlutterIconData.entypo(61773);
+      FlutterIconData.entypo(61773);
   static const IconData chevron_with_circle_left =
-      const FlutterIconData.entypo(61774);
+      FlutterIconData.entypo(61774);
   static const IconData chevron_with_circle_right =
-      const FlutterIconData.entypo(61775);
-  static const IconData chevron_with_circle_up =
-      const FlutterIconData.entypo(61776);
-  static const IconData circle = const FlutterIconData.entypo(61777);
-  static const IconData circle_with_cross = const FlutterIconData.entypo(61778);
-  static const IconData circle_with_minus = const FlutterIconData.entypo(61779);
-  static const IconData circle_with_plus = const FlutterIconData.entypo(61780);
-  static const IconData circular_graph = const FlutterIconData.entypo(61781);
-  static const IconData clapperboard = const FlutterIconData.entypo(61782);
-  static const IconData classic_computer = const FlutterIconData.entypo(61783);
-  static const IconData clipboard = const FlutterIconData.entypo(61784);
-  static const IconData clock = const FlutterIconData.entypo(61785);
-  static const IconData cloud = const FlutterIconData.entypo(61786);
-  static const IconData code = const FlutterIconData.entypo(61787);
-  static const IconData cog = const FlutterIconData.entypo(61788);
-  static const IconData colours = const FlutterIconData.entypo(61789);
-  static const IconData compass = const FlutterIconData.entypo(61790);
+      FlutterIconData.entypo(61775);
+  static const IconData chevron_with_circle_up = FlutterIconData.entypo(61776);
+  static const IconData circle = FlutterIconData.entypo(61777);
+  static const IconData circle_with_cross = FlutterIconData.entypo(61778);
+  static const IconData circle_with_minus = FlutterIconData.entypo(61779);
+  static const IconData circle_with_plus = FlutterIconData.entypo(61780);
+  static const IconData circular_graph = FlutterIconData.entypo(61781);
+  static const IconData clapperboard = FlutterIconData.entypo(61782);
+  static const IconData classic_computer = FlutterIconData.entypo(61783);
+  static const IconData clipboard = FlutterIconData.entypo(61784);
+  static const IconData clock = FlutterIconData.entypo(61785);
+  static const IconData cloud = FlutterIconData.entypo(61786);
+  static const IconData code = FlutterIconData.entypo(61787);
+  static const IconData cog = FlutterIconData.entypo(61788);
+  static const IconData colours = FlutterIconData.entypo(61789);
+  static const IconData compass = FlutterIconData.entypo(61790);
   static const IconData controller_fast_backward =
-      const FlutterIconData.entypo(61791);
-  static const IconData controller_fast_forward =
-      const FlutterIconData.entypo(61792);
+      FlutterIconData.entypo(61791);
+  static const IconData controller_fast_forward = FlutterIconData.entypo(61792);
   static const IconData controller_jump_to_start =
-      const FlutterIconData.entypo(61793);
-  static const IconData controller_next = const FlutterIconData.entypo(61794);
-  static const IconData controller_paus = const FlutterIconData.entypo(61795);
-  static const IconData controller_play = const FlutterIconData.entypo(61796);
-  static const IconData controller_record = const FlutterIconData.entypo(61797);
-  static const IconData controller_stop = const FlutterIconData.entypo(61798);
-  static const IconData controller_volume = const FlutterIconData.entypo(61799);
-  static const IconData copy = const FlutterIconData.entypo(61800);
-  static const IconData creative_cloud = const FlutterIconData.entypo(61801);
-  static const IconData creative_commons = const FlutterIconData.entypo(61802);
+      FlutterIconData.entypo(61793);
+  static const IconData controller_next = FlutterIconData.entypo(61794);
+  static const IconData controller_paus = FlutterIconData.entypo(61795);
+  static const IconData controller_play = FlutterIconData.entypo(61796);
+  static const IconData controller_record = FlutterIconData.entypo(61797);
+  static const IconData controller_stop = FlutterIconData.entypo(61798);
+  static const IconData controller_volume = FlutterIconData.entypo(61799);
+  static const IconData copy = FlutterIconData.entypo(61800);
+  static const IconData creative_cloud = FlutterIconData.entypo(61801);
+  static const IconData creative_commons = FlutterIconData.entypo(61802);
   static const IconData creative_commons_attribution =
-      const FlutterIconData.entypo(61803);
+      FlutterIconData.entypo(61803);
   static const IconData creative_commons_noderivs =
-      const FlutterIconData.entypo(61804);
+      FlutterIconData.entypo(61804);
   static const IconData creative_commons_noncommercial_eu =
-      const FlutterIconData.entypo(61805);
+      FlutterIconData.entypo(61805);
   static const IconData creative_commons_noncommercial_us =
-      const FlutterIconData.entypo(61806);
+      FlutterIconData.entypo(61806);
   static const IconData creative_commons_public_domain =
-      const FlutterIconData.entypo(61807);
-  static const IconData creative_commons_remix =
-      const FlutterIconData.entypo(61808);
-  static const IconData creative_commons_share =
-      const FlutterIconData.entypo(61809);
+      FlutterIconData.entypo(61807);
+  static const IconData creative_commons_remix = FlutterIconData.entypo(61808);
+  static const IconData creative_commons_share = FlutterIconData.entypo(61809);
   static const IconData creative_commons_sharealike =
-      const FlutterIconData.entypo(61810);
-  static const IconData credit = const FlutterIconData.entypo(61811);
-  static const IconData credit_card = const FlutterIconData.entypo(61812);
-  static const IconData crop = const FlutterIconData.entypo(61813);
-  static const IconData cross = const FlutterIconData.entypo(61814);
-  static const IconData cup = const FlutterIconData.entypo(61815);
-  static const IconData cw = const FlutterIconData.entypo(61816);
-  static const IconData cycle = const FlutterIconData.entypo(61817);
-  static const IconData database = const FlutterIconData.entypo(61818);
-  static const IconData dial_pad = const FlutterIconData.entypo(61819);
-  static const IconData direction = const FlutterIconData.entypo(61820);
-  static const IconData document = const FlutterIconData.entypo(61821);
-  static const IconData document_landscape =
-      const FlutterIconData.entypo(61822);
-  static const IconData documents = const FlutterIconData.entypo(61823);
-  static const IconData dot_single = const FlutterIconData.entypo(61824);
-  static const IconData dots_three_horizontal =
-      const FlutterIconData.entypo(61825);
-  static const IconData dots_three_vertical =
-      const FlutterIconData.entypo(61826);
-  static const IconData dots_two_horizontal =
-      const FlutterIconData.entypo(61827);
-  static const IconData dots_two_vertical = const FlutterIconData.entypo(61828);
-  static const IconData download = const FlutterIconData.entypo(61829);
-  static const IconData dribbble = const FlutterIconData.entypo(61830);
-  static const IconData dribbble_with_circle =
-      const FlutterIconData.entypo(61831);
-  static const IconData drink = const FlutterIconData.entypo(61832);
-  static const IconData drive = const FlutterIconData.entypo(61833);
-  static const IconData drop = const FlutterIconData.entypo(61834);
-  static const IconData dropbox = const FlutterIconData.entypo(61835);
-  static const IconData edit = const FlutterIconData.entypo(61836);
-  static const IconData email = const FlutterIconData.entypo(61837);
-  static const IconData emoji_flirt = const FlutterIconData.entypo(61838);
-  static const IconData emoji_happy = const FlutterIconData.entypo(61839);
-  static const IconData emoji_neutral = const FlutterIconData.entypo(61840);
-  static const IconData emoji_sad = const FlutterIconData.entypo(61841);
-  static const IconData erase = const FlutterIconData.entypo(61842);
-  static const IconData eraser = const FlutterIconData.entypo(61843);
-  static const IconData evernote = const FlutterIconData.entypo(61844);
-  static const IconData export = const FlutterIconData.entypo(61845);
-  static const IconData eye = const FlutterIconData.entypo(61846);
-  static const IconData eye_with_line = const FlutterIconData.entypo(61847);
-  static const IconData facebook = const FlutterIconData.entypo(61848);
-  static const IconData facebook_with_circle =
-      const FlutterIconData.entypo(61849);
-  static const IconData feather = const FlutterIconData.entypo(61850);
-  static const IconData fingerprint = const FlutterIconData.entypo(61851);
-  static const IconData flag = const FlutterIconData.entypo(61852);
-  static const IconData flash = const FlutterIconData.entypo(61853);
-  static const IconData flashlight = const FlutterIconData.entypo(61854);
-  static const IconData flat_brush = const FlutterIconData.entypo(61855);
-  static const IconData flattr = const FlutterIconData.entypo(61856);
-  static const IconData flickr = const FlutterIconData.entypo(61857);
-  static const IconData flickr_with_circle =
-      const FlutterIconData.entypo(61858);
-  static const IconData flow_branch = const FlutterIconData.entypo(61859);
-  static const IconData flow_cascade = const FlutterIconData.entypo(61860);
-  static const IconData flow_line = const FlutterIconData.entypo(61861);
-  static const IconData flow_parallel = const FlutterIconData.entypo(61862);
-  static const IconData flow_tree = const FlutterIconData.entypo(61863);
-  static const IconData flower = const FlutterIconData.entypo(61864);
-  static const IconData folder = const FlutterIconData.entypo(61865);
-  static const IconData folder_images = const FlutterIconData.entypo(61866);
-  static const IconData folder_music = const FlutterIconData.entypo(61867);
-  static const IconData folder_video = const FlutterIconData.entypo(61868);
-  static const IconData forward = const FlutterIconData.entypo(61869);
-  static const IconData foursquare = const FlutterIconData.entypo(61870);
-  static const IconData funnel = const FlutterIconData.entypo(61871);
-  static const IconData game_controller = const FlutterIconData.entypo(61872);
-  static const IconData gauge = const FlutterIconData.entypo(61873);
-  static const IconData github = const FlutterIconData.entypo(61874);
-  static const IconData github_with_circle =
-      const FlutterIconData.entypo(61875);
-  static const IconData globe = const FlutterIconData.entypo(61876);
-  static const IconData google_ = const FlutterIconData.entypo(61877);
-  static const IconData google__with_circle =
-      const FlutterIconData.entypo(61878);
-  static const IconData google_drive = const FlutterIconData.entypo(61879);
-  static const IconData google_hangouts = const FlutterIconData.entypo(61880);
-  static const IconData google_play = const FlutterIconData.entypo(61881);
-  static const IconData graduation_cap = const FlutterIconData.entypo(61882);
-  static const IconData grid = const FlutterIconData.entypo(61883);
-  static const IconData grooveshark = const FlutterIconData.entypo(61884);
-  static const IconData hair_cross = const FlutterIconData.entypo(61885);
-  static const IconData hand = const FlutterIconData.entypo(61886);
-  static const IconData heart = const FlutterIconData.entypo(61887);
-  static const IconData heart_outlined = const FlutterIconData.entypo(61888);
-  static const IconData help = const FlutterIconData.entypo(61889);
-  static const IconData help_with_circle = const FlutterIconData.entypo(61890);
-  static const IconData home = const FlutterIconData.entypo(61891);
-  static const IconData hour_glass = const FlutterIconData.entypo(61892);
-  static const IconData houzz = const FlutterIconData.entypo(61893);
-  static const IconData icloud = const FlutterIconData.entypo(61894);
-  static const IconData image = const FlutterIconData.entypo(61895);
-  static const IconData image_inverted = const FlutterIconData.entypo(61896);
-  static const IconData images = const FlutterIconData.entypo(61897);
-  static const IconData inbox = const FlutterIconData.entypo(61898);
-  static const IconData infinity = const FlutterIconData.entypo(61899);
-  static const IconData info = const FlutterIconData.entypo(61900);
-  static const IconData info_with_circle = const FlutterIconData.entypo(61901);
-  static const IconData instagram = const FlutterIconData.entypo(61902);
-  static const IconData instagram_with_circle =
-      const FlutterIconData.entypo(61903);
-  static const IconData install = const FlutterIconData.entypo(61904);
-  static const IconData key = const FlutterIconData.entypo(61905);
-  static const IconData keyboard = const FlutterIconData.entypo(61906);
-  static const IconData lab_flask = const FlutterIconData.entypo(61907);
-  static const IconData landline = const FlutterIconData.entypo(61908);
-  static const IconData language = const FlutterIconData.entypo(61909);
-  static const IconData laptop = const FlutterIconData.entypo(61910);
-  static const IconData lastfm = const FlutterIconData.entypo(61911);
-  static const IconData lastfm_with_circle =
-      const FlutterIconData.entypo(61912);
-  static const IconData layers = const FlutterIconData.entypo(61913);
-  static const IconData leaf = const FlutterIconData.entypo(61914);
-  static const IconData level_down = const FlutterIconData.entypo(61915);
-  static const IconData level_up = const FlutterIconData.entypo(61916);
-  static const IconData lifebuoy = const FlutterIconData.entypo(61917);
-  static const IconData light_bulb = const FlutterIconData.entypo(61918);
-  static const IconData light_down = const FlutterIconData.entypo(61919);
-  static const IconData light_up = const FlutterIconData.entypo(61920);
-  static const IconData line_graph = const FlutterIconData.entypo(61921);
-  static const IconData link = const FlutterIconData.entypo(61922);
-  static const IconData linkedin = const FlutterIconData.entypo(61923);
-  static const IconData linkedin_with_circle =
-      const FlutterIconData.entypo(61924);
-  static const IconData list = const FlutterIconData.entypo(61925);
-  static const IconData location = const FlutterIconData.entypo(61926);
-  static const IconData location_pin = const FlutterIconData.entypo(61927);
-  static const IconData lock = const FlutterIconData.entypo(61928);
-  static const IconData lock_open = const FlutterIconData.entypo(61929);
-  static const IconData log_out = const FlutterIconData.entypo(61930);
-  static const IconData login = const FlutterIconData.entypo(61931);
-  static const IconData loop = const FlutterIconData.entypo(61932);
-  static const IconData magnet = const FlutterIconData.entypo(61933);
-  static const IconData magnifying_glass = const FlutterIconData.entypo(61934);
-  static const IconData mail = const FlutterIconData.entypo(61935);
-  static const IconData mail_with_circle = const FlutterIconData.entypo(61936);
-  static const IconData man = const FlutterIconData.entypo(61937);
-  static const IconData map = const FlutterIconData.entypo(61938);
-  static const IconData mask = const FlutterIconData.entypo(61939);
-  static const IconData medal = const FlutterIconData.entypo(61940);
-  static const IconData medium = const FlutterIconData.entypo(61941);
-  static const IconData medium_with_circle =
-      const FlutterIconData.entypo(61942);
-  static const IconData megaphone = const FlutterIconData.entypo(61943);
-  static const IconData menu = const FlutterIconData.entypo(61944);
-  static const IconData merge = const FlutterIconData.entypo(61945);
-  static const IconData message = const FlutterIconData.entypo(61946);
-  static const IconData mic = const FlutterIconData.entypo(61947);
-  static const IconData minus = const FlutterIconData.entypo(61948);
-  static const IconData mixi = const FlutterIconData.entypo(61949);
-  static const IconData mobile = const FlutterIconData.entypo(61950);
-  static const IconData modern_mic = const FlutterIconData.entypo(61951);
-  static const IconData moon = const FlutterIconData.entypo(61952);
-  static const IconData mouse = const FlutterIconData.entypo(61953);
-  static const IconData mouse_pointer = const FlutterIconData.entypo(61954);
-  static const IconData music = const FlutterIconData.entypo(61955);
-  static const IconData network = const FlutterIconData.entypo(61956);
-  static const IconData new_ = const FlutterIconData.entypo(61957);
-  static const IconData new_message = const FlutterIconData.entypo(61958);
-  static const IconData news = const FlutterIconData.entypo(61959);
-  static const IconData newsletter = const FlutterIconData.entypo(61960);
-  static const IconData note = const FlutterIconData.entypo(61961);
-  static const IconData notification = const FlutterIconData.entypo(61962);
-  static const IconData notifications_off = const FlutterIconData.entypo(61963);
-  static const IconData old_mobile = const FlutterIconData.entypo(61964);
-  static const IconData old_phone = const FlutterIconData.entypo(61965);
-  static const IconData onedrive = const FlutterIconData.entypo(61966);
-  static const IconData open_book = const FlutterIconData.entypo(61967);
-  static const IconData palette = const FlutterIconData.entypo(61968);
-  static const IconData paper_plane = const FlutterIconData.entypo(61969);
-  static const IconData paypal = const FlutterIconData.entypo(61970);
-  static const IconData pencil = const FlutterIconData.entypo(61971);
-  static const IconData phone = const FlutterIconData.entypo(61972);
-  static const IconData picasa = const FlutterIconData.entypo(61973);
-  static const IconData pie_chart = const FlutterIconData.entypo(61974);
-  static const IconData pin = const FlutterIconData.entypo(61975);
-  static const IconData pinterest = const FlutterIconData.entypo(61976);
-  static const IconData pinterest_with_circle =
-      const FlutterIconData.entypo(61977);
-  static const IconData plus = const FlutterIconData.entypo(61978);
-  static const IconData popup = const FlutterIconData.entypo(61979);
-  static const IconData power_plug = const FlutterIconData.entypo(61980);
-  static const IconData price_ribbon = const FlutterIconData.entypo(61981);
-  static const IconData price_tag = const FlutterIconData.entypo(61982);
-  static const IconData print = const FlutterIconData.entypo(61983);
-  static const IconData progress_empty = const FlutterIconData.entypo(61984);
-  static const IconData progress_full = const FlutterIconData.entypo(61985);
-  static const IconData progress_one = const FlutterIconData.entypo(61986);
-  static const IconData progress_two = const FlutterIconData.entypo(61987);
-  static const IconData publish = const FlutterIconData.entypo(61988);
-  static const IconData qq = const FlutterIconData.entypo(61989);
-  static const IconData qq_with_circle = const FlutterIconData.entypo(61990);
-  static const IconData quote = const FlutterIconData.entypo(61991);
-  static const IconData radio = const FlutterIconData.entypo(61992);
-  static const IconData raft = const FlutterIconData.entypo(61993);
-  static const IconData raft_with_circle = const FlutterIconData.entypo(61994);
-  static const IconData rainbow = const FlutterIconData.entypo(61995);
-  static const IconData rdio = const FlutterIconData.entypo(61996);
-  static const IconData rdio_with_circle = const FlutterIconData.entypo(61997);
-  static const IconData remove_user = const FlutterIconData.entypo(61998);
-  static const IconData renren = const FlutterIconData.entypo(61999);
-  static const IconData reply = const FlutterIconData.entypo(62000);
-  static const IconData reply_all = const FlutterIconData.entypo(62001);
-  static const IconData resize_100_ = const FlutterIconData.entypo(62002);
-  static const IconData resize_full_screen =
-      const FlutterIconData.entypo(62003);
-  static const IconData retweet = const FlutterIconData.entypo(62004);
-  static const IconData rocket = const FlutterIconData.entypo(62005);
-  static const IconData round_brush = const FlutterIconData.entypo(62006);
-  static const IconData rss = const FlutterIconData.entypo(62007);
-  static const IconData ruler = const FlutterIconData.entypo(62008);
-  static const IconData save = const FlutterIconData.entypo(62009);
-  static const IconData scissors = const FlutterIconData.entypo(62010);
-  static const IconData scribd = const FlutterIconData.entypo(62011);
-  static const IconData select_arrows = const FlutterIconData.entypo(62012);
-  static const IconData share = const FlutterIconData.entypo(62013);
-  static const IconData share_alternative = const FlutterIconData.entypo(62014);
-  static const IconData shareable = const FlutterIconData.entypo(62015);
-  static const IconData shield = const FlutterIconData.entypo(62016);
-  static const IconData shop = const FlutterIconData.entypo(62017);
-  static const IconData shopping_bag = const FlutterIconData.entypo(62018);
-  static const IconData shopping_basket = const FlutterIconData.entypo(62019);
-  static const IconData shopping_cart = const FlutterIconData.entypo(62020);
-  static const IconData shuffle = const FlutterIconData.entypo(62021);
-  static const IconData signal = const FlutterIconData.entypo(62022);
-  static const IconData sina_weibo = const FlutterIconData.entypo(62023);
-  static const IconData skype = const FlutterIconData.entypo(62024);
-  static const IconData skype_with_circle = const FlutterIconData.entypo(62025);
-  static const IconData slideshare = const FlutterIconData.entypo(62026);
-  static const IconData smashing = const FlutterIconData.entypo(62027);
-  static const IconData sound = const FlutterIconData.entypo(62028);
-  static const IconData sound_mix = const FlutterIconData.entypo(62029);
-  static const IconData sound_mute = const FlutterIconData.entypo(62030);
-  static const IconData soundcloud = const FlutterIconData.entypo(62031);
-  static const IconData sports_club = const FlutterIconData.entypo(62032);
-  static const IconData spotify = const FlutterIconData.entypo(62033);
-  static const IconData spotify_with_circle =
-      const FlutterIconData.entypo(62034);
-  static const IconData spreadsheet = const FlutterIconData.entypo(62035);
-  static const IconData squared_cross = const FlutterIconData.entypo(62036);
-  static const IconData squared_minus = const FlutterIconData.entypo(62037);
-  static const IconData squared_plus = const FlutterIconData.entypo(62038);
-  static const IconData star = const FlutterIconData.entypo(62039);
-  static const IconData star_outlined = const FlutterIconData.entypo(62040);
-  static const IconData stopwatch = const FlutterIconData.entypo(62041);
-  static const IconData stumbleupon = const FlutterIconData.entypo(62042);
-  static const IconData stumbleupon_with_circle =
-      const FlutterIconData.entypo(62043);
-  static const IconData suitcase = const FlutterIconData.entypo(62044);
-  static const IconData swap = const FlutterIconData.entypo(62045);
-  static const IconData swarm = const FlutterIconData.entypo(62046);
-  static const IconData sweden = const FlutterIconData.entypo(62047);
-  static const IconData switch_ = const FlutterIconData.entypo(62048);
-  static const IconData tablet = const FlutterIconData.entypo(62049);
-  static const IconData tablet_mobile_combo =
-      const FlutterIconData.entypo(62050);
-  static const IconData tag = const FlutterIconData.entypo(62051);
-  static const IconData text = const FlutterIconData.entypo(62052);
-  static const IconData text_document = const FlutterIconData.entypo(62053);
-  static const IconData text_document_inverted =
-      const FlutterIconData.entypo(62054);
-  static const IconData thermometer = const FlutterIconData.entypo(62055);
-  static const IconData thumbs_down = const FlutterIconData.entypo(62056);
-  static const IconData thumbs_up = const FlutterIconData.entypo(62057);
-  static const IconData thunder_cloud = const FlutterIconData.entypo(62058);
-  static const IconData ticket = const FlutterIconData.entypo(62059);
-  static const IconData time_slot = const FlutterIconData.entypo(62060);
-  static const IconData tools = const FlutterIconData.entypo(62061);
-  static const IconData traffic_cone = const FlutterIconData.entypo(62062);
-  static const IconData trash = const FlutterIconData.entypo(62063);
-  static const IconData tree = const FlutterIconData.entypo(62064);
-  static const IconData triangle_down = const FlutterIconData.entypo(62065);
-  static const IconData triangle_left = const FlutterIconData.entypo(62066);
-  static const IconData triangle_right = const FlutterIconData.entypo(62067);
-  static const IconData triangle_up = const FlutterIconData.entypo(62068);
-  static const IconData tripadvisor = const FlutterIconData.entypo(62069);
-  static const IconData trophy = const FlutterIconData.entypo(62070);
-  static const IconData tumblr = const FlutterIconData.entypo(62071);
-  static const IconData tumblr_with_circle =
-      const FlutterIconData.entypo(62072);
-  static const IconData tv = const FlutterIconData.entypo(62073);
-  static const IconData twitter = const FlutterIconData.entypo(62074);
-  static const IconData twitter_with_circle =
-      const FlutterIconData.entypo(62075);
-  static const IconData typing = const FlutterIconData.entypo(62076);
-  static const IconData uninstall = const FlutterIconData.entypo(62077);
-  static const IconData unread = const FlutterIconData.entypo(62078);
-  static const IconData untag = const FlutterIconData.entypo(62079);
-  static const IconData upload = const FlutterIconData.entypo(62080);
-  static const IconData upload_to_cloud = const FlutterIconData.entypo(62081);
-  static const IconData user = const FlutterIconData.entypo(62082);
-  static const IconData users = const FlutterIconData.entypo(62083);
-  static const IconData v_card = const FlutterIconData.entypo(62084);
-  static const IconData video = const FlutterIconData.entypo(62085);
-  static const IconData video_camera = const FlutterIconData.entypo(62086);
-  static const IconData vimeo = const FlutterIconData.entypo(62087);
-  static const IconData vimeo_with_circle = const FlutterIconData.entypo(62088);
-  static const IconData vine = const FlutterIconData.entypo(62089);
-  static const IconData vine_with_circle = const FlutterIconData.entypo(62090);
-  static const IconData vinyl = const FlutterIconData.entypo(62091);
-  static const IconData vk = const FlutterIconData.entypo(62092);
-  static const IconData vk_alternitive = const FlutterIconData.entypo(62093);
-  static const IconData vk_with_circle = const FlutterIconData.entypo(62094);
-  static const IconData voicemail = const FlutterIconData.entypo(62095);
-  static const IconData wallet = const FlutterIconData.entypo(62096);
-  static const IconData warning = const FlutterIconData.entypo(62097);
-  static const IconData water = const FlutterIconData.entypo(62098);
-  static const IconData windows_store = const FlutterIconData.entypo(62099);
-  static const IconData xing = const FlutterIconData.entypo(62100);
-  static const IconData xing_with_circle = const FlutterIconData.entypo(62101);
-  static const IconData yelp = const FlutterIconData.entypo(62102);
-  static const IconData youko = const FlutterIconData.entypo(62103);
-  static const IconData youko_with_circle = const FlutterIconData.entypo(62104);
-  static const IconData youtube = const FlutterIconData.entypo(62105);
-  static const IconData youtube_with_circle =
-      const FlutterIconData.entypo(62106);
+      FlutterIconData.entypo(61810);
+  static const IconData credit = FlutterIconData.entypo(61811);
+  static const IconData credit_card = FlutterIconData.entypo(61812);
+  static const IconData crop = FlutterIconData.entypo(61813);
+  static const IconData cross = FlutterIconData.entypo(61814);
+  static const IconData cup = FlutterIconData.entypo(61815);
+  static const IconData cw = FlutterIconData.entypo(61816);
+  static const IconData cycle = FlutterIconData.entypo(61817);
+  static const IconData database = FlutterIconData.entypo(61818);
+  static const IconData dial_pad = FlutterIconData.entypo(61819);
+  static const IconData direction = FlutterIconData.entypo(61820);
+  static const IconData document = FlutterIconData.entypo(61821);
+  static const IconData document_landscape = FlutterIconData.entypo(61822);
+  static const IconData documents = FlutterIconData.entypo(61823);
+  static const IconData dot_single = FlutterIconData.entypo(61824);
+  static const IconData dots_three_horizontal = FlutterIconData.entypo(61825);
+  static const IconData dots_three_vertical = FlutterIconData.entypo(61826);
+  static const IconData dots_two_horizontal = FlutterIconData.entypo(61827);
+  static const IconData dots_two_vertical = FlutterIconData.entypo(61828);
+  static const IconData download = FlutterIconData.entypo(61829);
+  static const IconData dribbble = FlutterIconData.entypo(61830);
+  static const IconData dribbble_with_circle = FlutterIconData.entypo(61831);
+  static const IconData drink = FlutterIconData.entypo(61832);
+  static const IconData drive = FlutterIconData.entypo(61833);
+  static const IconData drop = FlutterIconData.entypo(61834);
+  static const IconData dropbox = FlutterIconData.entypo(61835);
+  static const IconData edit = FlutterIconData.entypo(61836);
+  static const IconData email = FlutterIconData.entypo(61837);
+  static const IconData emoji_flirt = FlutterIconData.entypo(61838);
+  static const IconData emoji_happy = FlutterIconData.entypo(61839);
+  static const IconData emoji_neutral = FlutterIconData.entypo(61840);
+  static const IconData emoji_sad = FlutterIconData.entypo(61841);
+  static const IconData erase = FlutterIconData.entypo(61842);
+  static const IconData eraser = FlutterIconData.entypo(61843);
+  static const IconData evernote = FlutterIconData.entypo(61844);
+  static const IconData export = FlutterIconData.entypo(61845);
+  static const IconData eye = FlutterIconData.entypo(61846);
+  static const IconData eye_with_line = FlutterIconData.entypo(61847);
+  static const IconData facebook = FlutterIconData.entypo(61848);
+  static const IconData facebook_with_circle = FlutterIconData.entypo(61849);
+  static const IconData feather = FlutterIconData.entypo(61850);
+  static const IconData fingerprint = FlutterIconData.entypo(61851);
+  static const IconData flag = FlutterIconData.entypo(61852);
+  static const IconData flash = FlutterIconData.entypo(61853);
+  static const IconData flashlight = FlutterIconData.entypo(61854);
+  static const IconData flat_brush = FlutterIconData.entypo(61855);
+  static const IconData flattr = FlutterIconData.entypo(61856);
+  static const IconData flickr = FlutterIconData.entypo(61857);
+  static const IconData flickr_with_circle = FlutterIconData.entypo(61858);
+  static const IconData flow_branch = FlutterIconData.entypo(61859);
+  static const IconData flow_cascade = FlutterIconData.entypo(61860);
+  static const IconData flow_line = FlutterIconData.entypo(61861);
+  static const IconData flow_parallel = FlutterIconData.entypo(61862);
+  static const IconData flow_tree = FlutterIconData.entypo(61863);
+  static const IconData flower = FlutterIconData.entypo(61864);
+  static const IconData folder = FlutterIconData.entypo(61865);
+  static const IconData folder_images = FlutterIconData.entypo(61866);
+  static const IconData folder_music = FlutterIconData.entypo(61867);
+  static const IconData folder_video = FlutterIconData.entypo(61868);
+  static const IconData forward = FlutterIconData.entypo(61869);
+  static const IconData foursquare = FlutterIconData.entypo(61870);
+  static const IconData funnel = FlutterIconData.entypo(61871);
+  static const IconData game_controller = FlutterIconData.entypo(61872);
+  static const IconData gauge = FlutterIconData.entypo(61873);
+  static const IconData github = FlutterIconData.entypo(61874);
+  static const IconData github_with_circle = FlutterIconData.entypo(61875);
+  static const IconData globe = FlutterIconData.entypo(61876);
+  static const IconData google_ = FlutterIconData.entypo(61877);
+  static const IconData google__with_circle = FlutterIconData.entypo(61878);
+  static const IconData google_drive = FlutterIconData.entypo(61879);
+  static const IconData google_hangouts = FlutterIconData.entypo(61880);
+  static const IconData google_play = FlutterIconData.entypo(61881);
+  static const IconData graduation_cap = FlutterIconData.entypo(61882);
+  static const IconData grid = FlutterIconData.entypo(61883);
+  static const IconData grooveshark = FlutterIconData.entypo(61884);
+  static const IconData hair_cross = FlutterIconData.entypo(61885);
+  static const IconData hand = FlutterIconData.entypo(61886);
+  static const IconData heart = FlutterIconData.entypo(61887);
+  static const IconData heart_outlined = FlutterIconData.entypo(61888);
+  static const IconData help = FlutterIconData.entypo(61889);
+  static const IconData help_with_circle = FlutterIconData.entypo(61890);
+  static const IconData home = FlutterIconData.entypo(61891);
+  static const IconData hour_glass = FlutterIconData.entypo(61892);
+  static const IconData houzz = FlutterIconData.entypo(61893);
+  static const IconData icloud = FlutterIconData.entypo(61894);
+  static const IconData image = FlutterIconData.entypo(61895);
+  static const IconData image_inverted = FlutterIconData.entypo(61896);
+  static const IconData images = FlutterIconData.entypo(61897);
+  static const IconData inbox = FlutterIconData.entypo(61898);
+  static const IconData infinity = FlutterIconData.entypo(61899);
+  static const IconData info = FlutterIconData.entypo(61900);
+  static const IconData info_with_circle = FlutterIconData.entypo(61901);
+  static const IconData instagram = FlutterIconData.entypo(61902);
+  static const IconData instagram_with_circle = FlutterIconData.entypo(61903);
+  static const IconData install = FlutterIconData.entypo(61904);
+  static const IconData key = FlutterIconData.entypo(61905);
+  static const IconData keyboard = FlutterIconData.entypo(61906);
+  static const IconData lab_flask = FlutterIconData.entypo(61907);
+  static const IconData landline = FlutterIconData.entypo(61908);
+  static const IconData language = FlutterIconData.entypo(61909);
+  static const IconData laptop = FlutterIconData.entypo(61910);
+  static const IconData lastfm = FlutterIconData.entypo(61911);
+  static const IconData lastfm_with_circle = FlutterIconData.entypo(61912);
+  static const IconData layers = FlutterIconData.entypo(61913);
+  static const IconData leaf = FlutterIconData.entypo(61914);
+  static const IconData level_down = FlutterIconData.entypo(61915);
+  static const IconData level_up = FlutterIconData.entypo(61916);
+  static const IconData lifebuoy = FlutterIconData.entypo(61917);
+  static const IconData light_bulb = FlutterIconData.entypo(61918);
+  static const IconData light_down = FlutterIconData.entypo(61919);
+  static const IconData light_up = FlutterIconData.entypo(61920);
+  static const IconData line_graph = FlutterIconData.entypo(61921);
+  static const IconData link = FlutterIconData.entypo(61922);
+  static const IconData linkedin = FlutterIconData.entypo(61923);
+  static const IconData linkedin_with_circle = FlutterIconData.entypo(61924);
+  static const IconData list = FlutterIconData.entypo(61925);
+  static const IconData location = FlutterIconData.entypo(61926);
+  static const IconData location_pin = FlutterIconData.entypo(61927);
+  static const IconData lock = FlutterIconData.entypo(61928);
+  static const IconData lock_open = FlutterIconData.entypo(61929);
+  static const IconData log_out = FlutterIconData.entypo(61930);
+  static const IconData login = FlutterIconData.entypo(61931);
+  static const IconData loop = FlutterIconData.entypo(61932);
+  static const IconData magnet = FlutterIconData.entypo(61933);
+  static const IconData magnifying_glass = FlutterIconData.entypo(61934);
+  static const IconData mail = FlutterIconData.entypo(61935);
+  static const IconData mail_with_circle = FlutterIconData.entypo(61936);
+  static const IconData man = FlutterIconData.entypo(61937);
+  static const IconData map = FlutterIconData.entypo(61938);
+  static const IconData mask = FlutterIconData.entypo(61939);
+  static const IconData medal = FlutterIconData.entypo(61940);
+  static const IconData medium = FlutterIconData.entypo(61941);
+  static const IconData medium_with_circle = FlutterIconData.entypo(61942);
+  static const IconData megaphone = FlutterIconData.entypo(61943);
+  static const IconData menu = FlutterIconData.entypo(61944);
+  static const IconData merge = FlutterIconData.entypo(61945);
+  static const IconData message = FlutterIconData.entypo(61946);
+  static const IconData mic = FlutterIconData.entypo(61947);
+  static const IconData minus = FlutterIconData.entypo(61948);
+  static const IconData mixi = FlutterIconData.entypo(61949);
+  static const IconData mobile = FlutterIconData.entypo(61950);
+  static const IconData modern_mic = FlutterIconData.entypo(61951);
+  static const IconData moon = FlutterIconData.entypo(61952);
+  static const IconData mouse = FlutterIconData.entypo(61953);
+  static const IconData mouse_pointer = FlutterIconData.entypo(61954);
+  static const IconData music = FlutterIconData.entypo(61955);
+  static const IconData network = FlutterIconData.entypo(61956);
+  static const IconData new_ = FlutterIconData.entypo(61957);
+  static const IconData new_message = FlutterIconData.entypo(61958);
+  static const IconData news = FlutterIconData.entypo(61959);
+  static const IconData newsletter = FlutterIconData.entypo(61960);
+  static const IconData note = FlutterIconData.entypo(61961);
+  static const IconData notification = FlutterIconData.entypo(61962);
+  static const IconData notifications_off = FlutterIconData.entypo(61963);
+  static const IconData old_mobile = FlutterIconData.entypo(61964);
+  static const IconData old_phone = FlutterIconData.entypo(61965);
+  static const IconData onedrive = FlutterIconData.entypo(61966);
+  static const IconData open_book = FlutterIconData.entypo(61967);
+  static const IconData palette = FlutterIconData.entypo(61968);
+  static const IconData paper_plane = FlutterIconData.entypo(61969);
+  static const IconData paypal = FlutterIconData.entypo(61970);
+  static const IconData pencil = FlutterIconData.entypo(61971);
+  static const IconData phone = FlutterIconData.entypo(61972);
+  static const IconData picasa = FlutterIconData.entypo(61973);
+  static const IconData pie_chart = FlutterIconData.entypo(61974);
+  static const IconData pin = FlutterIconData.entypo(61975);
+  static const IconData pinterest = FlutterIconData.entypo(61976);
+  static const IconData pinterest_with_circle = FlutterIconData.entypo(61977);
+  static const IconData plus = FlutterIconData.entypo(61978);
+  static const IconData popup = FlutterIconData.entypo(61979);
+  static const IconData power_plug = FlutterIconData.entypo(61980);
+  static const IconData price_ribbon = FlutterIconData.entypo(61981);
+  static const IconData price_tag = FlutterIconData.entypo(61982);
+  static const IconData print = FlutterIconData.entypo(61983);
+  static const IconData progress_empty = FlutterIconData.entypo(61984);
+  static const IconData progress_full = FlutterIconData.entypo(61985);
+  static const IconData progress_one = FlutterIconData.entypo(61986);
+  static const IconData progress_two = FlutterIconData.entypo(61987);
+  static const IconData publish = FlutterIconData.entypo(61988);
+  static const IconData qq = FlutterIconData.entypo(61989);
+  static const IconData qq_with_circle = FlutterIconData.entypo(61990);
+  static const IconData quote = FlutterIconData.entypo(61991);
+  static const IconData radio = FlutterIconData.entypo(61992);
+  static const IconData raft = FlutterIconData.entypo(61993);
+  static const IconData raft_with_circle = FlutterIconData.entypo(61994);
+  static const IconData rainbow = FlutterIconData.entypo(61995);
+  static const IconData rdio = FlutterIconData.entypo(61996);
+  static const IconData rdio_with_circle = FlutterIconData.entypo(61997);
+  static const IconData remove_user = FlutterIconData.entypo(61998);
+  static const IconData renren = FlutterIconData.entypo(61999);
+  static const IconData reply = FlutterIconData.entypo(62000);
+  static const IconData reply_all = FlutterIconData.entypo(62001);
+  static const IconData resize_100_ = FlutterIconData.entypo(62002);
+  static const IconData resize_full_screen = FlutterIconData.entypo(62003);
+  static const IconData retweet = FlutterIconData.entypo(62004);
+  static const IconData rocket = FlutterIconData.entypo(62005);
+  static const IconData round_brush = FlutterIconData.entypo(62006);
+  static const IconData rss = FlutterIconData.entypo(62007);
+  static const IconData ruler = FlutterIconData.entypo(62008);
+  static const IconData save = FlutterIconData.entypo(62009);
+  static const IconData scissors = FlutterIconData.entypo(62010);
+  static const IconData scribd = FlutterIconData.entypo(62011);
+  static const IconData select_arrows = FlutterIconData.entypo(62012);
+  static const IconData share = FlutterIconData.entypo(62013);
+  static const IconData share_alternative = FlutterIconData.entypo(62014);
+  static const IconData shareable = FlutterIconData.entypo(62015);
+  static const IconData shield = FlutterIconData.entypo(62016);
+  static const IconData shop = FlutterIconData.entypo(62017);
+  static const IconData shopping_bag = FlutterIconData.entypo(62018);
+  static const IconData shopping_basket = FlutterIconData.entypo(62019);
+  static const IconData shopping_cart = FlutterIconData.entypo(62020);
+  static const IconData shuffle = FlutterIconData.entypo(62021);
+  static const IconData signal = FlutterIconData.entypo(62022);
+  static const IconData sina_weibo = FlutterIconData.entypo(62023);
+  static const IconData skype = FlutterIconData.entypo(62024);
+  static const IconData skype_with_circle = FlutterIconData.entypo(62025);
+  static const IconData slideshare = FlutterIconData.entypo(62026);
+  static const IconData smashing = FlutterIconData.entypo(62027);
+  static const IconData sound = FlutterIconData.entypo(62028);
+  static const IconData sound_mix = FlutterIconData.entypo(62029);
+  static const IconData sound_mute = FlutterIconData.entypo(62030);
+  static const IconData soundcloud = FlutterIconData.entypo(62031);
+  static const IconData sports_club = FlutterIconData.entypo(62032);
+  static const IconData spotify = FlutterIconData.entypo(62033);
+  static const IconData spotify_with_circle = FlutterIconData.entypo(62034);
+  static const IconData spreadsheet = FlutterIconData.entypo(62035);
+  static const IconData squared_cross = FlutterIconData.entypo(62036);
+  static const IconData squared_minus = FlutterIconData.entypo(62037);
+  static const IconData squared_plus = FlutterIconData.entypo(62038);
+  static const IconData star = FlutterIconData.entypo(62039);
+  static const IconData star_outlined = FlutterIconData.entypo(62040);
+  static const IconData stopwatch = FlutterIconData.entypo(62041);
+  static const IconData stumbleupon = FlutterIconData.entypo(62042);
+  static const IconData stumbleupon_with_circle = FlutterIconData.entypo(62043);
+  static const IconData suitcase = FlutterIconData.entypo(62044);
+  static const IconData swap = FlutterIconData.entypo(62045);
+  static const IconData swarm = FlutterIconData.entypo(62046);
+  static const IconData sweden = FlutterIconData.entypo(62047);
+  static const IconData switch_ = FlutterIconData.entypo(62048);
+  static const IconData tablet = FlutterIconData.entypo(62049);
+  static const IconData tablet_mobile_combo = FlutterIconData.entypo(62050);
+  static const IconData tag = FlutterIconData.entypo(62051);
+  static const IconData text = FlutterIconData.entypo(62052);
+  static const IconData text_document = FlutterIconData.entypo(62053);
+  static const IconData text_document_inverted = FlutterIconData.entypo(62054);
+  static const IconData thermometer = FlutterIconData.entypo(62055);
+  static const IconData thumbs_down = FlutterIconData.entypo(62056);
+  static const IconData thumbs_up = FlutterIconData.entypo(62057);
+  static const IconData thunder_cloud = FlutterIconData.entypo(62058);
+  static const IconData ticket = FlutterIconData.entypo(62059);
+  static const IconData time_slot = FlutterIconData.entypo(62060);
+  static const IconData tools = FlutterIconData.entypo(62061);
+  static const IconData traffic_cone = FlutterIconData.entypo(62062);
+  static const IconData trash = FlutterIconData.entypo(62063);
+  static const IconData tree = FlutterIconData.entypo(62064);
+  static const IconData triangle_down = FlutterIconData.entypo(62065);
+  static const IconData triangle_left = FlutterIconData.entypo(62066);
+  static const IconData triangle_right = FlutterIconData.entypo(62067);
+  static const IconData triangle_up = FlutterIconData.entypo(62068);
+  static const IconData tripadvisor = FlutterIconData.entypo(62069);
+  static const IconData trophy = FlutterIconData.entypo(62070);
+  static const IconData tumblr = FlutterIconData.entypo(62071);
+  static const IconData tumblr_with_circle = FlutterIconData.entypo(62072);
+  static const IconData tv = FlutterIconData.entypo(62073);
+  static const IconData twitter = FlutterIconData.entypo(62074);
+  static const IconData twitter_with_circle = FlutterIconData.entypo(62075);
+  static const IconData typing = FlutterIconData.entypo(62076);
+  static const IconData uninstall = FlutterIconData.entypo(62077);
+  static const IconData unread = FlutterIconData.entypo(62078);
+  static const IconData untag = FlutterIconData.entypo(62079);
+  static const IconData upload = FlutterIconData.entypo(62080);
+  static const IconData upload_to_cloud = FlutterIconData.entypo(62081);
+  static const IconData user = FlutterIconData.entypo(62082);
+  static const IconData users = FlutterIconData.entypo(62083);
+  static const IconData v_card = FlutterIconData.entypo(62084);
+  static const IconData video = FlutterIconData.entypo(62085);
+  static const IconData video_camera = FlutterIconData.entypo(62086);
+  static const IconData vimeo = FlutterIconData.entypo(62087);
+  static const IconData vimeo_with_circle = FlutterIconData.entypo(62088);
+  static const IconData vine = FlutterIconData.entypo(62089);
+  static const IconData vine_with_circle = FlutterIconData.entypo(62090);
+  static const IconData vinyl = FlutterIconData.entypo(62091);
+  static const IconData vk = FlutterIconData.entypo(62092);
+  static const IconData vk_alternitive = FlutterIconData.entypo(62093);
+  static const IconData vk_with_circle = FlutterIconData.entypo(62094);
+  static const IconData voicemail = FlutterIconData.entypo(62095);
+  static const IconData wallet = FlutterIconData.entypo(62096);
+  static const IconData warning = FlutterIconData.entypo(62097);
+  static const IconData water = FlutterIconData.entypo(62098);
+  static const IconData windows_store = FlutterIconData.entypo(62099);
+  static const IconData xing = FlutterIconData.entypo(62100);
+  static const IconData xing_with_circle = FlutterIconData.entypo(62101);
+  static const IconData yelp = FlutterIconData.entypo(62102);
+  static const IconData youko = FlutterIconData.entypo(62103);
+  static const IconData youko_with_circle = FlutterIconData.entypo(62104);
+  static const IconData youtube = FlutterIconData.entypo(62105);
+  static const IconData youtube_with_circle = FlutterIconData.entypo(62106);
 }
