@@ -6,75 +6,76 @@ import 'flutter_icon_data.dart';
 /// http://evil-icons.io
 class EvilIcons {
   EvilIcons._();
-  static const IconData archive = const FlutterIconData.evilIcons(61696);
-  static const IconData arrow_down = const FlutterIconData.evilIcons(61697);
-  static const IconData arrow_left = const FlutterIconData.evilIcons(61698);
-  static const IconData arrow_right = const FlutterIconData.evilIcons(61699);
-  static const IconData arrow_up = const FlutterIconData.evilIcons(61700);
-  static const IconData bell = const FlutterIconData.evilIcons(61701);
-  static const IconData calendar = const FlutterIconData.evilIcons(61702);
-  static const IconData camera = const FlutterIconData.evilIcons(61703);
-  static const IconData cart = const FlutterIconData.evilIcons(61704);
-  static const IconData chart = const FlutterIconData.evilIcons(61705);
-  static const IconData check = const FlutterIconData.evilIcons(61706);
-  static const IconData chevron_down = const FlutterIconData.evilIcons(61707);
-  static const IconData chevron_left = const FlutterIconData.evilIcons(61708);
-  static const IconData chevron_right = const FlutterIconData.evilIcons(61709);
-  static const IconData chevron_up = const FlutterIconData.evilIcons(61710);
-  static const IconData clock = const FlutterIconData.evilIcons(61711);
-  static const IconData close = const FlutterIconData.evilIcons(61712);
-  static const IconData close_o = const FlutterIconData.evilIcons(61713);
-  static const IconData comment = const FlutterIconData.evilIcons(61714);
-  static const IconData credit_card = const FlutterIconData.evilIcons(61715);
-  static const IconData envelope = const FlutterIconData.evilIcons(61716);
-  static const IconData exclamation = const FlutterIconData.evilIcons(61717);
-  static const IconData external_link = const FlutterIconData.evilIcons(61718);
-  static const IconData eye = const FlutterIconData.evilIcons(61719);
-  static const IconData gear = const FlutterIconData.evilIcons(61720);
-  static const IconData heart = const FlutterIconData.evilIcons(61721);
-  static const IconData image = const FlutterIconData.evilIcons(61722);
-  static const IconData like = const FlutterIconData.evilIcons(61723);
-  static const IconData link = const FlutterIconData.evilIcons(61724);
-  static const IconData location = const FlutterIconData.evilIcons(61725);
-  static const IconData lock = const FlutterIconData.evilIcons(61726);
-  static const IconData minus = const FlutterIconData.evilIcons(61727);
-  static const IconData navicon = const FlutterIconData.evilIcons(61728);
-  static const IconData paperclip = const FlutterIconData.evilIcons(61729);
-  static const IconData pencil = const FlutterIconData.evilIcons(61730);
-  static const IconData play = const FlutterIconData.evilIcons(61731);
-  static const IconData plus = const FlutterIconData.evilIcons(61732);
-  static const IconData pointer = const FlutterIconData.evilIcons(61733);
-  static const IconData question = const FlutterIconData.evilIcons(61734);
-  static const IconData redo = const FlutterIconData.evilIcons(61735);
-  static const IconData refresh = const FlutterIconData.evilIcons(61736);
-  static const IconData retweet = const FlutterIconData.evilIcons(61737);
-  static const IconData sc_facebook = const FlutterIconData.evilIcons(61738);
-  static const IconData sc_github = const FlutterIconData.evilIcons(61739);
-  static const IconData sc_google_plus = const FlutterIconData.evilIcons(61740);
-  static const IconData sc_instagram = const FlutterIconData.evilIcons(61741);
-  static const IconData sc_linkedin = const FlutterIconData.evilIcons(61742);
+
+  static const IconData archive = FlutterIconData.evilIcons(61696);
+  static const IconData arrow_down = FlutterIconData.evilIcons(61697);
+  static const IconData arrow_left = FlutterIconData.evilIcons(61698);
+  static const IconData arrow_right = FlutterIconData.evilIcons(61699);
+  static const IconData arrow_up = FlutterIconData.evilIcons(61700);
+  static const IconData bell = FlutterIconData.evilIcons(61701);
+  static const IconData calendar = FlutterIconData.evilIcons(61702);
+  static const IconData camera = FlutterIconData.evilIcons(61703);
+  static const IconData cart = FlutterIconData.evilIcons(61704);
+  static const IconData chart = FlutterIconData.evilIcons(61705);
+  static const IconData check = FlutterIconData.evilIcons(61706);
+  static const IconData chevron_down = FlutterIconData.evilIcons(61707);
+  static const IconData chevron_left = FlutterIconData.evilIcons(61708);
+  static const IconData chevron_right = FlutterIconData.evilIcons(61709);
+  static const IconData chevron_up = FlutterIconData.evilIcons(61710);
+  static const IconData clock = FlutterIconData.evilIcons(61711);
+  static const IconData close = FlutterIconData.evilIcons(61712);
+  static const IconData close_o = FlutterIconData.evilIcons(61713);
+  static const IconData comment = FlutterIconData.evilIcons(61714);
+  static const IconData credit_card = FlutterIconData.evilIcons(61715);
+  static const IconData envelope = FlutterIconData.evilIcons(61716);
+  static const IconData exclamation = FlutterIconData.evilIcons(61717);
+  static const IconData external_link = FlutterIconData.evilIcons(61718);
+  static const IconData eye = FlutterIconData.evilIcons(61719);
+  static const IconData gear = FlutterIconData.evilIcons(61720);
+  static const IconData heart = FlutterIconData.evilIcons(61721);
+  static const IconData image = FlutterIconData.evilIcons(61722);
+  static const IconData like = FlutterIconData.evilIcons(61723);
+  static const IconData link = FlutterIconData.evilIcons(61724);
+  static const IconData location = FlutterIconData.evilIcons(61725);
+  static const IconData lock = FlutterIconData.evilIcons(61726);
+  static const IconData minus = FlutterIconData.evilIcons(61727);
+  static const IconData navicon = FlutterIconData.evilIcons(61728);
+  static const IconData paperclip = FlutterIconData.evilIcons(61729);
+  static const IconData pencil = FlutterIconData.evilIcons(61730);
+  static const IconData play = FlutterIconData.evilIcons(61731);
+  static const IconData plus = FlutterIconData.evilIcons(61732);
+  static const IconData pointer = FlutterIconData.evilIcons(61733);
+  static const IconData question = FlutterIconData.evilIcons(61734);
+  static const IconData redo = FlutterIconData.evilIcons(61735);
+  static const IconData refresh = FlutterIconData.evilIcons(61736);
+  static const IconData retweet = FlutterIconData.evilIcons(61737);
+  static const IconData sc_facebook = FlutterIconData.evilIcons(61738);
+  static const IconData sc_github = FlutterIconData.evilIcons(61739);
+  static const IconData sc_google_plus = FlutterIconData.evilIcons(61740);
+  static const IconData sc_instagram = FlutterIconData.evilIcons(61741);
+  static const IconData sc_linkedin = FlutterIconData.evilIcons(61742);
   static const IconData sc_odnoklassniki =
-      const FlutterIconData.evilIcons(61743);
-  static const IconData sc_pinterest = const FlutterIconData.evilIcons(61744);
-  static const IconData sc_skype = const FlutterIconData.evilIcons(61745);
-  static const IconData sc_soundcloud = const FlutterIconData.evilIcons(61746);
-  static const IconData sc_telegram = const FlutterIconData.evilIcons(61747);
-  static const IconData sc_tumblr = const FlutterIconData.evilIcons(61748);
-  static const IconData sc_twitter = const FlutterIconData.evilIcons(61749);
-  static const IconData sc_vimeo = const FlutterIconData.evilIcons(61750);
-  static const IconData sc_vk = const FlutterIconData.evilIcons(61751);
-  static const IconData sc_youtube = const FlutterIconData.evilIcons(61752);
-  static const IconData search = const FlutterIconData.evilIcons(61753);
-  static const IconData share_apple = const FlutterIconData.evilIcons(61754);
-  static const IconData share_google = const FlutterIconData.evilIcons(61755);
-  static const IconData spinner = const FlutterIconData.evilIcons(61756);
-  static const IconData spinner_2 = const FlutterIconData.evilIcons(61757);
-  static const IconData spinner_3 = const FlutterIconData.evilIcons(61758);
-  static const IconData star = const FlutterIconData.evilIcons(61759);
-  static const IconData tag = const FlutterIconData.evilIcons(61760);
-  static const IconData trash = const FlutterIconData.evilIcons(61761);
-  static const IconData trophy = const FlutterIconData.evilIcons(61762);
-  static const IconData undo = const FlutterIconData.evilIcons(61763);
-  static const IconData unlock = const FlutterIconData.evilIcons(61764);
-  static const IconData user = const FlutterIconData.evilIcons(61765);
+      FlutterIconData.evilIcons(61743);
+  static const IconData sc_pinterest = FlutterIconData.evilIcons(61744);
+  static const IconData sc_skype = FlutterIconData.evilIcons(61745);
+  static const IconData sc_soundcloud = FlutterIconData.evilIcons(61746);
+  static const IconData sc_telegram = FlutterIconData.evilIcons(61747);
+  static const IconData sc_tumblr = FlutterIconData.evilIcons(61748);
+  static const IconData sc_twitter = FlutterIconData.evilIcons(61749);
+  static const IconData sc_vimeo = FlutterIconData.evilIcons(61750);
+  static const IconData sc_vk = FlutterIconData.evilIcons(61751);
+  static const IconData sc_youtube = FlutterIconData.evilIcons(61752);
+  static const IconData search = FlutterIconData.evilIcons(61753);
+  static const IconData share_apple = FlutterIconData.evilIcons(61754);
+  static const IconData share_google = FlutterIconData.evilIcons(61755);
+  static const IconData spinner = FlutterIconData.evilIcons(61756);
+  static const IconData spinner_2 = FlutterIconData.evilIcons(61757);
+  static const IconData spinner_3 = FlutterIconData.evilIcons(61758);
+  static const IconData star = FlutterIconData.evilIcons(61759);
+  static const IconData tag = FlutterIconData.evilIcons(61760);
+  static const IconData trash = FlutterIconData.evilIcons(61761);
+  static const IconData trophy = FlutterIconData.evilIcons(61762);
+  static const IconData undo = FlutterIconData.evilIcons(61763);
+  static const IconData unlock = FlutterIconData.evilIcons(61764);
+  static const IconData user = FlutterIconData.evilIcons(61765);
 }
