@@ -12,11 +12,14 @@ class FlutterIconData extends IconData {
   const FlutterIconData.fontAwesome(int codePoint)
       : this(codePoint, "FontAwesome");
 
-  const FlutterIconData.fontAwesome5Brands(int codePoint):this(codePoint,"FontAwesome5_Brands");
+  const FlutterIconData.fontAwesome5Brands(int codePoint)
+      : this(codePoint, "FontAwesome5_Brands");
 
-  const FlutterIconData.fontAwesome5(int codePoint):this(codePoint,"FontAwesome5");
+  const FlutterIconData.fontAwesome5(int codePoint)
+      : this(codePoint, "FontAwesome5");
 
-  const FlutterIconData.fontAwesome5Solid(int codePoint):this(codePoint,"FontAwesome5_Solid");
+  const FlutterIconData.fontAwesome5Solid(int codePoint)
+      : this(codePoint, "FontAwesome5_Solid");
 
   const FlutterIconData.entypo(int codePoint) : this(codePoint, "Entypo");
 
@@ -40,5 +43,6 @@ class FlutterIconData extends IconData {
 
   const FlutterIconData.zocial(int codePoint) : this(codePoint, "Zocial");
 
-  const FlutterIconData.weatherIcons(int codePoint) : this(codePoint,"WeatherIcons");
+  const FlutterIconData.weatherIcons(int codePoint)
+      : this(codePoint, "WeatherIcons");
 }
