@@ -6,204 +6,164 @@ import 'flutter_icon_data.dart';
 /// https://fontawesome.com
 class FontAwesome5 {
   FontAwesome5._();
-  static const IconData address_book =
-      const FlutterIconData.fontAwesome5(62137);
-  static const IconData address_card =
-      const FlutterIconData.fontAwesome5(62139);
-  static const IconData angry = const FlutterIconData.fontAwesome5(62806);
+
+  static const IconData address_book = FlutterIconData.fontAwesome5(62137);
+  static const IconData address_card = FlutterIconData.fontAwesome5(62139);
+  static const IconData angry = FlutterIconData.fontAwesome5(62806);
   static const IconData arrow_alt_circle_down =
-      const FlutterIconData.fontAwesome5(62296);
+      FlutterIconData.fontAwesome5(62296);
   static const IconData arrow_alt_circle_left =
-      const FlutterIconData.fontAwesome5(62297);
+      FlutterIconData.fontAwesome5(62297);
   static const IconData arrow_alt_circle_right =
-      const FlutterIconData.fontAwesome5(62298);
+      FlutterIconData.fontAwesome5(62298);
   static const IconData arrow_alt_circle_up =
-      const FlutterIconData.fontAwesome5(62299);
-  static const IconData bell_slash = const FlutterIconData.fontAwesome5(61942);
-  static const IconData bell = const FlutterIconData.fontAwesome5(61683);
-  static const IconData bookmark = const FlutterIconData.fontAwesome5(61486);
-  static const IconData building = const FlutterIconData.fontAwesome5(61869);
-  static const IconData calendar_alt =
-      const FlutterIconData.fontAwesome5(61555);
-  static const IconData calendar_check =
-      const FlutterIconData.fontAwesome5(62068);
-  static const IconData calendar_minus =
-      const FlutterIconData.fontAwesome5(62066);
-  static const IconData calendar_plus =
-      const FlutterIconData.fontAwesome5(62065);
-  static const IconData calendar_times =
-      const FlutterIconData.fontAwesome5(62067);
-  static const IconData calendar = const FlutterIconData.fontAwesome5(61747);
-  static const IconData caret_square_down =
-      const FlutterIconData.fontAwesome5(61776);
-  static const IconData caret_square_left =
-      const FlutterIconData.fontAwesome5(61841);
+      FlutterIconData.fontAwesome5(62299);
+  static const IconData bell_slash = FlutterIconData.fontAwesome5(61942);
+  static const IconData bell = FlutterIconData.fontAwesome5(61683);
+  static const IconData bookmark = FlutterIconData.fontAwesome5(61486);
+  static const IconData building = FlutterIconData.fontAwesome5(61869);
+  static const IconData calendar_alt = FlutterIconData.fontAwesome5(61555);
+  static const IconData calendar_check = FlutterIconData.fontAwesome5(62068);
+  static const IconData calendar_minus = FlutterIconData.fontAwesome5(62066);
+  static const IconData calendar_plus = FlutterIconData.fontAwesome5(62065);
+  static const IconData calendar_times = FlutterIconData.fontAwesome5(62067);
+  static const IconData calendar = FlutterIconData.fontAwesome5(61747);
+  static const IconData caret_square_down = FlutterIconData.fontAwesome5(61776);
+  static const IconData caret_square_left = FlutterIconData.fontAwesome5(61841);
   static const IconData caret_square_right =
-      const FlutterIconData.fontAwesome5(61778);
-  static const IconData caret_square_up =
-      const FlutterIconData.fontAwesome5(61777);
-  static const IconData chart_bar = const FlutterIconData.fontAwesome5(61568);
-  static const IconData check_circle =
-      const FlutterIconData.fontAwesome5(61528);
-  static const IconData check_square =
-      const FlutterIconData.fontAwesome5(61770);
-  static const IconData circle = const FlutterIconData.fontAwesome5(61713);
-  static const IconData clipboard = const FlutterIconData.fontAwesome5(62248);
-  static const IconData clock = const FlutterIconData.fontAwesome5(61463);
-  static const IconData clone = const FlutterIconData.fontAwesome5(62029);
-  static const IconData closed_captioning =
-      const FlutterIconData.fontAwesome5(61962);
-  static const IconData comment_alt = const FlutterIconData.fontAwesome5(62074);
-  static const IconData comment_dots =
-      const FlutterIconData.fontAwesome5(62637);
-  static const IconData comment = const FlutterIconData.fontAwesome5(61557);
-  static const IconData comments = const FlutterIconData.fontAwesome5(61574);
-  static const IconData compass = const FlutterIconData.fontAwesome5(61774);
-  static const IconData copy = const FlutterIconData.fontAwesome5(61637);
-  static const IconData copyright = const FlutterIconData.fontAwesome5(61945);
-  static const IconData credit_card = const FlutterIconData.fontAwesome5(61597);
-  static const IconData dizzy = const FlutterIconData.fontAwesome5(62823);
-  static const IconData dot_circle = const FlutterIconData.fontAwesome5(61842);
-  static const IconData edit = const FlutterIconData.fontAwesome5(61508);
-  static const IconData envelope_open =
-      const FlutterIconData.fontAwesome5(62134);
-  static const IconData envelope = const FlutterIconData.fontAwesome5(61664);
-  static const IconData eye_slash = const FlutterIconData.fontAwesome5(61552);
-  static const IconData eye = const FlutterIconData.fontAwesome5(61550);
-  static const IconData file_alt = const FlutterIconData.fontAwesome5(61788);
-  static const IconData file_archive =
-      const FlutterIconData.fontAwesome5(61894);
-  static const IconData file_audio = const FlutterIconData.fontAwesome5(61895);
-  static const IconData file_code = const FlutterIconData.fontAwesome5(61897);
-  static const IconData file_excel = const FlutterIconData.fontAwesome5(61891);
-  static const IconData file_image = const FlutterIconData.fontAwesome5(61893);
-  static const IconData file_pdf = const FlutterIconData.fontAwesome5(61889);
-  static const IconData file_powerpoint =
-      const FlutterIconData.fontAwesome5(61892);
-  static const IconData file_video = const FlutterIconData.fontAwesome5(61896);
-  static const IconData file_word = const FlutterIconData.fontAwesome5(61890);
-  static const IconData file = const FlutterIconData.fontAwesome5(61787);
-  static const IconData flag = const FlutterIconData.fontAwesome5(61476);
-  static const IconData flushed = const FlutterIconData.fontAwesome5(62841);
-  static const IconData folder_open = const FlutterIconData.fontAwesome5(61564);
-  static const IconData folder = const FlutterIconData.fontAwesome5(61563);
+      FlutterIconData.fontAwesome5(61778);
+  static const IconData caret_square_up = FlutterIconData.fontAwesome5(61777);
+  static const IconData chart_bar = FlutterIconData.fontAwesome5(61568);
+  static const IconData check_circle = FlutterIconData.fontAwesome5(61528);
+  static const IconData check_square = FlutterIconData.fontAwesome5(61770);
+  static const IconData circle = FlutterIconData.fontAwesome5(61713);
+  static const IconData clipboard = FlutterIconData.fontAwesome5(62248);
+  static const IconData clock = FlutterIconData.fontAwesome5(61463);
+  static const IconData clone = FlutterIconData.fontAwesome5(62029);
+  static const IconData closed_captioning = FlutterIconData.fontAwesome5(61962);
+  static const IconData comment_alt = FlutterIconData.fontAwesome5(62074);
+  static const IconData comment_dots = FlutterIconData.fontAwesome5(62637);
+  static const IconData comment = FlutterIconData.fontAwesome5(61557);
+  static const IconData comments = FlutterIconData.fontAwesome5(61574);
+  static const IconData compass = FlutterIconData.fontAwesome5(61774);
+  static const IconData copy = FlutterIconData.fontAwesome5(61637);
+  static const IconData copyright = FlutterIconData.fontAwesome5(61945);
+  static const IconData credit_card = FlutterIconData.fontAwesome5(61597);
+  static const IconData dizzy = FlutterIconData.fontAwesome5(62823);
+  static const IconData dot_circle = FlutterIconData.fontAwesome5(61842);
+  static const IconData edit = FlutterIconData.fontAwesome5(61508);
+  static const IconData envelope_open = FlutterIconData.fontAwesome5(62134);
+  static const IconData envelope = FlutterIconData.fontAwesome5(61664);
+  static const IconData eye_slash = FlutterIconData.fontAwesome5(61552);
+  static const IconData eye = FlutterIconData.fontAwesome5(61550);
+  static const IconData file_alt = FlutterIconData.fontAwesome5(61788);
+  static const IconData file_archive = FlutterIconData.fontAwesome5(61894);
+  static const IconData file_audio = FlutterIconData.fontAwesome5(61895);
+  static const IconData file_code = FlutterIconData.fontAwesome5(61897);
+  static const IconData file_excel = FlutterIconData.fontAwesome5(61891);
+  static const IconData file_image = FlutterIconData.fontAwesome5(61893);
+  static const IconData file_pdf = FlutterIconData.fontAwesome5(61889);
+  static const IconData file_powerpoint = FlutterIconData.fontAwesome5(61892);
+  static const IconData file_video = FlutterIconData.fontAwesome5(61896);
+  static const IconData file_word = FlutterIconData.fontAwesome5(61890);
+  static const IconData file = FlutterIconData.fontAwesome5(61787);
+  static const IconData flag = FlutterIconData.fontAwesome5(61476);
+  static const IconData flushed = FlutterIconData.fontAwesome5(62841);
+  static const IconData folder_open = FlutterIconData.fontAwesome5(61564);
+  static const IconData folder = FlutterIconData.fontAwesome5(61563);
   static const IconData font_awesome_logo_full =
-      const FlutterIconData.fontAwesome5(62694);
-  static const IconData frown_open = const FlutterIconData.fontAwesome5(62842);
-  static const IconData frown = const FlutterIconData.fontAwesome5(61721);
-  static const IconData futbol = const FlutterIconData.fontAwesome5(61923);
-  static const IconData gem = const FlutterIconData.fontAwesome5(62373);
-  static const IconData grimace = const FlutterIconData.fontAwesome5(62847);
-  static const IconData grin_alt = const FlutterIconData.fontAwesome5(62849);
-  static const IconData grin_beam_sweat =
-      const FlutterIconData.fontAwesome5(62851);
-  static const IconData grin_beam = const FlutterIconData.fontAwesome5(62850);
-  static const IconData grin_hearts = const FlutterIconData.fontAwesome5(62852);
-  static const IconData grin_squint_tears =
-      const FlutterIconData.fontAwesome5(62854);
-  static const IconData grin_squint = const FlutterIconData.fontAwesome5(62853);
-  static const IconData grin_stars = const FlutterIconData.fontAwesome5(62855);
-  static const IconData grin_tears = const FlutterIconData.fontAwesome5(62856);
+      FlutterIconData.fontAwesome5(62694);
+  static const IconData frown_open = FlutterIconData.fontAwesome5(62842);
+  static const IconData frown = FlutterIconData.fontAwesome5(61721);
+  static const IconData futbol = FlutterIconData.fontAwesome5(61923);
+  static const IconData gem = FlutterIconData.fontAwesome5(62373);
+  static const IconData grimace = FlutterIconData.fontAwesome5(62847);
+  static const IconData grin_alt = FlutterIconData.fontAwesome5(62849);
+  static const IconData grin_beam_sweat = FlutterIconData.fontAwesome5(62851);
+  static const IconData grin_beam = FlutterIconData.fontAwesome5(62850);
+  static const IconData grin_hearts = FlutterIconData.fontAwesome5(62852);
+  static const IconData grin_squint_tears = FlutterIconData.fontAwesome5(62854);
+  static const IconData grin_squint = FlutterIconData.fontAwesome5(62853);
+  static const IconData grin_stars = FlutterIconData.fontAwesome5(62855);
+  static const IconData grin_tears = FlutterIconData.fontAwesome5(62856);
   static const IconData grin_tongue_squint =
-      const FlutterIconData.fontAwesome5(62858);
-  static const IconData grin_tongue_wink =
-      const FlutterIconData.fontAwesome5(62859);
-  static const IconData grin_tongue = const FlutterIconData.fontAwesome5(62857);
-  static const IconData grin_wink = const FlutterIconData.fontAwesome5(62860);
-  static const IconData grin = const FlutterIconData.fontAwesome5(62848);
-  static const IconData hand_lizard = const FlutterIconData.fontAwesome5(62040);
-  static const IconData hand_paper = const FlutterIconData.fontAwesome5(62038);
-  static const IconData hand_peace = const FlutterIconData.fontAwesome5(62043);
-  static const IconData hand_point_down =
-      const FlutterIconData.fontAwesome5(61607);
-  static const IconData hand_point_left =
-      const FlutterIconData.fontAwesome5(61605);
-  static const IconData hand_point_right =
-      const FlutterIconData.fontAwesome5(61604);
-  static const IconData hand_point_up =
-      const FlutterIconData.fontAwesome5(61606);
-  static const IconData hand_pointer =
-      const FlutterIconData.fontAwesome5(62042);
-  static const IconData hand_rock = const FlutterIconData.fontAwesome5(62037);
-  static const IconData hand_scissors =
-      const FlutterIconData.fontAwesome5(62039);
-  static const IconData hand_spock = const FlutterIconData.fontAwesome5(62041);
-  static const IconData handshake = const FlutterIconData.fontAwesome5(62133);
-  static const IconData hdd = const FlutterIconData.fontAwesome5(61600);
-  static const IconData heart = const FlutterIconData.fontAwesome5(61444);
-  static const IconData hospital = const FlutterIconData.fontAwesome5(61688);
-  static const IconData hourglass = const FlutterIconData.fontAwesome5(62036);
-  static const IconData id_badge = const FlutterIconData.fontAwesome5(62145);
-  static const IconData id_card = const FlutterIconData.fontAwesome5(62146);
-  static const IconData image = const FlutterIconData.fontAwesome5(61502);
-  static const IconData images = const FlutterIconData.fontAwesome5(62210);
-  static const IconData keyboard = const FlutterIconData.fontAwesome5(61724);
-  static const IconData kiss_beam = const FlutterIconData.fontAwesome5(62871);
-  static const IconData kiss_wink_heart =
-      const FlutterIconData.fontAwesome5(62872);
-  static const IconData kiss = const FlutterIconData.fontAwesome5(62870);
-  static const IconData laugh_beam = const FlutterIconData.fontAwesome5(62874);
-  static const IconData laugh_squint =
-      const FlutterIconData.fontAwesome5(62875);
-  static const IconData laugh_wink = const FlutterIconData.fontAwesome5(62876);
-  static const IconData laugh = const FlutterIconData.fontAwesome5(62873);
-  static const IconData lemon = const FlutterIconData.fontAwesome5(61588);
-  static const IconData life_ring = const FlutterIconData.fontAwesome5(61901);
-  static const IconData lightbulb = const FlutterIconData.fontAwesome5(61675);
-  static const IconData list_alt = const FlutterIconData.fontAwesome5(61474);
-  static const IconData map = const FlutterIconData.fontAwesome5(62073);
-  static const IconData meh_blank = const FlutterIconData.fontAwesome5(62884);
-  static const IconData meh_rolling_eyes =
-      const FlutterIconData.fontAwesome5(62885);
-  static const IconData meh = const FlutterIconData.fontAwesome5(61722);
-  static const IconData minus_square =
-      const FlutterIconData.fontAwesome5(61766);
-  static const IconData money_bill_alt =
-      const FlutterIconData.fontAwesome5(62417);
-  static const IconData moon = const FlutterIconData.fontAwesome5(61830);
-  static const IconData newspaper = const FlutterIconData.fontAwesome5(61930);
-  static const IconData object_group =
-      const FlutterIconData.fontAwesome5(62023);
-  static const IconData object_ungroup =
-      const FlutterIconData.fontAwesome5(62024);
-  static const IconData paper_plane = const FlutterIconData.fontAwesome5(61912);
-  static const IconData pause_circle =
-      const FlutterIconData.fontAwesome5(62091);
-  static const IconData play_circle = const FlutterIconData.fontAwesome5(61764);
-  static const IconData plus_square = const FlutterIconData.fontAwesome5(61694);
-  static const IconData question_circle =
-      const FlutterIconData.fontAwesome5(61529);
-  static const IconData registered = const FlutterIconData.fontAwesome5(62045);
-  static const IconData sad_cry = const FlutterIconData.fontAwesome5(62899);
-  static const IconData sad_tear = const FlutterIconData.fontAwesome5(62900);
-  static const IconData save = const FlutterIconData.fontAwesome5(61639);
-  static const IconData share_square =
-      const FlutterIconData.fontAwesome5(61773);
-  static const IconData smile_beam = const FlutterIconData.fontAwesome5(62904);
-  static const IconData smile_wink = const FlutterIconData.fontAwesome5(62682);
-  static const IconData smile = const FlutterIconData.fontAwesome5(61720);
-  static const IconData snowflake = const FlutterIconData.fontAwesome5(62172);
-  static const IconData square = const FlutterIconData.fontAwesome5(61640);
-  static const IconData star_half = const FlutterIconData.fontAwesome5(61577);
-  static const IconData star = const FlutterIconData.fontAwesome5(61445);
-  static const IconData sticky_note = const FlutterIconData.fontAwesome5(62025);
-  static const IconData stop_circle = const FlutterIconData.fontAwesome5(62093);
-  static const IconData sun = const FlutterIconData.fontAwesome5(61829);
-  static const IconData surprise = const FlutterIconData.fontAwesome5(62914);
-  static const IconData thumbs_down = const FlutterIconData.fontAwesome5(61797);
-  static const IconData thumbs_up = const FlutterIconData.fontAwesome5(61796);
-  static const IconData times_circle =
-      const FlutterIconData.fontAwesome5(61527);
-  static const IconData tired = const FlutterIconData.fontAwesome5(62920);
-  static const IconData trash_alt = const FlutterIconData.fontAwesome5(62189);
-  static const IconData user_circle = const FlutterIconData.fontAwesome5(62141);
-  static const IconData user = const FlutterIconData.fontAwesome5(61447);
-  static const IconData window_close =
-      const FlutterIconData.fontAwesome5(62480);
-  static const IconData window_maximize =
-      const FlutterIconData.fontAwesome5(62160);
-  static const IconData window_minimize =
-      const FlutterIconData.fontAwesome5(62161);
-  static const IconData window_restore =
-      const FlutterIconData.fontAwesome5(62162);
+      FlutterIconData.fontAwesome5(62858);
+  static const IconData grin_tongue_wink = FlutterIconData.fontAwesome5(62859);
+  static const IconData grin_tongue = FlutterIconData.fontAwesome5(62857);
+  static const IconData grin_wink = FlutterIconData.fontAwesome5(62860);
+  static const IconData grin = FlutterIconData.fontAwesome5(62848);
+  static const IconData hand_lizard = FlutterIconData.fontAwesome5(62040);
+  static const IconData hand_paper = FlutterIconData.fontAwesome5(62038);
+  static const IconData hand_peace = FlutterIconData.fontAwesome5(62043);
+  static const IconData hand_point_down = FlutterIconData.fontAwesome5(61607);
+  static const IconData hand_point_left = FlutterIconData.fontAwesome5(61605);
+  static const IconData hand_point_right = FlutterIconData.fontAwesome5(61604);
+  static const IconData hand_point_up = FlutterIconData.fontAwesome5(61606);
+  static const IconData hand_pointer = FlutterIconData.fontAwesome5(62042);
+  static const IconData hand_rock = FlutterIconData.fontAwesome5(62037);
+  static const IconData hand_scissors = FlutterIconData.fontAwesome5(62039);
+  static const IconData hand_spock = FlutterIconData.fontAwesome5(62041);
+  static const IconData handshake = FlutterIconData.fontAwesome5(62133);
+  static const IconData hdd = FlutterIconData.fontAwesome5(61600);
+  static const IconData heart = FlutterIconData.fontAwesome5(61444);
+  static const IconData hospital = FlutterIconData.fontAwesome5(61688);
+  static const IconData hourglass = FlutterIconData.fontAwesome5(62036);
+  static const IconData id_badge = FlutterIconData.fontAwesome5(62145);
+  static const IconData id_card = FlutterIconData.fontAwesome5(62146);
+  static const IconData image = FlutterIconData.fontAwesome5(61502);
+  static const IconData images = FlutterIconData.fontAwesome5(62210);
+  static const IconData keyboard = FlutterIconData.fontAwesome5(61724);
+  static const IconData kiss_beam = FlutterIconData.fontAwesome5(62871);
+  static const IconData kiss_wink_heart = FlutterIconData.fontAwesome5(62872);
+  static const IconData kiss = FlutterIconData.fontAwesome5(62870);
+  static const IconData laugh_beam = FlutterIconData.fontAwesome5(62874);
+  static const IconData laugh_squint = FlutterIconData.fontAwesome5(62875);
+  static const IconData laugh_wink = FlutterIconData.fontAwesome5(62876);
+  static const IconData laugh = FlutterIconData.fontAwesome5(62873);
+  static const IconData lemon = FlutterIconData.fontAwesome5(61588);
+  static const IconData life_ring = FlutterIconData.fontAwesome5(61901);
+  static const IconData lightbulb = FlutterIconData.fontAwesome5(61675);
+  static const IconData list_alt = FlutterIconData.fontAwesome5(61474);
+  static const IconData map = FlutterIconData.fontAwesome5(62073);
+  static const IconData meh_blank = FlutterIconData.fontAwesome5(62884);
+  static const IconData meh_rolling_eyes = FlutterIconData.fontAwesome5(62885);
+  static const IconData meh = FlutterIconData.fontAwesome5(61722);
+  static const IconData minus_square = FlutterIconData.fontAwesome5(61766);
+  static const IconData money_bill_alt = FlutterIconData.fontAwesome5(62417);
+  static const IconData moon = FlutterIconData.fontAwesome5(61830);
+  static const IconData newspaper = FlutterIconData.fontAwesome5(61930);
+  static const IconData object_group = FlutterIconData.fontAwesome5(62023);
+  static const IconData object_ungroup = FlutterIconData.fontAwesome5(62024);
+  static const IconData paper_plane = FlutterIconData.fontAwesome5(61912);
+  static const IconData pause_circle = FlutterIconData.fontAwesome5(62091);
+  static const IconData play_circle = FlutterIconData.fontAwesome5(61764);
+  static const IconData plus_square = FlutterIconData.fontAwesome5(61694);
+  static const IconData question_circle = FlutterIconData.fontAwesome5(61529);
+  static const IconData registered = FlutterIconData.fontAwesome5(62045);
+  static const IconData sad_cry = FlutterIconData.fontAwesome5(62899);
+  static const IconData sad_tear = FlutterIconData.fontAwesome5(62900);
+  static const IconData save = FlutterIconData.fontAwesome5(61639);
+  static const IconData share_square = FlutterIconData.fontAwesome5(61773);
+  static const IconData smile_beam = FlutterIconData.fontAwesome5(62904);
+  static const IconData smile_wink = FlutterIconData.fontAwesome5(62682);
+  static const IconData smile = FlutterIconData.fontAwesome5(61720);
+  static const IconData snowflake = FlutterIconData.fontAwesome5(62172);
+  static const IconData square = FlutterIconData.fontAwesome5(61640);
+  static const IconData star_half = FlutterIconData.fontAwesome5(61577);
+  static const IconData star = FlutterIconData.fontAwesome5(61445);
+  static const IconData sticky_note = FlutterIconData.fontAwesome5(62025);
+  static const IconData stop_circle = FlutterIconData.fontAwesome5(62093);
+  static const IconData sun = FlutterIconData.fontAwesome5(61829);
+  static const IconData surprise = FlutterIconData.fontAwesome5(62914);
+  static const IconData thumbs_down = FlutterIconData.fontAwesome5(61797);
+  static const IconData thumbs_up = FlutterIconData.fontAwesome5(61796);
+  static const IconData times_circle = FlutterIconData.fontAwesome5(61527);
+  static const IconData tired = FlutterIconData.fontAwesome5(62920);
+  static const IconData trash_alt = FlutterIconData.fontAwesome5(62189);
+  static const IconData user_circle = FlutterIconData.fontAwesome5(62141);
+  static const IconData user = FlutterIconData.fontAwesome5(61447);
+  static const IconData window_close = FlutterIconData.fontAwesome5(62480);
+  static const IconData window_maximize = FlutterIconData.fontAwesome5(62160);
+  static const IconData window_minimize = FlutterIconData.fontAwesome5(62161);
+  static const IconData window_restore = FlutterIconData.fontAwesome5(62162);
 }
