@@ -6,752 +6,522 @@ import 'flutter_icon_data.dart';
 /// https://fontawesome.com
 class FontAwesome5Brands {
   FontAwesome5Brands._();
-  static const IconData $500px =
-      const FlutterIconData.fontAwesome5Brands(62062);
+
+  static const IconData $500px = FlutterIconData.fontAwesome5Brands(62062);
   static const IconData accessible_icon =
-      const FlutterIconData.fontAwesome5Brands(62312);
-  static const IconData accusoft =
-      const FlutterIconData.fontAwesome5Brands(62313);
+      FlutterIconData.fontAwesome5Brands(62312);
+  static const IconData accusoft = FlutterIconData.fontAwesome5Brands(62313);
   static const IconData acquisitions_incorporated =
-      const FlutterIconData.fontAwesome5Brands(63151);
-  static const IconData adn = const FlutterIconData.fontAwesome5Brands(61808);
-  static const IconData adobe = const FlutterIconData.fontAwesome5Brands(63352);
-  static const IconData adversal =
-      const FlutterIconData.fontAwesome5Brands(62314);
+      FlutterIconData.fontAwesome5Brands(63151);
+  static const IconData adn = FlutterIconData.fontAwesome5Brands(61808);
+  static const IconData adobe = FlutterIconData.fontAwesome5Brands(63352);
+  static const IconData adversal = FlutterIconData.fontAwesome5Brands(62314);
   static const IconData affiliatetheme =
-      const FlutterIconData.fontAwesome5Brands(62315);
-  static const IconData algolia =
-      const FlutterIconData.fontAwesome5Brands(62316);
-  static const IconData alipay =
-      const FlutterIconData.fontAwesome5Brands(63042);
-  static const IconData amazon_pay =
-      const FlutterIconData.fontAwesome5Brands(62508);
-  static const IconData amazon =
-      const FlutterIconData.fontAwesome5Brands(62064);
-  static const IconData amilia =
-      const FlutterIconData.fontAwesome5Brands(62317);
-  static const IconData android =
-      const FlutterIconData.fontAwesome5Brands(61819);
-  static const IconData angellist =
-      const FlutterIconData.fontAwesome5Brands(61961);
+      FlutterIconData.fontAwesome5Brands(62315);
+  static const IconData algolia = FlutterIconData.fontAwesome5Brands(62316);
+  static const IconData alipay = FlutterIconData.fontAwesome5Brands(63042);
+  static const IconData amazon_pay = FlutterIconData.fontAwesome5Brands(62508);
+  static const IconData amazon = FlutterIconData.fontAwesome5Brands(62064);
+  static const IconData amilia = FlutterIconData.fontAwesome5Brands(62317);
+  static const IconData android = FlutterIconData.fontAwesome5Brands(61819);
+  static const IconData angellist = FlutterIconData.fontAwesome5Brands(61961);
   static const IconData angrycreative =
-      const FlutterIconData.fontAwesome5Brands(62318);
-  static const IconData angular =
-      const FlutterIconData.fontAwesome5Brands(62496);
+      FlutterIconData.fontAwesome5Brands(62318);
+  static const IconData angular = FlutterIconData.fontAwesome5Brands(62496);
   static const IconData app_store_ios =
-      const FlutterIconData.fontAwesome5Brands(62320);
-  static const IconData app_store =
-      const FlutterIconData.fontAwesome5Brands(62319);
-  static const IconData apper = const FlutterIconData.fontAwesome5Brands(62321);
-  static const IconData apple_pay =
-      const FlutterIconData.fontAwesome5Brands(62485);
-  static const IconData apple = const FlutterIconData.fontAwesome5Brands(61817);
-  static const IconData artstation =
-      const FlutterIconData.fontAwesome5Brands(63354);
-  static const IconData asymmetrik =
-      const FlutterIconData.fontAwesome5Brands(62322);
-  static const IconData atlassian =
-      const FlutterIconData.fontAwesome5Brands(63355);
-  static const IconData audible =
-      const FlutterIconData.fontAwesome5Brands(62323);
+      FlutterIconData.fontAwesome5Brands(62320);
+  static const IconData app_store = FlutterIconData.fontAwesome5Brands(62319);
+  static const IconData apper = FlutterIconData.fontAwesome5Brands(62321);
+  static const IconData apple_pay = FlutterIconData.fontAwesome5Brands(62485);
+  static const IconData apple = FlutterIconData.fontAwesome5Brands(61817);
+  static const IconData artstation = FlutterIconData.fontAwesome5Brands(63354);
+  static const IconData asymmetrik = FlutterIconData.fontAwesome5Brands(62322);
+  static const IconData atlassian = FlutterIconData.fontAwesome5Brands(63355);
+  static const IconData audible = FlutterIconData.fontAwesome5Brands(62323);
   static const IconData autoprefixer =
-      const FlutterIconData.fontAwesome5Brands(62492);
-  static const IconData avianex =
-      const FlutterIconData.fontAwesome5Brands(62324);
-  static const IconData aviato =
-      const FlutterIconData.fontAwesome5Brands(62497);
-  static const IconData aws = const FlutterIconData.fontAwesome5Brands(62325);
-  static const IconData bandcamp =
-      const FlutterIconData.fontAwesome5Brands(62165);
+      FlutterIconData.fontAwesome5Brands(62492);
+  static const IconData avianex = FlutterIconData.fontAwesome5Brands(62324);
+  static const IconData aviato = FlutterIconData.fontAwesome5Brands(62497);
+  static const IconData aws = FlutterIconData.fontAwesome5Brands(62325);
+  static const IconData bandcamp = FlutterIconData.fontAwesome5Brands(62165);
   static const IconData behance_square =
-      const FlutterIconData.fontAwesome5Brands(61877);
-  static const IconData behance =
-      const FlutterIconData.fontAwesome5Brands(61876);
-  static const IconData bimobject =
-      const FlutterIconData.fontAwesome5Brands(62328);
-  static const IconData bitbucket =
-      const FlutterIconData.fontAwesome5Brands(61809);
-  static const IconData bitcoin =
-      const FlutterIconData.fontAwesome5Brands(62329);
-  static const IconData bity = const FlutterIconData.fontAwesome5Brands(62330);
-  static const IconData black_tie =
-      const FlutterIconData.fontAwesome5Brands(62078);
-  static const IconData blackberry =
-      const FlutterIconData.fontAwesome5Brands(62331);
-  static const IconData blogger_b =
-      const FlutterIconData.fontAwesome5Brands(62333);
-  static const IconData blogger =
-      const FlutterIconData.fontAwesome5Brands(62332);
-  static const IconData bluetooth_b =
-      const FlutterIconData.fontAwesome5Brands(62100);
-  static const IconData bluetooth =
-      const FlutterIconData.fontAwesome5Brands(62099);
-  static const IconData btc = const FlutterIconData.fontAwesome5Brands(61786);
+      FlutterIconData.fontAwesome5Brands(61877);
+  static const IconData behance = FlutterIconData.fontAwesome5Brands(61876);
+  static const IconData bimobject = FlutterIconData.fontAwesome5Brands(62328);
+  static const IconData bitbucket = FlutterIconData.fontAwesome5Brands(61809);
+  static const IconData bitcoin = FlutterIconData.fontAwesome5Brands(62329);
+  static const IconData bity = FlutterIconData.fontAwesome5Brands(62330);
+  static const IconData black_tie = FlutterIconData.fontAwesome5Brands(62078);
+  static const IconData blackberry = FlutterIconData.fontAwesome5Brands(62331);
+  static const IconData blogger_b = FlutterIconData.fontAwesome5Brands(62333);
+  static const IconData blogger = FlutterIconData.fontAwesome5Brands(62332);
+  static const IconData bluetooth_b = FlutterIconData.fontAwesome5Brands(62100);
+  static const IconData bluetooth = FlutterIconData.fontAwesome5Brands(62099);
+  static const IconData btc = FlutterIconData.fontAwesome5Brands(61786);
   static const IconData buromobelexperte =
-      const FlutterIconData.fontAwesome5Brands(62335);
-  static const IconData buysellads =
-      const FlutterIconData.fontAwesome5Brands(61965);
+      FlutterIconData.fontAwesome5Brands(62335);
+  static const IconData buysellads = FlutterIconData.fontAwesome5Brands(61965);
   static const IconData canadian_maple_leaf =
-      const FlutterIconData.fontAwesome5Brands(63365);
+      FlutterIconData.fontAwesome5Brands(63365);
   static const IconData cc_amazon_pay =
-      const FlutterIconData.fontAwesome5Brands(62509);
-  static const IconData cc_amex =
-      const FlutterIconData.fontAwesome5Brands(61939);
+      FlutterIconData.fontAwesome5Brands(62509);
+  static const IconData cc_amex = FlutterIconData.fontAwesome5Brands(61939);
   static const IconData cc_apple_pay =
-      const FlutterIconData.fontAwesome5Brands(62486);
+      FlutterIconData.fontAwesome5Brands(62486);
   static const IconData cc_diners_club =
-      const FlutterIconData.fontAwesome5Brands(62028);
-  static const IconData cc_discover =
-      const FlutterIconData.fontAwesome5Brands(61938);
-  static const IconData cc_jcb =
-      const FlutterIconData.fontAwesome5Brands(62027);
+      FlutterIconData.fontAwesome5Brands(62028);
+  static const IconData cc_discover = FlutterIconData.fontAwesome5Brands(61938);
+  static const IconData cc_jcb = FlutterIconData.fontAwesome5Brands(62027);
   static const IconData cc_mastercard =
-      const FlutterIconData.fontAwesome5Brands(61937);
-  static const IconData cc_paypal =
-      const FlutterIconData.fontAwesome5Brands(61940);
-  static const IconData cc_stripe =
-      const FlutterIconData.fontAwesome5Brands(61941);
-  static const IconData cc_visa =
-      const FlutterIconData.fontAwesome5Brands(61936);
-  static const IconData centercode =
-      const FlutterIconData.fontAwesome5Brands(62336);
-  static const IconData centos =
-      const FlutterIconData.fontAwesome5Brands(63369);
-  static const IconData chrome =
-      const FlutterIconData.fontAwesome5Brands(62056);
-  static const IconData cloudscale =
-      const FlutterIconData.fontAwesome5Brands(62339);
-  static const IconData cloudsmith =
-      const FlutterIconData.fontAwesome5Brands(62340);
+      FlutterIconData.fontAwesome5Brands(61937);
+  static const IconData cc_paypal = FlutterIconData.fontAwesome5Brands(61940);
+  static const IconData cc_stripe = FlutterIconData.fontAwesome5Brands(61941);
+  static const IconData cc_visa = FlutterIconData.fontAwesome5Brands(61936);
+  static const IconData centercode = FlutterIconData.fontAwesome5Brands(62336);
+  static const IconData centos = FlutterIconData.fontAwesome5Brands(63369);
+  static const IconData chrome = FlutterIconData.fontAwesome5Brands(62056);
+  static const IconData cloudscale = FlutterIconData.fontAwesome5Brands(62339);
+  static const IconData cloudsmith = FlutterIconData.fontAwesome5Brands(62340);
   static const IconData cloudversify =
-      const FlutterIconData.fontAwesome5Brands(62341);
-  static const IconData codepen =
-      const FlutterIconData.fontAwesome5Brands(61899);
-  static const IconData codiepie =
-      const FlutterIconData.fontAwesome5Brands(62084);
-  static const IconData confluence =
-      const FlutterIconData.fontAwesome5Brands(63373);
+      FlutterIconData.fontAwesome5Brands(62341);
+  static const IconData codepen = FlutterIconData.fontAwesome5Brands(61899);
+  static const IconData codiepie = FlutterIconData.fontAwesome5Brands(62084);
+  static const IconData confluence = FlutterIconData.fontAwesome5Brands(63373);
   static const IconData connectdevelop =
-      const FlutterIconData.fontAwesome5Brands(61966);
-  static const IconData contao =
-      const FlutterIconData.fontAwesome5Brands(62061);
-  static const IconData cpanel =
-      const FlutterIconData.fontAwesome5Brands(62344);
+      FlutterIconData.fontAwesome5Brands(61966);
+  static const IconData contao = FlutterIconData.fontAwesome5Brands(62061);
+  static const IconData cpanel = FlutterIconData.fontAwesome5Brands(62344);
   static const IconData creative_commons_by =
-      const FlutterIconData.fontAwesome5Brands(62695);
+      FlutterIconData.fontAwesome5Brands(62695);
   static const IconData creative_commons_nc_eu =
-      const FlutterIconData.fontAwesome5Brands(62697);
+      FlutterIconData.fontAwesome5Brands(62697);
   static const IconData creative_commons_nc_jp =
-      const FlutterIconData.fontAwesome5Brands(62698);
+      FlutterIconData.fontAwesome5Brands(62698);
   static const IconData creative_commons_nc =
-      const FlutterIconData.fontAwesome5Brands(62696);
+      FlutterIconData.fontAwesome5Brands(62696);
   static const IconData creative_commons_nd =
-      const FlutterIconData.fontAwesome5Brands(62699);
+      FlutterIconData.fontAwesome5Brands(62699);
   static const IconData creative_commons_pd_alt =
-      const FlutterIconData.fontAwesome5Brands(62701);
+      FlutterIconData.fontAwesome5Brands(62701);
   static const IconData creative_commons_pd =
-      const FlutterIconData.fontAwesome5Brands(62700);
+      FlutterIconData.fontAwesome5Brands(62700);
   static const IconData creative_commons_remix =
-      const FlutterIconData.fontAwesome5Brands(62702);
+      FlutterIconData.fontAwesome5Brands(62702);
   static const IconData creative_commons_sa =
-      const FlutterIconData.fontAwesome5Brands(62703);
+      FlutterIconData.fontAwesome5Brands(62703);
   static const IconData creative_commons_sampling_plus =
-      const FlutterIconData.fontAwesome5Brands(62705);
+      FlutterIconData.fontAwesome5Brands(62705);
   static const IconData creative_commons_sampling =
-      const FlutterIconData.fontAwesome5Brands(62704);
+      FlutterIconData.fontAwesome5Brands(62704);
   static const IconData creative_commons_share =
-      const FlutterIconData.fontAwesome5Brands(62706);
+      FlutterIconData.fontAwesome5Brands(62706);
   static const IconData creative_commons_zero =
-      const FlutterIconData.fontAwesome5Brands(62707);
+      FlutterIconData.fontAwesome5Brands(62707);
   static const IconData creative_commons =
-      const FlutterIconData.fontAwesome5Brands(62046);
+      FlutterIconData.fontAwesome5Brands(62046);
   static const IconData critical_role =
-      const FlutterIconData.fontAwesome5Brands(63177);
-  static const IconData css3_alt =
-      const FlutterIconData.fontAwesome5Brands(62347);
-  static const IconData css3 = const FlutterIconData.fontAwesome5Brands(61756);
-  static const IconData cuttlefish =
-      const FlutterIconData.fontAwesome5Brands(62348);
+      FlutterIconData.fontAwesome5Brands(63177);
+  static const IconData css3_alt = FlutterIconData.fontAwesome5Brands(62347);
+  static const IconData css3 = FlutterIconData.fontAwesome5Brands(61756);
+  static const IconData cuttlefish = FlutterIconData.fontAwesome5Brands(62348);
   static const IconData d_and_d_beyond =
-      const FlutterIconData.fontAwesome5Brands(63178);
-  static const IconData d_and_d =
-      const FlutterIconData.fontAwesome5Brands(62349);
-  static const IconData dashcube =
-      const FlutterIconData.fontAwesome5Brands(61968);
-  static const IconData delicious =
-      const FlutterIconData.fontAwesome5Brands(61861);
-  static const IconData deploydog =
-      const FlutterIconData.fontAwesome5Brands(62350);
-  static const IconData deskpro =
-      const FlutterIconData.fontAwesome5Brands(62351);
-  static const IconData dev = const FlutterIconData.fontAwesome5Brands(63180);
-  static const IconData deviantart =
-      const FlutterIconData.fontAwesome5Brands(61885);
-  static const IconData dhl = const FlutterIconData.fontAwesome5Brands(63376);
-  static const IconData diaspora =
-      const FlutterIconData.fontAwesome5Brands(63377);
-  static const IconData digg = const FlutterIconData.fontAwesome5Brands(61862);
+      FlutterIconData.fontAwesome5Brands(63178);
+  static const IconData d_and_d = FlutterIconData.fontAwesome5Brands(62349);
+  static const IconData dashcube = FlutterIconData.fontAwesome5Brands(61968);
+  static const IconData delicious = FlutterIconData.fontAwesome5Brands(61861);
+  static const IconData deploydog = FlutterIconData.fontAwesome5Brands(62350);
+  static const IconData deskpro = FlutterIconData.fontAwesome5Brands(62351);
+  static const IconData dev = FlutterIconData.fontAwesome5Brands(63180);
+  static const IconData deviantart = FlutterIconData.fontAwesome5Brands(61885);
+  static const IconData dhl = FlutterIconData.fontAwesome5Brands(63376);
+  static const IconData diaspora = FlutterIconData.fontAwesome5Brands(63377);
+  static const IconData digg = FlutterIconData.fontAwesome5Brands(61862);
   static const IconData digital_ocean =
-      const FlutterIconData.fontAwesome5Brands(62353);
-  static const IconData discord =
-      const FlutterIconData.fontAwesome5Brands(62354);
-  static const IconData discourse =
-      const FlutterIconData.fontAwesome5Brands(62355);
-  static const IconData dochub =
-      const FlutterIconData.fontAwesome5Brands(62356);
-  static const IconData docker =
-      const FlutterIconData.fontAwesome5Brands(62357);
+      FlutterIconData.fontAwesome5Brands(62353);
+  static const IconData discord = FlutterIconData.fontAwesome5Brands(62354);
+  static const IconData discourse = FlutterIconData.fontAwesome5Brands(62355);
+  static const IconData dochub = FlutterIconData.fontAwesome5Brands(62356);
+  static const IconData docker = FlutterIconData.fontAwesome5Brands(62357);
   static const IconData draft2digital =
-      const FlutterIconData.fontAwesome5Brands(62358);
+      FlutterIconData.fontAwesome5Brands(62358);
   static const IconData dribbble_square =
-      const FlutterIconData.fontAwesome5Brands(62359);
-  static const IconData dribbble =
-      const FlutterIconData.fontAwesome5Brands(61821);
-  static const IconData dropbox =
-      const FlutterIconData.fontAwesome5Brands(61803);
-  static const IconData drupal =
-      const FlutterIconData.fontAwesome5Brands(61865);
-  static const IconData dyalog =
-      const FlutterIconData.fontAwesome5Brands(62361);
-  static const IconData earlybirds =
-      const FlutterIconData.fontAwesome5Brands(62362);
-  static const IconData ebay = const FlutterIconData.fontAwesome5Brands(62708);
-  static const IconData edge = const FlutterIconData.fontAwesome5Brands(62082);
-  static const IconData elementor =
-      const FlutterIconData.fontAwesome5Brands(62512);
-  static const IconData ello = const FlutterIconData.fontAwesome5Brands(62961);
-  static const IconData ember = const FlutterIconData.fontAwesome5Brands(62499);
-  static const IconData empire =
-      const FlutterIconData.fontAwesome5Brands(61905);
-  static const IconData envira =
-      const FlutterIconData.fontAwesome5Brands(62105);
-  static const IconData erlang =
-      const FlutterIconData.fontAwesome5Brands(62365);
-  static const IconData ethereum =
-      const FlutterIconData.fontAwesome5Brands(62510);
-  static const IconData etsy = const FlutterIconData.fontAwesome5Brands(62167);
+      FlutterIconData.fontAwesome5Brands(62359);
+  static const IconData dribbble = FlutterIconData.fontAwesome5Brands(61821);
+  static const IconData dropbox = FlutterIconData.fontAwesome5Brands(61803);
+  static const IconData drupal = FlutterIconData.fontAwesome5Brands(61865);
+  static const IconData dyalog = FlutterIconData.fontAwesome5Brands(62361);
+  static const IconData earlybirds = FlutterIconData.fontAwesome5Brands(62362);
+  static const IconData ebay = FlutterIconData.fontAwesome5Brands(62708);
+  static const IconData edge = FlutterIconData.fontAwesome5Brands(62082);
+  static const IconData elementor = FlutterIconData.fontAwesome5Brands(62512);
+  static const IconData ello = FlutterIconData.fontAwesome5Brands(62961);
+  static const IconData ember = FlutterIconData.fontAwesome5Brands(62499);
+  static const IconData empire = FlutterIconData.fontAwesome5Brands(61905);
+  static const IconData envira = FlutterIconData.fontAwesome5Brands(62105);
+  static const IconData erlang = FlutterIconData.fontAwesome5Brands(62365);
+  static const IconData ethereum = FlutterIconData.fontAwesome5Brands(62510);
+  static const IconData etsy = FlutterIconData.fontAwesome5Brands(62167);
   static const IconData expeditedssl =
-      const FlutterIconData.fontAwesome5Brands(62014);
-  static const IconData facebook_f =
-      const FlutterIconData.fontAwesome5Brands(62366);
+      FlutterIconData.fontAwesome5Brands(62014);
+  static const IconData facebook_f = FlutterIconData.fontAwesome5Brands(62366);
   static const IconData facebook_messenger =
-      const FlutterIconData.fontAwesome5Brands(62367);
+      FlutterIconData.fontAwesome5Brands(62367);
   static const IconData facebook_square =
-      const FlutterIconData.fontAwesome5Brands(61570);
-  static const IconData facebook =
-      const FlutterIconData.fontAwesome5Brands(61594);
+      FlutterIconData.fontAwesome5Brands(61570);
+  static const IconData facebook = FlutterIconData.fontAwesome5Brands(61594);
   static const IconData fantasy_flight_games =
-      const FlutterIconData.fontAwesome5Brands(63196);
-  static const IconData fedex = const FlutterIconData.fontAwesome5Brands(63383);
-  static const IconData fedora =
-      const FlutterIconData.fontAwesome5Brands(63384);
-  static const IconData figma = const FlutterIconData.fontAwesome5Brands(63385);
-  static const IconData firefox =
-      const FlutterIconData.fontAwesome5Brands(62057);
+      FlutterIconData.fontAwesome5Brands(63196);
+  static const IconData fedex = FlutterIconData.fontAwesome5Brands(63383);
+  static const IconData fedora = FlutterIconData.fontAwesome5Brands(63384);
+  static const IconData figma = FlutterIconData.fontAwesome5Brands(63385);
+  static const IconData firefox = FlutterIconData.fontAwesome5Brands(62057);
   static const IconData first_order_alt =
-      const FlutterIconData.fontAwesome5Brands(62730);
-  static const IconData first_order =
-      const FlutterIconData.fontAwesome5Brands(62128);
-  static const IconData firstdraft =
-      const FlutterIconData.fontAwesome5Brands(62369);
-  static const IconData flickr =
-      const FlutterIconData.fontAwesome5Brands(61806);
-  static const IconData flipboard =
-      const FlutterIconData.fontAwesome5Brands(62541);
-  static const IconData fly = const FlutterIconData.fontAwesome5Brands(62487);
+      FlutterIconData.fontAwesome5Brands(62730);
+  static const IconData first_order = FlutterIconData.fontAwesome5Brands(62128);
+  static const IconData firstdraft = FlutterIconData.fontAwesome5Brands(62369);
+  static const IconData flickr = FlutterIconData.fontAwesome5Brands(61806);
+  static const IconData flipboard = FlutterIconData.fontAwesome5Brands(62541);
+  static const IconData fly = FlutterIconData.fontAwesome5Brands(62487);
   static const IconData font_awesome_alt =
-      const FlutterIconData.fontAwesome5Brands(62300);
+      FlutterIconData.fontAwesome5Brands(62300);
   static const IconData font_awesome_flag =
-      const FlutterIconData.fontAwesome5Brands(62501);
+      FlutterIconData.fontAwesome5Brands(62501);
   static const IconData font_awesome_logo_full =
-      const FlutterIconData.fontAwesome5Brands(62694);
+      FlutterIconData.fontAwesome5Brands(62694);
   static const IconData font_awesome =
-      const FlutterIconData.fontAwesome5Brands(62132);
+      FlutterIconData.fontAwesome5Brands(62132);
   static const IconData fonticons_fi =
-      const FlutterIconData.fontAwesome5Brands(62370);
-  static const IconData fonticons =
-      const FlutterIconData.fontAwesome5Brands(62080);
+      FlutterIconData.fontAwesome5Brands(62370);
+  static const IconData fonticons = FlutterIconData.fontAwesome5Brands(62080);
   static const IconData fort_awesome_alt =
-      const FlutterIconData.fontAwesome5Brands(62371);
+      FlutterIconData.fontAwesome5Brands(62371);
   static const IconData fort_awesome =
-      const FlutterIconData.fontAwesome5Brands(62086);
-  static const IconData forumbee =
-      const FlutterIconData.fontAwesome5Brands(61969);
-  static const IconData foursquare =
-      const FlutterIconData.fontAwesome5Brands(61824);
+      FlutterIconData.fontAwesome5Brands(62086);
+  static const IconData forumbee = FlutterIconData.fontAwesome5Brands(61969);
+  static const IconData foursquare = FlutterIconData.fontAwesome5Brands(61824);
   static const IconData free_code_camp =
-      const FlutterIconData.fontAwesome5Brands(62149);
-  static const IconData freebsd =
-      const FlutterIconData.fontAwesome5Brands(62372);
-  static const IconData fulcrum =
-      const FlutterIconData.fontAwesome5Brands(62731);
+      FlutterIconData.fontAwesome5Brands(62149);
+  static const IconData freebsd = FlutterIconData.fontAwesome5Brands(62372);
+  static const IconData fulcrum = FlutterIconData.fontAwesome5Brands(62731);
   static const IconData galactic_republic =
-      const FlutterIconData.fontAwesome5Brands(62732);
+      FlutterIconData.fontAwesome5Brands(62732);
   static const IconData galactic_senate =
-      const FlutterIconData.fontAwesome5Brands(62733);
-  static const IconData get_pocket =
-      const FlutterIconData.fontAwesome5Brands(62053);
-  static const IconData gg_circle =
-      const FlutterIconData.fontAwesome5Brands(62049);
-  static const IconData gg = const FlutterIconData.fontAwesome5Brands(62048);
-  static const IconData git_square =
-      const FlutterIconData.fontAwesome5Brands(61906);
-  static const IconData git = const FlutterIconData.fontAwesome5Brands(61907);
-  static const IconData github_alt =
-      const FlutterIconData.fontAwesome5Brands(61715);
+      FlutterIconData.fontAwesome5Brands(62733);
+  static const IconData get_pocket = FlutterIconData.fontAwesome5Brands(62053);
+  static const IconData gg_circle = FlutterIconData.fontAwesome5Brands(62049);
+  static const IconData gg = FlutterIconData.fontAwesome5Brands(62048);
+  static const IconData git_square = FlutterIconData.fontAwesome5Brands(61906);
+  static const IconData git = FlutterIconData.fontAwesome5Brands(61907);
+  static const IconData github_alt = FlutterIconData.fontAwesome5Brands(61715);
   static const IconData github_square =
-      const FlutterIconData.fontAwesome5Brands(61586);
-  static const IconData github =
-      const FlutterIconData.fontAwesome5Brands(61595);
-  static const IconData gitkraken =
-      const FlutterIconData.fontAwesome5Brands(62374);
-  static const IconData gitlab =
-      const FlutterIconData.fontAwesome5Brands(62102);
-  static const IconData gitter =
-      const FlutterIconData.fontAwesome5Brands(62502);
-  static const IconData glide_g =
-      const FlutterIconData.fontAwesome5Brands(62118);
-  static const IconData glide = const FlutterIconData.fontAwesome5Brands(62117);
-  static const IconData gofore =
-      const FlutterIconData.fontAwesome5Brands(62375);
-  static const IconData goodreads_g =
-      const FlutterIconData.fontAwesome5Brands(62377);
-  static const IconData goodreads =
-      const FlutterIconData.fontAwesome5Brands(62376);
+      FlutterIconData.fontAwesome5Brands(61586);
+  static const IconData github = FlutterIconData.fontAwesome5Brands(61595);
+  static const IconData gitkraken = FlutterIconData.fontAwesome5Brands(62374);
+  static const IconData gitlab = FlutterIconData.fontAwesome5Brands(62102);
+  static const IconData gitter = FlutterIconData.fontAwesome5Brands(62502);
+  static const IconData glide_g = FlutterIconData.fontAwesome5Brands(62118);
+  static const IconData glide = FlutterIconData.fontAwesome5Brands(62117);
+  static const IconData gofore = FlutterIconData.fontAwesome5Brands(62375);
+  static const IconData goodreads_g = FlutterIconData.fontAwesome5Brands(62377);
+  static const IconData goodreads = FlutterIconData.fontAwesome5Brands(62376);
   static const IconData google_drive =
-      const FlutterIconData.fontAwesome5Brands(62378);
-  static const IconData google_play =
-      const FlutterIconData.fontAwesome5Brands(62379);
+      FlutterIconData.fontAwesome5Brands(62378);
+  static const IconData google_play = FlutterIconData.fontAwesome5Brands(62379);
   static const IconData google_plus_g =
-      const FlutterIconData.fontAwesome5Brands(61653);
+      FlutterIconData.fontAwesome5Brands(61653);
   static const IconData google_plus_square =
-      const FlutterIconData.fontAwesome5Brands(61652);
-  static const IconData google_plus =
-      const FlutterIconData.fontAwesome5Brands(62131);
+      FlutterIconData.fontAwesome5Brands(61652);
+  static const IconData google_plus = FlutterIconData.fontAwesome5Brands(62131);
   static const IconData google_wallet =
-      const FlutterIconData.fontAwesome5Brands(61934);
-  static const IconData google =
-      const FlutterIconData.fontAwesome5Brands(61856);
-  static const IconData gratipay =
-      const FlutterIconData.fontAwesome5Brands(61828);
-  static const IconData grav = const FlutterIconData.fontAwesome5Brands(62166);
-  static const IconData gripfire =
-      const FlutterIconData.fontAwesome5Brands(62380);
-  static const IconData grunt = const FlutterIconData.fontAwesome5Brands(62381);
-  static const IconData gulp = const FlutterIconData.fontAwesome5Brands(62382);
+      FlutterIconData.fontAwesome5Brands(61934);
+  static const IconData google = FlutterIconData.fontAwesome5Brands(61856);
+  static const IconData gratipay = FlutterIconData.fontAwesome5Brands(61828);
+  static const IconData grav = FlutterIconData.fontAwesome5Brands(62166);
+  static const IconData gripfire = FlutterIconData.fontAwesome5Brands(62380);
+  static const IconData grunt = FlutterIconData.fontAwesome5Brands(62381);
+  static const IconData gulp = FlutterIconData.fontAwesome5Brands(62382);
   static const IconData hacker_news_square =
-      const FlutterIconData.fontAwesome5Brands(62383);
-  static const IconData hacker_news =
-      const FlutterIconData.fontAwesome5Brands(61908);
-  static const IconData hackerrank =
-      const FlutterIconData.fontAwesome5Brands(62967);
-  static const IconData hips = const FlutterIconData.fontAwesome5Brands(62546);
+      FlutterIconData.fontAwesome5Brands(62383);
+  static const IconData hacker_news = FlutterIconData.fontAwesome5Brands(61908);
+  static const IconData hackerrank = FlutterIconData.fontAwesome5Brands(62967);
+  static const IconData hips = FlutterIconData.fontAwesome5Brands(62546);
   static const IconData hire_a_helper =
-      const FlutterIconData.fontAwesome5Brands(62384);
-  static const IconData hooli = const FlutterIconData.fontAwesome5Brands(62503);
-  static const IconData hornbill =
-      const FlutterIconData.fontAwesome5Brands(62866);
-  static const IconData hotjar =
-      const FlutterIconData.fontAwesome5Brands(62385);
-  static const IconData houzz = const FlutterIconData.fontAwesome5Brands(62076);
-  static const IconData html5 = const FlutterIconData.fontAwesome5Brands(61755);
-  static const IconData hubspot =
-      const FlutterIconData.fontAwesome5Brands(62386);
-  static const IconData imdb = const FlutterIconData.fontAwesome5Brands(62168);
-  static const IconData instagram =
-      const FlutterIconData.fontAwesome5Brands(61805);
-  static const IconData intercom =
-      const FlutterIconData.fontAwesome5Brands(63407);
+      FlutterIconData.fontAwesome5Brands(62384);
+  static const IconData hooli = FlutterIconData.fontAwesome5Brands(62503);
+  static const IconData hornbill = FlutterIconData.fontAwesome5Brands(62866);
+  static const IconData hotjar = FlutterIconData.fontAwesome5Brands(62385);
+  static const IconData houzz = FlutterIconData.fontAwesome5Brands(62076);
+  static const IconData html5 = FlutterIconData.fontAwesome5Brands(61755);
+  static const IconData hubspot = FlutterIconData.fontAwesome5Brands(62386);
+  static const IconData imdb = FlutterIconData.fontAwesome5Brands(62168);
+  static const IconData instagram = FlutterIconData.fontAwesome5Brands(61805);
+  static const IconData intercom = FlutterIconData.fontAwesome5Brands(63407);
   static const IconData internet_explorer =
-      const FlutterIconData.fontAwesome5Brands(62059);
-  static const IconData invision =
-      const FlutterIconData.fontAwesome5Brands(63408);
-  static const IconData ioxhost =
-      const FlutterIconData.fontAwesome5Brands(61960);
-  static const IconData itunes_note =
-      const FlutterIconData.fontAwesome5Brands(62389);
-  static const IconData itunes =
-      const FlutterIconData.fontAwesome5Brands(62388);
-  static const IconData java = const FlutterIconData.fontAwesome5Brands(62692);
-  static const IconData jedi_order =
-      const FlutterIconData.fontAwesome5Brands(62734);
-  static const IconData jenkins =
-      const FlutterIconData.fontAwesome5Brands(62390);
-  static const IconData jira = const FlutterIconData.fontAwesome5Brands(63409);
-  static const IconData joget = const FlutterIconData.fontAwesome5Brands(62391);
-  static const IconData joomla =
-      const FlutterIconData.fontAwesome5Brands(61866);
-  static const IconData js_square =
-      const FlutterIconData.fontAwesome5Brands(62393);
-  static const IconData js = const FlutterIconData.fontAwesome5Brands(62392);
-  static const IconData jsfiddle =
-      const FlutterIconData.fontAwesome5Brands(61900);
-  static const IconData kaggle =
-      const FlutterIconData.fontAwesome5Brands(62970);
-  static const IconData keybase =
-      const FlutterIconData.fontAwesome5Brands(62709);
-  static const IconData keycdn =
-      const FlutterIconData.fontAwesome5Brands(62394);
+      FlutterIconData.fontAwesome5Brands(62059);
+  static const IconData invision = FlutterIconData.fontAwesome5Brands(63408);
+  static const IconData ioxhost = FlutterIconData.fontAwesome5Brands(61960);
+  static const IconData itunes_note = FlutterIconData.fontAwesome5Brands(62389);
+  static const IconData itunes = FlutterIconData.fontAwesome5Brands(62388);
+  static const IconData java = FlutterIconData.fontAwesome5Brands(62692);
+  static const IconData jedi_order = FlutterIconData.fontAwesome5Brands(62734);
+  static const IconData jenkins = FlutterIconData.fontAwesome5Brands(62390);
+  static const IconData jira = FlutterIconData.fontAwesome5Brands(63409);
+  static const IconData joget = FlutterIconData.fontAwesome5Brands(62391);
+  static const IconData joomla = FlutterIconData.fontAwesome5Brands(61866);
+  static const IconData js_square = FlutterIconData.fontAwesome5Brands(62393);
+  static const IconData js = FlutterIconData.fontAwesome5Brands(62392);
+  static const IconData jsfiddle = FlutterIconData.fontAwesome5Brands(61900);
+  static const IconData kaggle = FlutterIconData.fontAwesome5Brands(62970);
+  static const IconData keybase = FlutterIconData.fontAwesome5Brands(62709);
+  static const IconData keycdn = FlutterIconData.fontAwesome5Brands(62394);
   static const IconData kickstarter_k =
-      const FlutterIconData.fontAwesome5Brands(62396);
-  static const IconData kickstarter =
-      const FlutterIconData.fontAwesome5Brands(62395);
-  static const IconData korvue =
-      const FlutterIconData.fontAwesome5Brands(62511);
-  static const IconData laravel =
-      const FlutterIconData.fontAwesome5Brands(62397);
+      FlutterIconData.fontAwesome5Brands(62396);
+  static const IconData kickstarter = FlutterIconData.fontAwesome5Brands(62395);
+  static const IconData korvue = FlutterIconData.fontAwesome5Brands(62511);
+  static const IconData laravel = FlutterIconData.fontAwesome5Brands(62397);
   static const IconData lastfm_square =
-      const FlutterIconData.fontAwesome5Brands(61955);
-  static const IconData lastfm =
-      const FlutterIconData.fontAwesome5Brands(61954);
-  static const IconData leanpub =
-      const FlutterIconData.fontAwesome5Brands(61970);
-  static const IconData less = const FlutterIconData.fontAwesome5Brands(62493);
-  static const IconData line = const FlutterIconData.fontAwesome5Brands(62400);
-  static const IconData linkedin_in =
-      const FlutterIconData.fontAwesome5Brands(61665);
-  static const IconData linkedin =
-      const FlutterIconData.fontAwesome5Brands(61580);
-  static const IconData linode =
-      const FlutterIconData.fontAwesome5Brands(62136);
-  static const IconData linux = const FlutterIconData.fontAwesome5Brands(61820);
-  static const IconData lyft = const FlutterIconData.fontAwesome5Brands(62403);
-  static const IconData magento =
-      const FlutterIconData.fontAwesome5Brands(62404);
-  static const IconData mailchimp =
-      const FlutterIconData.fontAwesome5Brands(62878);
-  static const IconData mandalorian =
-      const FlutterIconData.fontAwesome5Brands(62735);
-  static const IconData markdown =
-      const FlutterIconData.fontAwesome5Brands(62991);
-  static const IconData mastodon =
-      const FlutterIconData.fontAwesome5Brands(62710);
-  static const IconData maxcdn =
-      const FlutterIconData.fontAwesome5Brands(61750);
-  static const IconData medapps =
-      const FlutterIconData.fontAwesome5Brands(62406);
-  static const IconData medium_m =
-      const FlutterIconData.fontAwesome5Brands(62407);
-  static const IconData medium =
-      const FlutterIconData.fontAwesome5Brands(62010);
-  static const IconData medrt = const FlutterIconData.fontAwesome5Brands(62408);
-  static const IconData meetup =
-      const FlutterIconData.fontAwesome5Brands(62176);
-  static const IconData megaport =
-      const FlutterIconData.fontAwesome5Brands(62883);
-  static const IconData mendeley =
-      const FlutterIconData.fontAwesome5Brands(63411);
-  static const IconData microsoft =
-      const FlutterIconData.fontAwesome5Brands(62410);
-  static const IconData mix = const FlutterIconData.fontAwesome5Brands(62411);
-  static const IconData mixcloud =
-      const FlutterIconData.fontAwesome5Brands(62089);
-  static const IconData mizuni =
-      const FlutterIconData.fontAwesome5Brands(62412);
-  static const IconData modx = const FlutterIconData.fontAwesome5Brands(62085);
-  static const IconData monero =
-      const FlutterIconData.fontAwesome5Brands(62416);
-  static const IconData napster =
-      const FlutterIconData.fontAwesome5Brands(62418);
-  static const IconData neos = const FlutterIconData.fontAwesome5Brands(62994);
-  static const IconData nimblr =
-      const FlutterIconData.fontAwesome5Brands(62888);
+      FlutterIconData.fontAwesome5Brands(61955);
+  static const IconData lastfm = FlutterIconData.fontAwesome5Brands(61954);
+  static const IconData leanpub = FlutterIconData.fontAwesome5Brands(61970);
+  static const IconData less = FlutterIconData.fontAwesome5Brands(62493);
+  static const IconData line = FlutterIconData.fontAwesome5Brands(62400);
+  static const IconData linkedin_in = FlutterIconData.fontAwesome5Brands(61665);
+  static const IconData linkedin = FlutterIconData.fontAwesome5Brands(61580);
+  static const IconData linode = FlutterIconData.fontAwesome5Brands(62136);
+  static const IconData linux = FlutterIconData.fontAwesome5Brands(61820);
+  static const IconData lyft = FlutterIconData.fontAwesome5Brands(62403);
+  static const IconData magento = FlutterIconData.fontAwesome5Brands(62404);
+  static const IconData mailchimp = FlutterIconData.fontAwesome5Brands(62878);
+  static const IconData mandalorian = FlutterIconData.fontAwesome5Brands(62735);
+  static const IconData markdown = FlutterIconData.fontAwesome5Brands(62991);
+  static const IconData mastodon = FlutterIconData.fontAwesome5Brands(62710);
+  static const IconData maxcdn = FlutterIconData.fontAwesome5Brands(61750);
+  static const IconData medapps = FlutterIconData.fontAwesome5Brands(62406);
+  static const IconData medium_m = FlutterIconData.fontAwesome5Brands(62407);
+  static const IconData medium = FlutterIconData.fontAwesome5Brands(62010);
+  static const IconData medrt = FlutterIconData.fontAwesome5Brands(62408);
+  static const IconData meetup = FlutterIconData.fontAwesome5Brands(62176);
+  static const IconData megaport = FlutterIconData.fontAwesome5Brands(62883);
+  static const IconData mendeley = FlutterIconData.fontAwesome5Brands(63411);
+  static const IconData microsoft = FlutterIconData.fontAwesome5Brands(62410);
+  static const IconData mix = FlutterIconData.fontAwesome5Brands(62411);
+  static const IconData mixcloud = FlutterIconData.fontAwesome5Brands(62089);
+  static const IconData mizuni = FlutterIconData.fontAwesome5Brands(62412);
+  static const IconData modx = FlutterIconData.fontAwesome5Brands(62085);
+  static const IconData monero = FlutterIconData.fontAwesome5Brands(62416);
+  static const IconData napster = FlutterIconData.fontAwesome5Brands(62418);
+  static const IconData neos = FlutterIconData.fontAwesome5Brands(62994);
+  static const IconData nimblr = FlutterIconData.fontAwesome5Brands(62888);
   static const IconData nintendo_switch =
-      const FlutterIconData.fontAwesome5Brands(62488);
-  static const IconData node_js =
-      const FlutterIconData.fontAwesome5Brands(62419);
-  static const IconData node = const FlutterIconData.fontAwesome5Brands(62489);
-  static const IconData npm = const FlutterIconData.fontAwesome5Brands(62420);
-  static const IconData ns8 = const FlutterIconData.fontAwesome5Brands(62421);
-  static const IconData nutritionix =
-      const FlutterIconData.fontAwesome5Brands(62422);
+      FlutterIconData.fontAwesome5Brands(62488);
+  static const IconData node_js = FlutterIconData.fontAwesome5Brands(62419);
+  static const IconData node = FlutterIconData.fontAwesome5Brands(62489);
+  static const IconData npm = FlutterIconData.fontAwesome5Brands(62420);
+  static const IconData ns8 = FlutterIconData.fontAwesome5Brands(62421);
+  static const IconData nutritionix = FlutterIconData.fontAwesome5Brands(62422);
   static const IconData odnoklassniki_square =
-      const FlutterIconData.fontAwesome5Brands(62052);
+      FlutterIconData.fontAwesome5Brands(62052);
   static const IconData odnoklassniki =
-      const FlutterIconData.fontAwesome5Brands(62051);
+      FlutterIconData.fontAwesome5Brands(62051);
   static const IconData old_republic =
-      const FlutterIconData.fontAwesome5Brands(62736);
-  static const IconData opencart =
-      const FlutterIconData.fontAwesome5Brands(62013);
-  static const IconData openid =
-      const FlutterIconData.fontAwesome5Brands(61851);
-  static const IconData opera = const FlutterIconData.fontAwesome5Brands(62058);
+      FlutterIconData.fontAwesome5Brands(62736);
+  static const IconData opencart = FlutterIconData.fontAwesome5Brands(62013);
+  static const IconData openid = FlutterIconData.fontAwesome5Brands(61851);
+  static const IconData opera = FlutterIconData.fontAwesome5Brands(62058);
   static const IconData optin_monster =
-      const FlutterIconData.fontAwesome5Brands(62012);
-  static const IconData osi = const FlutterIconData.fontAwesome5Brands(62490);
-  static const IconData page4 = const FlutterIconData.fontAwesome5Brands(62423);
-  static const IconData pagelines =
-      const FlutterIconData.fontAwesome5Brands(61836);
-  static const IconData palfed =
-      const FlutterIconData.fontAwesome5Brands(62424);
-  static const IconData patreon =
-      const FlutterIconData.fontAwesome5Brands(62425);
-  static const IconData paypal =
-      const FlutterIconData.fontAwesome5Brands(61933);
+      FlutterIconData.fontAwesome5Brands(62012);
+  static const IconData osi = FlutterIconData.fontAwesome5Brands(62490);
+  static const IconData page4 = FlutterIconData.fontAwesome5Brands(62423);
+  static const IconData pagelines = FlutterIconData.fontAwesome5Brands(61836);
+  static const IconData palfed = FlutterIconData.fontAwesome5Brands(62424);
+  static const IconData patreon = FlutterIconData.fontAwesome5Brands(62425);
+  static const IconData paypal = FlutterIconData.fontAwesome5Brands(61933);
   static const IconData penny_arcade =
-      const FlutterIconData.fontAwesome5Brands(63236);
-  static const IconData periscope =
-      const FlutterIconData.fontAwesome5Brands(62426);
-  static const IconData phabricator =
-      const FlutterIconData.fontAwesome5Brands(62427);
+      FlutterIconData.fontAwesome5Brands(63236);
+  static const IconData periscope = FlutterIconData.fontAwesome5Brands(62426);
+  static const IconData phabricator = FlutterIconData.fontAwesome5Brands(62427);
   static const IconData phoenix_framework =
-      const FlutterIconData.fontAwesome5Brands(62428);
+      FlutterIconData.fontAwesome5Brands(62428);
   static const IconData phoenix_squadron =
-      const FlutterIconData.fontAwesome5Brands(62737);
-  static const IconData php = const FlutterIconData.fontAwesome5Brands(62551);
+      FlutterIconData.fontAwesome5Brands(62737);
+  static const IconData php = FlutterIconData.fontAwesome5Brands(62551);
   static const IconData pied_piper_alt =
-      const FlutterIconData.fontAwesome5Brands(61864);
+      FlutterIconData.fontAwesome5Brands(61864);
   static const IconData pied_piper_hat =
-      const FlutterIconData.fontAwesome5Brands(62693);
+      FlutterIconData.fontAwesome5Brands(62693);
   static const IconData pied_piper_pp =
-      const FlutterIconData.fontAwesome5Brands(61863);
-  static const IconData pied_piper =
-      const FlutterIconData.fontAwesome5Brands(62126);
-  static const IconData pinterest_p =
-      const FlutterIconData.fontAwesome5Brands(62001);
+      FlutterIconData.fontAwesome5Brands(61863);
+  static const IconData pied_piper = FlutterIconData.fontAwesome5Brands(62126);
+  static const IconData pinterest_p = FlutterIconData.fontAwesome5Brands(62001);
   static const IconData pinterest_square =
-      const FlutterIconData.fontAwesome5Brands(61651);
-  static const IconData pinterest =
-      const FlutterIconData.fontAwesome5Brands(61650);
-  static const IconData playstation =
-      const FlutterIconData.fontAwesome5Brands(62431);
+      FlutterIconData.fontAwesome5Brands(61651);
+  static const IconData pinterest = FlutterIconData.fontAwesome5Brands(61650);
+  static const IconData playstation = FlutterIconData.fontAwesome5Brands(62431);
   static const IconData product_hunt =
-      const FlutterIconData.fontAwesome5Brands(62088);
-  static const IconData pushed =
-      const FlutterIconData.fontAwesome5Brands(62433);
-  static const IconData python =
-      const FlutterIconData.fontAwesome5Brands(62434);
-  static const IconData qq = const FlutterIconData.fontAwesome5Brands(61910);
-  static const IconData quinscape =
-      const FlutterIconData.fontAwesome5Brands(62553);
-  static const IconData quora = const FlutterIconData.fontAwesome5Brands(62148);
-  static const IconData r_project =
-      const FlutterIconData.fontAwesome5Brands(62711);
+      FlutterIconData.fontAwesome5Brands(62088);
+  static const IconData pushed = FlutterIconData.fontAwesome5Brands(62433);
+  static const IconData python = FlutterIconData.fontAwesome5Brands(62434);
+  static const IconData qq = FlutterIconData.fontAwesome5Brands(61910);
+  static const IconData quinscape = FlutterIconData.fontAwesome5Brands(62553);
+  static const IconData quora = FlutterIconData.fontAwesome5Brands(62148);
+  static const IconData r_project = FlutterIconData.fontAwesome5Brands(62711);
   static const IconData raspberry_pi =
-      const FlutterIconData.fontAwesome5Brands(63419);
-  static const IconData ravelry =
-      const FlutterIconData.fontAwesome5Brands(62169);
-  static const IconData react = const FlutterIconData.fontAwesome5Brands(62491);
-  static const IconData reacteurope =
-      const FlutterIconData.fontAwesome5Brands(63325);
-  static const IconData readme =
-      const FlutterIconData.fontAwesome5Brands(62677);
-  static const IconData rebel = const FlutterIconData.fontAwesome5Brands(61904);
-  static const IconData red_river =
-      const FlutterIconData.fontAwesome5Brands(62435);
+      FlutterIconData.fontAwesome5Brands(63419);
+  static const IconData ravelry = FlutterIconData.fontAwesome5Brands(62169);
+  static const IconData react = FlutterIconData.fontAwesome5Brands(62491);
+  static const IconData reacteurope = FlutterIconData.fontAwesome5Brands(63325);
+  static const IconData readme = FlutterIconData.fontAwesome5Brands(62677);
+  static const IconData rebel = FlutterIconData.fontAwesome5Brands(61904);
+  static const IconData red_river = FlutterIconData.fontAwesome5Brands(62435);
   static const IconData reddit_alien =
-      const FlutterIconData.fontAwesome5Brands(62081);
+      FlutterIconData.fontAwesome5Brands(62081);
   static const IconData reddit_square =
-      const FlutterIconData.fontAwesome5Brands(61858);
-  static const IconData reddit =
-      const FlutterIconData.fontAwesome5Brands(61857);
-  static const IconData redhat =
-      const FlutterIconData.fontAwesome5Brands(63420);
-  static const IconData renren =
-      const FlutterIconData.fontAwesome5Brands(61835);
-  static const IconData replyd =
-      const FlutterIconData.fontAwesome5Brands(62438);
+      FlutterIconData.fontAwesome5Brands(61858);
+  static const IconData reddit = FlutterIconData.fontAwesome5Brands(61857);
+  static const IconData redhat = FlutterIconData.fontAwesome5Brands(63420);
+  static const IconData renren = FlutterIconData.fontAwesome5Brands(61835);
+  static const IconData replyd = FlutterIconData.fontAwesome5Brands(62438);
   static const IconData researchgate =
-      const FlutterIconData.fontAwesome5Brands(62712);
-  static const IconData resolving =
-      const FlutterIconData.fontAwesome5Brands(62439);
-  static const IconData rev = const FlutterIconData.fontAwesome5Brands(62898);
-  static const IconData rocketchat =
-      const FlutterIconData.fontAwesome5Brands(62440);
-  static const IconData rockrms =
-      const FlutterIconData.fontAwesome5Brands(62441);
-  static const IconData safari =
-      const FlutterIconData.fontAwesome5Brands(62055);
-  static const IconData sass = const FlutterIconData.fontAwesome5Brands(62494);
-  static const IconData schlix =
-      const FlutterIconData.fontAwesome5Brands(62442);
-  static const IconData scribd =
-      const FlutterIconData.fontAwesome5Brands(62090);
-  static const IconData searchengin =
-      const FlutterIconData.fontAwesome5Brands(62443);
-  static const IconData sellcast =
-      const FlutterIconData.fontAwesome5Brands(62170);
-  static const IconData sellsy =
-      const FlutterIconData.fontAwesome5Brands(61971);
+      FlutterIconData.fontAwesome5Brands(62712);
+  static const IconData resolving = FlutterIconData.fontAwesome5Brands(62439);
+  static const IconData rev = FlutterIconData.fontAwesome5Brands(62898);
+  static const IconData rocketchat = FlutterIconData.fontAwesome5Brands(62440);
+  static const IconData rockrms = FlutterIconData.fontAwesome5Brands(62441);
+  static const IconData safari = FlutterIconData.fontAwesome5Brands(62055);
+  static const IconData sass = FlutterIconData.fontAwesome5Brands(62494);
+  static const IconData schlix = FlutterIconData.fontAwesome5Brands(62442);
+  static const IconData scribd = FlutterIconData.fontAwesome5Brands(62090);
+  static const IconData searchengin = FlutterIconData.fontAwesome5Brands(62443);
+  static const IconData sellcast = FlutterIconData.fontAwesome5Brands(62170);
+  static const IconData sellsy = FlutterIconData.fontAwesome5Brands(61971);
   static const IconData servicestack =
-      const FlutterIconData.fontAwesome5Brands(62444);
+      FlutterIconData.fontAwesome5Brands(62444);
   static const IconData shirtsinbulk =
-      const FlutterIconData.fontAwesome5Brands(61972);
-  static const IconData shopware =
-      const FlutterIconData.fontAwesome5Brands(62901);
-  static const IconData simplybuilt =
-      const FlutterIconData.fontAwesome5Brands(61973);
-  static const IconData sistrix =
-      const FlutterIconData.fontAwesome5Brands(62446);
-  static const IconData sith = const FlutterIconData.fontAwesome5Brands(62738);
-  static const IconData sketch =
-      const FlutterIconData.fontAwesome5Brands(63430);
-  static const IconData skyatlas =
-      const FlutterIconData.fontAwesome5Brands(61974);
-  static const IconData skype = const FlutterIconData.fontAwesome5Brands(61822);
-  static const IconData slack_hash =
-      const FlutterIconData.fontAwesome5Brands(62447);
-  static const IconData slack = const FlutterIconData.fontAwesome5Brands(61848);
-  static const IconData slideshare =
-      const FlutterIconData.fontAwesome5Brands(61927);
+      FlutterIconData.fontAwesome5Brands(61972);
+  static const IconData shopware = FlutterIconData.fontAwesome5Brands(62901);
+  static const IconData simplybuilt = FlutterIconData.fontAwesome5Brands(61973);
+  static const IconData sistrix = FlutterIconData.fontAwesome5Brands(62446);
+  static const IconData sith = FlutterIconData.fontAwesome5Brands(62738);
+  static const IconData sketch = FlutterIconData.fontAwesome5Brands(63430);
+  static const IconData skyatlas = FlutterIconData.fontAwesome5Brands(61974);
+  static const IconData skype = FlutterIconData.fontAwesome5Brands(61822);
+  static const IconData slack_hash = FlutterIconData.fontAwesome5Brands(62447);
+  static const IconData slack = FlutterIconData.fontAwesome5Brands(61848);
+  static const IconData slideshare = FlutterIconData.fontAwesome5Brands(61927);
   static const IconData snapchat_ghost =
-      const FlutterIconData.fontAwesome5Brands(62124);
+      FlutterIconData.fontAwesome5Brands(62124);
   static const IconData snapchat_square =
-      const FlutterIconData.fontAwesome5Brands(62125);
-  static const IconData snapchat =
-      const FlutterIconData.fontAwesome5Brands(62123);
-  static const IconData soundcloud =
-      const FlutterIconData.fontAwesome5Brands(61886);
-  static const IconData sourcetree =
-      const FlutterIconData.fontAwesome5Brands(63443);
-  static const IconData speakap =
-      const FlutterIconData.fontAwesome5Brands(62451);
-  static const IconData spotify =
-      const FlutterIconData.fontAwesome5Brands(61884);
-  static const IconData squarespace =
-      const FlutterIconData.fontAwesome5Brands(62910);
+      FlutterIconData.fontAwesome5Brands(62125);
+  static const IconData snapchat = FlutterIconData.fontAwesome5Brands(62123);
+  static const IconData soundcloud = FlutterIconData.fontAwesome5Brands(61886);
+  static const IconData sourcetree = FlutterIconData.fontAwesome5Brands(63443);
+  static const IconData speakap = FlutterIconData.fontAwesome5Brands(62451);
+  static const IconData spotify = FlutterIconData.fontAwesome5Brands(61884);
+  static const IconData squarespace = FlutterIconData.fontAwesome5Brands(62910);
   static const IconData stack_exchange =
-      const FlutterIconData.fontAwesome5Brands(61837);
+      FlutterIconData.fontAwesome5Brands(61837);
   static const IconData stack_overflow =
-      const FlutterIconData.fontAwesome5Brands(61804);
-  static const IconData staylinked =
-      const FlutterIconData.fontAwesome5Brands(62453);
+      FlutterIconData.fontAwesome5Brands(61804);
+  static const IconData staylinked = FlutterIconData.fontAwesome5Brands(62453);
   static const IconData steam_square =
-      const FlutterIconData.fontAwesome5Brands(61879);
+      FlutterIconData.fontAwesome5Brands(61879);
   static const IconData steam_symbol =
-      const FlutterIconData.fontAwesome5Brands(62454);
-  static const IconData steam = const FlutterIconData.fontAwesome5Brands(61878);
+      FlutterIconData.fontAwesome5Brands(62454);
+  static const IconData steam = FlutterIconData.fontAwesome5Brands(61878);
   static const IconData sticker_mule =
-      const FlutterIconData.fontAwesome5Brands(62455);
-  static const IconData strava =
-      const FlutterIconData.fontAwesome5Brands(62504);
-  static const IconData stripe_s =
-      const FlutterIconData.fontAwesome5Brands(62506);
-  static const IconData stripe =
-      const FlutterIconData.fontAwesome5Brands(62505);
+      FlutterIconData.fontAwesome5Brands(62455);
+  static const IconData strava = FlutterIconData.fontAwesome5Brands(62504);
+  static const IconData stripe_s = FlutterIconData.fontAwesome5Brands(62506);
+  static const IconData stripe = FlutterIconData.fontAwesome5Brands(62505);
   static const IconData studiovinari =
-      const FlutterIconData.fontAwesome5Brands(62456);
+      FlutterIconData.fontAwesome5Brands(62456);
   static const IconData stumbleupon_circle =
-      const FlutterIconData.fontAwesome5Brands(61859);
-  static const IconData stumbleupon =
-      const FlutterIconData.fontAwesome5Brands(61860);
-  static const IconData superpowers =
-      const FlutterIconData.fontAwesome5Brands(62173);
-  static const IconData supple =
-      const FlutterIconData.fontAwesome5Brands(62457);
-  static const IconData suse = const FlutterIconData.fontAwesome5Brands(63446);
-  static const IconData teamspeak =
-      const FlutterIconData.fontAwesome5Brands(62713);
+      FlutterIconData.fontAwesome5Brands(61859);
+  static const IconData stumbleupon = FlutterIconData.fontAwesome5Brands(61860);
+  static const IconData superpowers = FlutterIconData.fontAwesome5Brands(62173);
+  static const IconData supple = FlutterIconData.fontAwesome5Brands(62457);
+  static const IconData suse = FlutterIconData.fontAwesome5Brands(63446);
+  static const IconData teamspeak = FlutterIconData.fontAwesome5Brands(62713);
   static const IconData telegram_plane =
-      const FlutterIconData.fontAwesome5Brands(62462);
-  static const IconData telegram =
-      const FlutterIconData.fontAwesome5Brands(62150);
+      FlutterIconData.fontAwesome5Brands(62462);
+  static const IconData telegram = FlutterIconData.fontAwesome5Brands(62150);
   static const IconData tencent_weibo =
-      const FlutterIconData.fontAwesome5Brands(61909);
+      FlutterIconData.fontAwesome5Brands(61909);
   static const IconData the_red_yeti =
-      const FlutterIconData.fontAwesome5Brands(63133);
-  static const IconData themeco =
-      const FlutterIconData.fontAwesome5Brands(62918);
-  static const IconData themeisle =
-      const FlutterIconData.fontAwesome5Brands(62130);
-  static const IconData think_peaks =
-      const FlutterIconData.fontAwesome5Brands(63281);
+      FlutterIconData.fontAwesome5Brands(63133);
+  static const IconData themeco = FlutterIconData.fontAwesome5Brands(62918);
+  static const IconData themeisle = FlutterIconData.fontAwesome5Brands(62130);
+  static const IconData think_peaks = FlutterIconData.fontAwesome5Brands(63281);
   static const IconData trade_federation =
-      const FlutterIconData.fontAwesome5Brands(62739);
-  static const IconData trello =
-      const FlutterIconData.fontAwesome5Brands(61825);
-  static const IconData tripadvisor =
-      const FlutterIconData.fontAwesome5Brands(62050);
+      FlutterIconData.fontAwesome5Brands(62739);
+  static const IconData trello = FlutterIconData.fontAwesome5Brands(61825);
+  static const IconData tripadvisor = FlutterIconData.fontAwesome5Brands(62050);
   static const IconData tumblr_square =
-      const FlutterIconData.fontAwesome5Brands(61812);
-  static const IconData tumblr =
-      const FlutterIconData.fontAwesome5Brands(61811);
-  static const IconData twitch =
-      const FlutterIconData.fontAwesome5Brands(61928);
+      FlutterIconData.fontAwesome5Brands(61812);
+  static const IconData tumblr = FlutterIconData.fontAwesome5Brands(61811);
+  static const IconData twitch = FlutterIconData.fontAwesome5Brands(61928);
   static const IconData twitter_square =
-      const FlutterIconData.fontAwesome5Brands(61569);
-  static const IconData twitter =
-      const FlutterIconData.fontAwesome5Brands(61593);
-  static const IconData typo3 = const FlutterIconData.fontAwesome5Brands(62507);
-  static const IconData uber = const FlutterIconData.fontAwesome5Brands(62466);
-  static const IconData ubuntu =
-      const FlutterIconData.fontAwesome5Brands(63455);
-  static const IconData uikit = const FlutterIconData.fontAwesome5Brands(62467);
-  static const IconData uniregistry =
-      const FlutterIconData.fontAwesome5Brands(62468);
-  static const IconData untappd =
-      const FlutterIconData.fontAwesome5Brands(62469);
-  static const IconData ups = const FlutterIconData.fontAwesome5Brands(63456);
-  static const IconData usb = const FlutterIconData.fontAwesome5Brands(62087);
-  static const IconData usps = const FlutterIconData.fontAwesome5Brands(63457);
-  static const IconData ussunnah =
-      const FlutterIconData.fontAwesome5Brands(62471);
-  static const IconData vaadin =
-      const FlutterIconData.fontAwesome5Brands(62472);
-  static const IconData viacoin =
-      const FlutterIconData.fontAwesome5Brands(62007);
+      FlutterIconData.fontAwesome5Brands(61569);
+  static const IconData twitter = FlutterIconData.fontAwesome5Brands(61593);
+  static const IconData typo3 = FlutterIconData.fontAwesome5Brands(62507);
+  static const IconData uber = FlutterIconData.fontAwesome5Brands(62466);
+  static const IconData ubuntu = FlutterIconData.fontAwesome5Brands(63455);
+  static const IconData uikit = FlutterIconData.fontAwesome5Brands(62467);
+  static const IconData uniregistry = FlutterIconData.fontAwesome5Brands(62468);
+  static const IconData untappd = FlutterIconData.fontAwesome5Brands(62469);
+  static const IconData ups = FlutterIconData.fontAwesome5Brands(63456);
+  static const IconData usb = FlutterIconData.fontAwesome5Brands(62087);
+  static const IconData usps = FlutterIconData.fontAwesome5Brands(63457);
+  static const IconData ussunnah = FlutterIconData.fontAwesome5Brands(62471);
+  static const IconData vaadin = FlutterIconData.fontAwesome5Brands(62472);
+  static const IconData viacoin = FlutterIconData.fontAwesome5Brands(62007);
   static const IconData viadeo_square =
-      const FlutterIconData.fontAwesome5Brands(62122);
-  static const IconData viadeo =
-      const FlutterIconData.fontAwesome5Brands(62121);
-  static const IconData viber = const FlutterIconData.fontAwesome5Brands(62473);
+      FlutterIconData.fontAwesome5Brands(62122);
+  static const IconData viadeo = FlutterIconData.fontAwesome5Brands(62121);
+  static const IconData viber = FlutterIconData.fontAwesome5Brands(62473);
   static const IconData vimeo_square =
-      const FlutterIconData.fontAwesome5Brands(61844);
-  static const IconData vimeo_v =
-      const FlutterIconData.fontAwesome5Brands(62077);
-  static const IconData vimeo = const FlutterIconData.fontAwesome5Brands(62474);
-  static const IconData vine = const FlutterIconData.fontAwesome5Brands(61898);
-  static const IconData vk = const FlutterIconData.fontAwesome5Brands(61833);
-  static const IconData vnv = const FlutterIconData.fontAwesome5Brands(62475);
-  static const IconData vuejs = const FlutterIconData.fontAwesome5Brands(62495);
-  static const IconData weebly =
-      const FlutterIconData.fontAwesome5Brands(62924);
-  static const IconData weibo = const FlutterIconData.fontAwesome5Brands(61834);
-  static const IconData weixin =
-      const FlutterIconData.fontAwesome5Brands(61911);
+      FlutterIconData.fontAwesome5Brands(61844);
+  static const IconData vimeo_v = FlutterIconData.fontAwesome5Brands(62077);
+  static const IconData vimeo = FlutterIconData.fontAwesome5Brands(62474);
+  static const IconData vine = FlutterIconData.fontAwesome5Brands(61898);
+  static const IconData vk = FlutterIconData.fontAwesome5Brands(61833);
+  static const IconData vnv = FlutterIconData.fontAwesome5Brands(62475);
+  static const IconData vuejs = FlutterIconData.fontAwesome5Brands(62495);
+  static const IconData weebly = FlutterIconData.fontAwesome5Brands(62924);
+  static const IconData weibo = FlutterIconData.fontAwesome5Brands(61834);
+  static const IconData weixin = FlutterIconData.fontAwesome5Brands(61911);
   static const IconData whatsapp_square =
-      const FlutterIconData.fontAwesome5Brands(62476);
-  static const IconData whatsapp =
-      const FlutterIconData.fontAwesome5Brands(62002);
-  static const IconData whmcs = const FlutterIconData.fontAwesome5Brands(62477);
-  static const IconData wikipedia_w =
-      const FlutterIconData.fontAwesome5Brands(62054);
-  static const IconData windows =
-      const FlutterIconData.fontAwesome5Brands(61818);
-  static const IconData wix = const FlutterIconData.fontAwesome5Brands(62927);
+      FlutterIconData.fontAwesome5Brands(62476);
+  static const IconData whatsapp = FlutterIconData.fontAwesome5Brands(62002);
+  static const IconData whmcs = FlutterIconData.fontAwesome5Brands(62477);
+  static const IconData wikipedia_w = FlutterIconData.fontAwesome5Brands(62054);
+  static const IconData windows = FlutterIconData.fontAwesome5Brands(61818);
+  static const IconData wix = FlutterIconData.fontAwesome5Brands(62927);
   static const IconData wizards_of_the_coast =
-      const FlutterIconData.fontAwesome5Brands(63280);
+      FlutterIconData.fontAwesome5Brands(63280);
   static const IconData wolf_pack_battalion =
-      const FlutterIconData.fontAwesome5Brands(62740);
+      FlutterIconData.fontAwesome5Brands(62740);
   static const IconData wordpress_simple =
-      const FlutterIconData.fontAwesome5Brands(62481);
-  static const IconData wordpress =
-      const FlutterIconData.fontAwesome5Brands(61850);
-  static const IconData wpbeginner =
-      const FlutterIconData.fontAwesome5Brands(62103);
-  static const IconData wpexplorer =
-      const FlutterIconData.fontAwesome5Brands(62174);
-  static const IconData wpforms =
-      const FlutterIconData.fontAwesome5Brands(62104);
-  static const IconData wpressr =
-      const FlutterIconData.fontAwesome5Brands(62436);
-  static const IconData xbox = const FlutterIconData.fontAwesome5Brands(62482);
-  static const IconData xing_square =
-      const FlutterIconData.fontAwesome5Brands(61801);
-  static const IconData xing = const FlutterIconData.fontAwesome5Brands(61800);
+      FlutterIconData.fontAwesome5Brands(62481);
+  static const IconData wordpress = FlutterIconData.fontAwesome5Brands(61850);
+  static const IconData wpbeginner = FlutterIconData.fontAwesome5Brands(62103);
+  static const IconData wpexplorer = FlutterIconData.fontAwesome5Brands(62174);
+  static const IconData wpforms = FlutterIconData.fontAwesome5Brands(62104);
+  static const IconData wpressr = FlutterIconData.fontAwesome5Brands(62436);
+  static const IconData xbox = FlutterIconData.fontAwesome5Brands(62482);
+  static const IconData xing_square = FlutterIconData.fontAwesome5Brands(61801);
+  static const IconData xing = FlutterIconData.fontAwesome5Brands(61800);
   static const IconData y_combinator =
-      const FlutterIconData.fontAwesome5Brands(62011);
-  static const IconData yahoo = const FlutterIconData.fontAwesome5Brands(61854);
+      FlutterIconData.fontAwesome5Brands(62011);
+  static const IconData yahoo = FlutterIconData.fontAwesome5Brands(61854);
   static const IconData yandex_international =
-      const FlutterIconData.fontAwesome5Brands(62484);
-  static const IconData yandex =
-      const FlutterIconData.fontAwesome5Brands(62483);
-  static const IconData yarn = const FlutterIconData.fontAwesome5Brands(63459);
-  static const IconData yelp = const FlutterIconData.fontAwesome5Brands(61929);
-  static const IconData yoast = const FlutterIconData.fontAwesome5Brands(62129);
+      FlutterIconData.fontAwesome5Brands(62484);
+  static const IconData yandex = FlutterIconData.fontAwesome5Brands(62483);
+  static const IconData yarn = FlutterIconData.fontAwesome5Brands(63459);
+  static const IconData yelp = FlutterIconData.fontAwesome5Brands(61929);
+  static const IconData yoast = FlutterIconData.fontAwesome5Brands(62129);
   static const IconData youtube_square =
-      const FlutterIconData.fontAwesome5Brands(62513);
-  static const IconData youtube =
-      const FlutterIconData.fontAwesome5Brands(61799);
-  static const IconData zhihu = const FlutterIconData.fontAwesome5Brands(63039);
+      FlutterIconData.fontAwesome5Brands(62513);
+  static const IconData youtube = FlutterIconData.fontAwesome5Brands(61799);
+  static const IconData zhihu = FlutterIconData.fontAwesome5Brands(63039);
 }
