@@ -6,340 +6,292 @@ import 'flutter_icon_data.dart';
 /// http://zurb.com/playground/foundation-icon-fonts-3
 class Foundation {
   Foundation._();
-  static const IconData address_book = const FlutterIconData.foundation(61696);
-  static const IconData alert = const FlutterIconData.foundation(61697);
-  static const IconData align_center = const FlutterIconData.foundation(61698);
-  static const IconData align_justify = const FlutterIconData.foundation(61699);
-  static const IconData align_left = const FlutterIconData.foundation(61700);
-  static const IconData align_right = const FlutterIconData.foundation(61701);
-  static const IconData anchor = const FlutterIconData.foundation(61702);
-  static const IconData annotate = const FlutterIconData.foundation(61703);
-  static const IconData archive = const FlutterIconData.foundation(61704);
-  static const IconData arrow_down = const FlutterIconData.foundation(61705);
-  static const IconData arrow_left = const FlutterIconData.foundation(61706);
-  static const IconData arrow_right = const FlutterIconData.foundation(61707);
-  static const IconData arrow_up = const FlutterIconData.foundation(61708);
-  static const IconData arrows_compress =
-      const FlutterIconData.foundation(61709);
-  static const IconData arrows_expand = const FlutterIconData.foundation(61710);
-  static const IconData arrows_in = const FlutterIconData.foundation(61711);
-  static const IconData arrows_out = const FlutterIconData.foundation(61712);
-  static const IconData asl = const FlutterIconData.foundation(61713);
-  static const IconData asterisk = const FlutterIconData.foundation(61714);
-  static const IconData at_sign = const FlutterIconData.foundation(61715);
-  static const IconData background_color =
-      const FlutterIconData.foundation(61716);
-  static const IconData battery_empty = const FlutterIconData.foundation(61717);
-  static const IconData battery_full = const FlutterIconData.foundation(61718);
-  static const IconData battery_half = const FlutterIconData.foundation(61719);
-  static const IconData bitcoin_circle =
-      const FlutterIconData.foundation(61720);
-  static const IconData bitcoin = const FlutterIconData.foundation(61721);
-  static const IconData blind = const FlutterIconData.foundation(61722);
-  static const IconData bluetooth = const FlutterIconData.foundation(61723);
-  static const IconData bold = const FlutterIconData.foundation(61724);
-  static const IconData book_bookmark = const FlutterIconData.foundation(61725);
-  static const IconData book = const FlutterIconData.foundation(61726);
-  static const IconData bookmark = const FlutterIconData.foundation(61727);
-  static const IconData braille = const FlutterIconData.foundation(61728);
-  static const IconData burst_new = const FlutterIconData.foundation(61729);
-  static const IconData burst_sale = const FlutterIconData.foundation(61730);
-  static const IconData burst = const FlutterIconData.foundation(61731);
-  static const IconData calendar = const FlutterIconData.foundation(61732);
-  static const IconData camera = const FlutterIconData.foundation(61733);
-  static const IconData check = const FlutterIconData.foundation(61734);
-  static const IconData checkbox = const FlutterIconData.foundation(61735);
-  static const IconData clipboard_notes =
-      const FlutterIconData.foundation(61736);
-  static const IconData clipboard_pencil =
-      const FlutterIconData.foundation(61737);
-  static const IconData clipboard = const FlutterIconData.foundation(61738);
-  static const IconData clock = const FlutterIconData.foundation(61739);
-  static const IconData closed_caption =
-      const FlutterIconData.foundation(61740);
-  static const IconData cloud = const FlutterIconData.foundation(61741);
-  static const IconData comment_minus = const FlutterIconData.foundation(61742);
-  static const IconData comment_quotes =
-      const FlutterIconData.foundation(61743);
-  static const IconData comment_video = const FlutterIconData.foundation(61744);
-  static const IconData comment = const FlutterIconData.foundation(61745);
-  static const IconData comments = const FlutterIconData.foundation(61746);
-  static const IconData compass = const FlutterIconData.foundation(61747);
-  static const IconData contrast = const FlutterIconData.foundation(61748);
-  static const IconData credit_card = const FlutterIconData.foundation(61749);
-  static const IconData crop = const FlutterIconData.foundation(61750);
-  static const IconData crown = const FlutterIconData.foundation(61751);
-  static const IconData css3 = const FlutterIconData.foundation(61752);
-  static const IconData database = const FlutterIconData.foundation(61753);
-  static const IconData die_five = const FlutterIconData.foundation(61754);
-  static const IconData die_four = const FlutterIconData.foundation(61755);
-  static const IconData die_one = const FlutterIconData.foundation(61756);
-  static const IconData die_six = const FlutterIconData.foundation(61757);
-  static const IconData die_three = const FlutterIconData.foundation(61758);
-  static const IconData die_two = const FlutterIconData.foundation(61759);
-  static const IconData dislike = const FlutterIconData.foundation(61760);
-  static const IconData dollar_bill = const FlutterIconData.foundation(61761);
-  static const IconData dollar = const FlutterIconData.foundation(61762);
-  static const IconData download = const FlutterIconData.foundation(61763);
-  static const IconData eject = const FlutterIconData.foundation(61764);
-  static const IconData elevator = const FlutterIconData.foundation(61765);
-  static const IconData euro = const FlutterIconData.foundation(61766);
-  static const IconData eye = const FlutterIconData.foundation(61767);
-  static const IconData fast_forward = const FlutterIconData.foundation(61768);
-  static const IconData female_symbol = const FlutterIconData.foundation(61769);
-  static const IconData female = const FlutterIconData.foundation(61770);
-  static const IconData filter = const FlutterIconData.foundation(61771);
-  static const IconData first_aid = const FlutterIconData.foundation(61772);
-  static const IconData flag = const FlutterIconData.foundation(61773);
-  static const IconData folder_add = const FlutterIconData.foundation(61774);
-  static const IconData folder_lock = const FlutterIconData.foundation(61775);
-  static const IconData folder = const FlutterIconData.foundation(61776);
-  static const IconData foot = const FlutterIconData.foundation(61777);
-  static const IconData foundation = const FlutterIconData.foundation(61778);
-  static const IconData graph_bar = const FlutterIconData.foundation(61779);
-  static const IconData graph_horizontal =
-      const FlutterIconData.foundation(61780);
-  static const IconData graph_pie = const FlutterIconData.foundation(61781);
-  static const IconData graph_trend = const FlutterIconData.foundation(61782);
-  static const IconData guide_dog = const FlutterIconData.foundation(61783);
-  static const IconData hearing_aid = const FlutterIconData.foundation(61784);
-  static const IconData heart = const FlutterIconData.foundation(61785);
-  static const IconData home = const FlutterIconData.foundation(61786);
-  static const IconData html5 = const FlutterIconData.foundation(61787);
-  static const IconData indent_less = const FlutterIconData.foundation(61788);
-  static const IconData indent_more = const FlutterIconData.foundation(61789);
-  static const IconData info = const FlutterIconData.foundation(61790);
-  static const IconData italic = const FlutterIconData.foundation(61791);
-  static const IconData key = const FlutterIconData.foundation(61792);
-  static const IconData laptop = const FlutterIconData.foundation(61793);
-  static const IconData layout = const FlutterIconData.foundation(61794);
-  static const IconData lightbulb = const FlutterIconData.foundation(61795);
-  static const IconData like = const FlutterIconData.foundation(61796);
-  static const IconData link = const FlutterIconData.foundation(61797);
-  static const IconData list_bullet = const FlutterIconData.foundation(61798);
-  static const IconData list_number = const FlutterIconData.foundation(61799);
-  static const IconData list_thumbnails =
-      const FlutterIconData.foundation(61800);
-  static const IconData list = const FlutterIconData.foundation(61801);
-  static const IconData lock = const FlutterIconData.foundation(61802);
-  static const IconData loop = const FlutterIconData.foundation(61803);
-  static const IconData magnifying_glass =
-      const FlutterIconData.foundation(61804);
-  static const IconData mail = const FlutterIconData.foundation(61805);
-  static const IconData male_female = const FlutterIconData.foundation(61806);
-  static const IconData male_symbol = const FlutterIconData.foundation(61807);
-  static const IconData male = const FlutterIconData.foundation(61808);
-  static const IconData map = const FlutterIconData.foundation(61809);
-  static const IconData marker = const FlutterIconData.foundation(61810);
-  static const IconData megaphone = const FlutterIconData.foundation(61811);
-  static const IconData microphone = const FlutterIconData.foundation(61812);
-  static const IconData minus_circle = const FlutterIconData.foundation(61813);
-  static const IconData minus = const FlutterIconData.foundation(61814);
-  static const IconData mobile_signal = const FlutterIconData.foundation(61815);
-  static const IconData mobile = const FlutterIconData.foundation(61816);
-  static const IconData monitor = const FlutterIconData.foundation(61817);
-  static const IconData mountains = const FlutterIconData.foundation(61818);
-  static const IconData music = const FlutterIconData.foundation(61819);
-  static const IconData next = const FlutterIconData.foundation(61820);
-  static const IconData no_dogs = const FlutterIconData.foundation(61821);
-  static const IconData no_smoking = const FlutterIconData.foundation(61822);
-  static const IconData page_add = const FlutterIconData.foundation(61823);
-  static const IconData page_copy = const FlutterIconData.foundation(61824);
-  static const IconData page_csv = const FlutterIconData.foundation(61825);
-  static const IconData page_delete = const FlutterIconData.foundation(61826);
-  static const IconData page_doc = const FlutterIconData.foundation(61827);
-  static const IconData page_edit = const FlutterIconData.foundation(61828);
-  static const IconData page_export_csv =
-      const FlutterIconData.foundation(61829);
-  static const IconData page_export_doc =
-      const FlutterIconData.foundation(61830);
-  static const IconData page_export_pdf =
-      const FlutterIconData.foundation(61831);
-  static const IconData page_export = const FlutterIconData.foundation(61832);
-  static const IconData page_filled = const FlutterIconData.foundation(61833);
-  static const IconData page_multiple = const FlutterIconData.foundation(61834);
-  static const IconData page_pdf = const FlutterIconData.foundation(61835);
-  static const IconData page_remove = const FlutterIconData.foundation(61836);
-  static const IconData page_search = const FlutterIconData.foundation(61837);
-  static const IconData page = const FlutterIconData.foundation(61838);
-  static const IconData paint_bucket = const FlutterIconData.foundation(61839);
-  static const IconData paperclip = const FlutterIconData.foundation(61840);
-  static const IconData pause = const FlutterIconData.foundation(61841);
-  static const IconData paw = const FlutterIconData.foundation(61842);
-  static const IconData paypal = const FlutterIconData.foundation(61843);
-  static const IconData pencil = const FlutterIconData.foundation(61844);
-  static const IconData photo = const FlutterIconData.foundation(61845);
-  static const IconData play_circle = const FlutterIconData.foundation(61846);
-  static const IconData play_video = const FlutterIconData.foundation(61847);
-  static const IconData play = const FlutterIconData.foundation(61848);
-  static const IconData plus = const FlutterIconData.foundation(61849);
-  static const IconData pound = const FlutterIconData.foundation(61850);
-  static const IconData power = const FlutterIconData.foundation(61851);
-  static const IconData previous = const FlutterIconData.foundation(61852);
-  static const IconData price_tag = const FlutterIconData.foundation(61853);
-  static const IconData pricetag_multiple =
-      const FlutterIconData.foundation(61854);
-  static const IconData print = const FlutterIconData.foundation(61855);
-  static const IconData prohibited = const FlutterIconData.foundation(61856);
-  static const IconData projection_screen =
-      const FlutterIconData.foundation(61857);
-  static const IconData puzzle = const FlutterIconData.foundation(61858);
-  static const IconData quote = const FlutterIconData.foundation(61859);
-  static const IconData record = const FlutterIconData.foundation(61860);
-  static const IconData refresh = const FlutterIconData.foundation(61861);
+
+  static const IconData address_book = FlutterIconData.foundation(61696);
+  static const IconData alert = FlutterIconData.foundation(61697);
+  static const IconData align_center = FlutterIconData.foundation(61698);
+  static const IconData align_justify = FlutterIconData.foundation(61699);
+  static const IconData align_left = FlutterIconData.foundation(61700);
+  static const IconData align_right = FlutterIconData.foundation(61701);
+  static const IconData anchor = FlutterIconData.foundation(61702);
+  static const IconData annotate = FlutterIconData.foundation(61703);
+  static const IconData archive = FlutterIconData.foundation(61704);
+  static const IconData arrow_down = FlutterIconData.foundation(61705);
+  static const IconData arrow_left = FlutterIconData.foundation(61706);
+  static const IconData arrow_right = FlutterIconData.foundation(61707);
+  static const IconData arrow_up = FlutterIconData.foundation(61708);
+  static const IconData arrows_compress = FlutterIconData.foundation(61709);
+  static const IconData arrows_expand = FlutterIconData.foundation(61710);
+  static const IconData arrows_in = FlutterIconData.foundation(61711);
+  static const IconData arrows_out = FlutterIconData.foundation(61712);
+  static const IconData asl = FlutterIconData.foundation(61713);
+  static const IconData asterisk = FlutterIconData.foundation(61714);
+  static const IconData at_sign = FlutterIconData.foundation(61715);
+  static const IconData background_color = FlutterIconData.foundation(61716);
+  static const IconData battery_empty = FlutterIconData.foundation(61717);
+  static const IconData battery_full = FlutterIconData.foundation(61718);
+  static const IconData battery_half = FlutterIconData.foundation(61719);
+  static const IconData bitcoin_circle = FlutterIconData.foundation(61720);
+  static const IconData bitcoin = FlutterIconData.foundation(61721);
+  static const IconData blind = FlutterIconData.foundation(61722);
+  static const IconData bluetooth = FlutterIconData.foundation(61723);
+  static const IconData bold = FlutterIconData.foundation(61724);
+  static const IconData book_bookmark = FlutterIconData.foundation(61725);
+  static const IconData book = FlutterIconData.foundation(61726);
+  static const IconData bookmark = FlutterIconData.foundation(61727);
+  static const IconData braille = FlutterIconData.foundation(61728);
+  static const IconData burst_new = FlutterIconData.foundation(61729);
+  static const IconData burst_sale = FlutterIconData.foundation(61730);
+  static const IconData burst = FlutterIconData.foundation(61731);
+  static const IconData calendar = FlutterIconData.foundation(61732);
+  static const IconData camera = FlutterIconData.foundation(61733);
+  static const IconData check = FlutterIconData.foundation(61734);
+  static const IconData checkbox = FlutterIconData.foundation(61735);
+  static const IconData clipboard_notes = FlutterIconData.foundation(61736);
+  static const IconData clipboard_pencil = FlutterIconData.foundation(61737);
+  static const IconData clipboard = FlutterIconData.foundation(61738);
+  static const IconData clock = FlutterIconData.foundation(61739);
+  static const IconData closed_caption = FlutterIconData.foundation(61740);
+  static const IconData cloud = FlutterIconData.foundation(61741);
+  static const IconData comment_minus = FlutterIconData.foundation(61742);
+  static const IconData comment_quotes = FlutterIconData.foundation(61743);
+  static const IconData comment_video = FlutterIconData.foundation(61744);
+  static const IconData comment = FlutterIconData.foundation(61745);
+  static const IconData comments = FlutterIconData.foundation(61746);
+  static const IconData compass = FlutterIconData.foundation(61747);
+  static const IconData contrast = FlutterIconData.foundation(61748);
+  static const IconData credit_card = FlutterIconData.foundation(61749);
+  static const IconData crop = FlutterIconData.foundation(61750);
+  static const IconData crown = FlutterIconData.foundation(61751);
+  static const IconData css3 = FlutterIconData.foundation(61752);
+  static const IconData database = FlutterIconData.foundation(61753);
+  static const IconData die_five = FlutterIconData.foundation(61754);
+  static const IconData die_four = FlutterIconData.foundation(61755);
+  static const IconData die_one = FlutterIconData.foundation(61756);
+  static const IconData die_six = FlutterIconData.foundation(61757);
+  static const IconData die_three = FlutterIconData.foundation(61758);
+  static const IconData die_two = FlutterIconData.foundation(61759);
+  static const IconData dislike = FlutterIconData.foundation(61760);
+  static const IconData dollar_bill = FlutterIconData.foundation(61761);
+  static const IconData dollar = FlutterIconData.foundation(61762);
+  static const IconData download = FlutterIconData.foundation(61763);
+  static const IconData eject = FlutterIconData.foundation(61764);
+  static const IconData elevator = FlutterIconData.foundation(61765);
+  static const IconData euro = FlutterIconData.foundation(61766);
+  static const IconData eye = FlutterIconData.foundation(61767);
+  static const IconData fast_forward = FlutterIconData.foundation(61768);
+  static const IconData female_symbol = FlutterIconData.foundation(61769);
+  static const IconData female = FlutterIconData.foundation(61770);
+  static const IconData filter = FlutterIconData.foundation(61771);
+  static const IconData first_aid = FlutterIconData.foundation(61772);
+  static const IconData flag = FlutterIconData.foundation(61773);
+  static const IconData folder_add = FlutterIconData.foundation(61774);
+  static const IconData folder_lock = FlutterIconData.foundation(61775);
+  static const IconData folder = FlutterIconData.foundation(61776);
+  static const IconData foot = FlutterIconData.foundation(61777);
+  static const IconData foundation = FlutterIconData.foundation(61778);
+  static const IconData graph_bar = FlutterIconData.foundation(61779);
+  static const IconData graph_horizontal = FlutterIconData.foundation(61780);
+  static const IconData graph_pie = FlutterIconData.foundation(61781);
+  static const IconData graph_trend = FlutterIconData.foundation(61782);
+  static const IconData guide_dog = FlutterIconData.foundation(61783);
+  static const IconData hearing_aid = FlutterIconData.foundation(61784);
+  static const IconData heart = FlutterIconData.foundation(61785);
+  static const IconData home = FlutterIconData.foundation(61786);
+  static const IconData html5 = FlutterIconData.foundation(61787);
+  static const IconData indent_less = FlutterIconData.foundation(61788);
+  static const IconData indent_more = FlutterIconData.foundation(61789);
+  static const IconData info = FlutterIconData.foundation(61790);
+  static const IconData italic = FlutterIconData.foundation(61791);
+  static const IconData key = FlutterIconData.foundation(61792);
+  static const IconData laptop = FlutterIconData.foundation(61793);
+  static const IconData layout = FlutterIconData.foundation(61794);
+  static const IconData lightbulb = FlutterIconData.foundation(61795);
+  static const IconData like = FlutterIconData.foundation(61796);
+  static const IconData link = FlutterIconData.foundation(61797);
+  static const IconData list_bullet = FlutterIconData.foundation(61798);
+  static const IconData list_number = FlutterIconData.foundation(61799);
+  static const IconData list_thumbnails = FlutterIconData.foundation(61800);
+  static const IconData list = FlutterIconData.foundation(61801);
+  static const IconData lock = FlutterIconData.foundation(61802);
+  static const IconData loop = FlutterIconData.foundation(61803);
+  static const IconData magnifying_glass = FlutterIconData.foundation(61804);
+  static const IconData mail = FlutterIconData.foundation(61805);
+  static const IconData male_female = FlutterIconData.foundation(61806);
+  static const IconData male_symbol = FlutterIconData.foundation(61807);
+  static const IconData male = FlutterIconData.foundation(61808);
+  static const IconData map = FlutterIconData.foundation(61809);
+  static const IconData marker = FlutterIconData.foundation(61810);
+  static const IconData megaphone = FlutterIconData.foundation(61811);
+  static const IconData microphone = FlutterIconData.foundation(61812);
+  static const IconData minus_circle = FlutterIconData.foundation(61813);
+  static const IconData minus = FlutterIconData.foundation(61814);
+  static const IconData mobile_signal = FlutterIconData.foundation(61815);
+  static const IconData mobile = FlutterIconData.foundation(61816);
+  static const IconData monitor = FlutterIconData.foundation(61817);
+  static const IconData mountains = FlutterIconData.foundation(61818);
+  static const IconData music = FlutterIconData.foundation(61819);
+  static const IconData next = FlutterIconData.foundation(61820);
+  static const IconData no_dogs = FlutterIconData.foundation(61821);
+  static const IconData no_smoking = FlutterIconData.foundation(61822);
+  static const IconData page_add = FlutterIconData.foundation(61823);
+  static const IconData page_copy = FlutterIconData.foundation(61824);
+  static const IconData page_csv = FlutterIconData.foundation(61825);
+  static const IconData page_delete = FlutterIconData.foundation(61826);
+  static const IconData page_doc = FlutterIconData.foundation(61827);
+  static const IconData page_edit = FlutterIconData.foundation(61828);
+  static const IconData page_export_csv = FlutterIconData.foundation(61829);
+  static const IconData page_export_doc = FlutterIconData.foundation(61830);
+  static const IconData page_export_pdf = FlutterIconData.foundation(61831);
+  static const IconData page_export = FlutterIconData.foundation(61832);
+  static const IconData page_filled = FlutterIconData.foundation(61833);
+  static const IconData page_multiple = FlutterIconData.foundation(61834);
+  static const IconData page_pdf = FlutterIconData.foundation(61835);
+  static const IconData page_remove = FlutterIconData.foundation(61836);
+  static const IconData page_search = FlutterIconData.foundation(61837);
+  static const IconData page = FlutterIconData.foundation(61838);
+  static const IconData paint_bucket = FlutterIconData.foundation(61839);
+  static const IconData paperclip = FlutterIconData.foundation(61840);
+  static const IconData pause = FlutterIconData.foundation(61841);
+  static const IconData paw = FlutterIconData.foundation(61842);
+  static const IconData paypal = FlutterIconData.foundation(61843);
+  static const IconData pencil = FlutterIconData.foundation(61844);
+  static const IconData photo = FlutterIconData.foundation(61845);
+  static const IconData play_circle = FlutterIconData.foundation(61846);
+  static const IconData play_video = FlutterIconData.foundation(61847);
+  static const IconData play = FlutterIconData.foundation(61848);
+  static const IconData plus = FlutterIconData.foundation(61849);
+  static const IconData pound = FlutterIconData.foundation(61850);
+  static const IconData power = FlutterIconData.foundation(61851);
+  static const IconData previous = FlutterIconData.foundation(61852);
+  static const IconData price_tag = FlutterIconData.foundation(61853);
+  static const IconData pricetag_multiple = FlutterIconData.foundation(61854);
+  static const IconData print = FlutterIconData.foundation(61855);
+  static const IconData prohibited = FlutterIconData.foundation(61856);
+  static const IconData projection_screen = FlutterIconData.foundation(61857);
+  static const IconData puzzle = FlutterIconData.foundation(61858);
+  static const IconData quote = FlutterIconData.foundation(61859);
+  static const IconData record = FlutterIconData.foundation(61860);
+  static const IconData refresh = FlutterIconData.foundation(61861);
   static const IconData results_demographics =
-      const FlutterIconData.foundation(61862);
-  static const IconData results = const FlutterIconData.foundation(61863);
-  static const IconData rewind_ten = const FlutterIconData.foundation(61864);
-  static const IconData rewind = const FlutterIconData.foundation(61865);
-  static const IconData rss = const FlutterIconData.foundation(61866);
-  static const IconData safety_cone = const FlutterIconData.foundation(61867);
-  static const IconData save = const FlutterIconData.foundation(61868);
-  static const IconData share = const FlutterIconData.foundation(61869);
-  static const IconData sheriff_badge = const FlutterIconData.foundation(61870);
-  static const IconData shield = const FlutterIconData.foundation(61871);
-  static const IconData shopping_bag = const FlutterIconData.foundation(61872);
-  static const IconData shopping_cart = const FlutterIconData.foundation(61873);
-  static const IconData shuffle = const FlutterIconData.foundation(61874);
-  static const IconData skull = const FlutterIconData.foundation(61875);
-  static const IconData social_500px = const FlutterIconData.foundation(61876);
-  static const IconData social_adobe = const FlutterIconData.foundation(61877);
-  static const IconData social_amazon = const FlutterIconData.foundation(61878);
-  static const IconData social_android =
-      const FlutterIconData.foundation(61879);
-  static const IconData social_apple = const FlutterIconData.foundation(61880);
-  static const IconData social_behance =
-      const FlutterIconData.foundation(61881);
-  static const IconData social_bing = const FlutterIconData.foundation(61882);
-  static const IconData social_blogger =
-      const FlutterIconData.foundation(61883);
-  static const IconData social_delicious =
-      const FlutterIconData.foundation(61884);
+      FlutterIconData.foundation(61862);
+  static const IconData results = FlutterIconData.foundation(61863);
+  static const IconData rewind_ten = FlutterIconData.foundation(61864);
+  static const IconData rewind = FlutterIconData.foundation(61865);
+  static const IconData rss = FlutterIconData.foundation(61866);
+  static const IconData safety_cone = FlutterIconData.foundation(61867);
+  static const IconData save = FlutterIconData.foundation(61868);
+  static const IconData share = FlutterIconData.foundation(61869);
+  static const IconData sheriff_badge = FlutterIconData.foundation(61870);
+  static const IconData shield = FlutterIconData.foundation(61871);
+  static const IconData shopping_bag = FlutterIconData.foundation(61872);
+  static const IconData shopping_cart = FlutterIconData.foundation(61873);
+  static const IconData shuffle = FlutterIconData.foundation(61874);
+  static const IconData skull = FlutterIconData.foundation(61875);
+  static const IconData social_500px = FlutterIconData.foundation(61876);
+  static const IconData social_adobe = FlutterIconData.foundation(61877);
+  static const IconData social_amazon = FlutterIconData.foundation(61878);
+  static const IconData social_android = FlutterIconData.foundation(61879);
+  static const IconData social_apple = FlutterIconData.foundation(61880);
+  static const IconData social_behance = FlutterIconData.foundation(61881);
+  static const IconData social_bing = FlutterIconData.foundation(61882);
+  static const IconData social_blogger = FlutterIconData.foundation(61883);
+  static const IconData social_delicious = FlutterIconData.foundation(61884);
   static const IconData social_designer_news =
-      const FlutterIconData.foundation(61885);
-  static const IconData social_deviant_art =
-      const FlutterIconData.foundation(61886);
-  static const IconData social_digg = const FlutterIconData.foundation(61887);
-  static const IconData social_dribbble =
-      const FlutterIconData.foundation(61888);
-  static const IconData social_drive = const FlutterIconData.foundation(61889);
-  static const IconData social_dropbox =
-      const FlutterIconData.foundation(61890);
-  static const IconData social_evernote =
-      const FlutterIconData.foundation(61891);
-  static const IconData social_facebook =
-      const FlutterIconData.foundation(61892);
-  static const IconData social_flickr = const FlutterIconData.foundation(61893);
-  static const IconData social_forrst = const FlutterIconData.foundation(61894);
-  static const IconData social_foursquare =
-      const FlutterIconData.foundation(61895);
-  static const IconData social_game_center =
-      const FlutterIconData.foundation(61896);
-  static const IconData social_github = const FlutterIconData.foundation(61897);
-  static const IconData social_google_plus =
-      const FlutterIconData.foundation(61898);
-  static const IconData social_hacker_news =
-      const FlutterIconData.foundation(61899);
-  static const IconData social_hi5 = const FlutterIconData.foundation(61900);
-  static const IconData social_instagram =
-      const FlutterIconData.foundation(61901);
-  static const IconData social_joomla = const FlutterIconData.foundation(61902);
-  static const IconData social_lastfm = const FlutterIconData.foundation(61903);
-  static const IconData social_linkedin =
-      const FlutterIconData.foundation(61904);
-  static const IconData social_medium = const FlutterIconData.foundation(61905);
-  static const IconData social_myspace =
-      const FlutterIconData.foundation(61906);
-  static const IconData social_orkut = const FlutterIconData.foundation(61907);
-  static const IconData social_path = const FlutterIconData.foundation(61908);
-  static const IconData social_picasa = const FlutterIconData.foundation(61909);
-  static const IconData social_pinterest =
-      const FlutterIconData.foundation(61910);
-  static const IconData social_rdio = const FlutterIconData.foundation(61911);
-  static const IconData social_reddit = const FlutterIconData.foundation(61912);
-  static const IconData social_skillshare =
-      const FlutterIconData.foundation(61913);
-  static const IconData social_skype = const FlutterIconData.foundation(61914);
-  static const IconData social_smashing_mag =
-      const FlutterIconData.foundation(61915);
-  static const IconData social_snapchat =
-      const FlutterIconData.foundation(61916);
-  static const IconData social_spotify =
-      const FlutterIconData.foundation(61917);
-  static const IconData social_squidoo =
-      const FlutterIconData.foundation(61918);
+      FlutterIconData.foundation(61885);
+  static const IconData social_deviant_art = FlutterIconData.foundation(61886);
+  static const IconData social_digg = FlutterIconData.foundation(61887);
+  static const IconData social_dribbble = FlutterIconData.foundation(61888);
+  static const IconData social_drive = FlutterIconData.foundation(61889);
+  static const IconData social_dropbox = FlutterIconData.foundation(61890);
+  static const IconData social_evernote = FlutterIconData.foundation(61891);
+  static const IconData social_facebook = FlutterIconData.foundation(61892);
+  static const IconData social_flickr = FlutterIconData.foundation(61893);
+  static const IconData social_forrst = FlutterIconData.foundation(61894);
+  static const IconData social_foursquare = FlutterIconData.foundation(61895);
+  static const IconData social_game_center = FlutterIconData.foundation(61896);
+  static const IconData social_github = FlutterIconData.foundation(61897);
+  static const IconData social_google_plus = FlutterIconData.foundation(61898);
+  static const IconData social_hacker_news = FlutterIconData.foundation(61899);
+  static const IconData social_hi5 = FlutterIconData.foundation(61900);
+  static const IconData social_instagram = FlutterIconData.foundation(61901);
+  static const IconData social_joomla = FlutterIconData.foundation(61902);
+  static const IconData social_lastfm = FlutterIconData.foundation(61903);
+  static const IconData social_linkedin = FlutterIconData.foundation(61904);
+  static const IconData social_medium = FlutterIconData.foundation(61905);
+  static const IconData social_myspace = FlutterIconData.foundation(61906);
+  static const IconData social_orkut = FlutterIconData.foundation(61907);
+  static const IconData social_path = FlutterIconData.foundation(61908);
+  static const IconData social_picasa = FlutterIconData.foundation(61909);
+  static const IconData social_pinterest = FlutterIconData.foundation(61910);
+  static const IconData social_rdio = FlutterIconData.foundation(61911);
+  static const IconData social_reddit = FlutterIconData.foundation(61912);
+  static const IconData social_skillshare = FlutterIconData.foundation(61913);
+  static const IconData social_skype = FlutterIconData.foundation(61914);
+  static const IconData social_smashing_mag = FlutterIconData.foundation(61915);
+  static const IconData social_snapchat = FlutterIconData.foundation(61916);
+  static const IconData social_spotify = FlutterIconData.foundation(61917);
+  static const IconData social_squidoo = FlutterIconData.foundation(61918);
   static const IconData social_stack_overflow =
-      const FlutterIconData.foundation(61919);
-  static const IconData social_steam = const FlutterIconData.foundation(61920);
-  static const IconData social_stumbleupon =
-      const FlutterIconData.foundation(61921);
-  static const IconData social_treehouse =
-      const FlutterIconData.foundation(61922);
-  static const IconData social_tumblr = const FlutterIconData.foundation(61923);
-  static const IconData social_twitter =
-      const FlutterIconData.foundation(61924);
-  static const IconData social_vimeo = const FlutterIconData.foundation(61925);
-  static const IconData social_windows =
-      const FlutterIconData.foundation(61926);
-  static const IconData social_xbox = const FlutterIconData.foundation(61927);
-  static const IconData social_yahoo = const FlutterIconData.foundation(61928);
-  static const IconData social_yelp = const FlutterIconData.foundation(61929);
-  static const IconData social_youtube =
-      const FlutterIconData.foundation(61930);
-  static const IconData social_zerply = const FlutterIconData.foundation(61931);
-  static const IconData social_zurb = const FlutterIconData.foundation(61932);
-  static const IconData sound = const FlutterIconData.foundation(61933);
-  static const IconData star = const FlutterIconData.foundation(61934);
-  static const IconData stop = const FlutterIconData.foundation(61935);
-  static const IconData strikethrough = const FlutterIconData.foundation(61936);
-  static const IconData subscript = const FlutterIconData.foundation(61937);
-  static const IconData superscript = const FlutterIconData.foundation(61938);
-  static const IconData tablet_landscape =
-      const FlutterIconData.foundation(61939);
-  static const IconData tablet_portrait =
-      const FlutterIconData.foundation(61940);
-  static const IconData target_two = const FlutterIconData.foundation(61941);
-  static const IconData target = const FlutterIconData.foundation(61942);
+      FlutterIconData.foundation(61919);
+  static const IconData social_steam = FlutterIconData.foundation(61920);
+  static const IconData social_stumbleupon = FlutterIconData.foundation(61921);
+  static const IconData social_treehouse = FlutterIconData.foundation(61922);
+  static const IconData social_tumblr = FlutterIconData.foundation(61923);
+  static const IconData social_twitter = FlutterIconData.foundation(61924);
+  static const IconData social_vimeo = FlutterIconData.foundation(61925);
+  static const IconData social_windows = FlutterIconData.foundation(61926);
+  static const IconData social_xbox = FlutterIconData.foundation(61927);
+  static const IconData social_yahoo = FlutterIconData.foundation(61928);
+  static const IconData social_yelp = FlutterIconData.foundation(61929);
+  static const IconData social_youtube = FlutterIconData.foundation(61930);
+  static const IconData social_zerply = FlutterIconData.foundation(61931);
+  static const IconData social_zurb = FlutterIconData.foundation(61932);
+  static const IconData sound = FlutterIconData.foundation(61933);
+  static const IconData star = FlutterIconData.foundation(61934);
+  static const IconData stop = FlutterIconData.foundation(61935);
+  static const IconData strikethrough = FlutterIconData.foundation(61936);
+  static const IconData subscript = FlutterIconData.foundation(61937);
+  static const IconData superscript = FlutterIconData.foundation(61938);
+  static const IconData tablet_landscape = FlutterIconData.foundation(61939);
+  static const IconData tablet_portrait = FlutterIconData.foundation(61940);
+  static const IconData target_two = FlutterIconData.foundation(61941);
+  static const IconData target = FlutterIconData.foundation(61942);
   static const IconData telephone_accessible =
-      const FlutterIconData.foundation(61943);
-  static const IconData telephone = const FlutterIconData.foundation(61944);
-  static const IconData text_color = const FlutterIconData.foundation(61945);
-  static const IconData thumbnails = const FlutterIconData.foundation(61946);
-  static const IconData ticket = const FlutterIconData.foundation(61947);
-  static const IconData torso_business =
-      const FlutterIconData.foundation(61948);
-  static const IconData torso_female = const FlutterIconData.foundation(61949);
-  static const IconData torso = const FlutterIconData.foundation(61950);
-  static const IconData torsos_all_female =
-      const FlutterIconData.foundation(61951);
-  static const IconData torsos_all = const FlutterIconData.foundation(61952);
-  static const IconData torsos_female_male =
-      const FlutterIconData.foundation(61953);
-  static const IconData torsos_male_female =
-      const FlutterIconData.foundation(61954);
-  static const IconData torsos = const FlutterIconData.foundation(61955);
-  static const IconData trash = const FlutterIconData.foundation(61956);
-  static const IconData trees = const FlutterIconData.foundation(61957);
-  static const IconData trophy = const FlutterIconData.foundation(61958);
-  static const IconData underline = const FlutterIconData.foundation(61959);
-  static const IconData universal_access =
-      const FlutterIconData.foundation(61960);
-  static const IconData unlink = const FlutterIconData.foundation(61961);
-  static const IconData unlock = const FlutterIconData.foundation(61962);
-  static const IconData upload_cloud = const FlutterIconData.foundation(61963);
-  static const IconData upload = const FlutterIconData.foundation(61964);
-  static const IconData usb = const FlutterIconData.foundation(61965);
-  static const IconData video = const FlutterIconData.foundation(61966);
-  static const IconData volume_none = const FlutterIconData.foundation(61967);
-  static const IconData volume_strike = const FlutterIconData.foundation(61968);
-  static const IconData volume = const FlutterIconData.foundation(61969);
-  static const IconData web = const FlutterIconData.foundation(61970);
-  static const IconData wheelchair = const FlutterIconData.foundation(61971);
-  static const IconData widget = const FlutterIconData.foundation(61972);
-  static const IconData wrench = const FlutterIconData.foundation(61973);
-  static const IconData x_circle = const FlutterIconData.foundation(61974);
-  static const IconData x = const FlutterIconData.foundation(61975);
-  static const IconData yen = const FlutterIconData.foundation(61976);
-  static const IconData zoom_in = const FlutterIconData.foundation(61977);
-  static const IconData zoom_out = const FlutterIconData.foundation(61978);
+      FlutterIconData.foundation(61943);
+  static const IconData telephone = FlutterIconData.foundation(61944);
+  static const IconData text_color = FlutterIconData.foundation(61945);
+  static const IconData thumbnails = FlutterIconData.foundation(61946);
+  static const IconData ticket = FlutterIconData.foundation(61947);
+  static const IconData torso_business = FlutterIconData.foundation(61948);
+  static const IconData torso_female = FlutterIconData.foundation(61949);
+  static const IconData torso = FlutterIconData.foundation(61950);
+  static const IconData torsos_all_female = FlutterIconData.foundation(61951);
+  static const IconData torsos_all = FlutterIconData.foundation(61952);
+  static const IconData torsos_female_male = FlutterIconData.foundation(61953);
+  static const IconData torsos_male_female = FlutterIconData.foundation(61954);
+  static const IconData torsos = FlutterIconData.foundation(61955);
+  static const IconData trash = FlutterIconData.foundation(61956);
+  static const IconData trees = FlutterIconData.foundation(61957);
+  static const IconData trophy = FlutterIconData.foundation(61958);
+  static const IconData underline = FlutterIconData.foundation(61959);
+  static const IconData universal_access = FlutterIconData.foundation(61960);
+  static const IconData unlink = FlutterIconData.foundation(61961);
+  static const IconData unlock = FlutterIconData.foundation(61962);
+  static const IconData upload_cloud = FlutterIconData.foundation(61963);
+  static const IconData upload = FlutterIconData.foundation(61964);
+  static const IconData usb = FlutterIconData.foundation(61965);
+  static const IconData video = FlutterIconData.foundation(61966);
+  static const IconData volume_none = FlutterIconData.foundation(61967);
+  static const IconData volume_strike = FlutterIconData.foundation(61968);
+  static const IconData volume = FlutterIconData.foundation(61969);
+  static const IconData web = FlutterIconData.foundation(61970);
+  static const IconData wheelchair = FlutterIconData.foundation(61971);
+  static const IconData widget = FlutterIconData.foundation(61972);
+  static const IconData wrench = FlutterIconData.foundation(61973);
+  static const IconData x_circle = FlutterIconData.foundation(61974);
+  static const IconData x = FlutterIconData.foundation(61975);
+  static const IconData yen = FlutterIconData.foundation(61976);
+  static const IconData zoom_in = FlutterIconData.foundation(61977);
+  static const IconData zoom_out = FlutterIconData.foundation(61978);
 }
