@@ -6,192 +6,183 @@ import 'flutter_icon_data.dart';
 /// http://octicons.github.com
 class Octicons {
   Octicons._();
-  static const IconData alert = const FlutterIconData.octicons(61696);
-  static const IconData archive = const FlutterIconData.octicons(61697);
-  static const IconData arrow_down = const FlutterIconData.octicons(61698);
-  static const IconData arrow_left = const FlutterIconData.octicons(61699);
-  static const IconData arrow_right = const FlutterIconData.octicons(61700);
-  static const IconData arrow_small_down =
-      const FlutterIconData.octicons(61701);
-  static const IconData arrow_small_left =
-      const FlutterIconData.octicons(61702);
-  static const IconData arrow_small_right =
-      const FlutterIconData.octicons(61703);
-  static const IconData arrow_small_up = const FlutterIconData.octicons(61704);
-  static const IconData arrow_up = const FlutterIconData.octicons(61705);
-  static const IconData beaker = const FlutterIconData.octicons(61706);
-  static const IconData bell = const FlutterIconData.octicons(61707);
-  static const IconData bold = const FlutterIconData.octicons(61708);
-  static const IconData book = const FlutterIconData.octicons(61709);
-  static const IconData bookmark = const FlutterIconData.octicons(61710);
-  static const IconData briefcase = const FlutterIconData.octicons(61711);
-  static const IconData broadcast = const FlutterIconData.octicons(61712);
-  static const IconData browser = const FlutterIconData.octicons(61713);
-  static const IconData bug = const FlutterIconData.octicons(61714);
-  static const IconData calendar = const FlutterIconData.octicons(61715);
-  static const IconData check = const FlutterIconData.octicons(61716);
-  static const IconData checklist = const FlutterIconData.octicons(61717);
-  static const IconData chevron_down = const FlutterIconData.octicons(61718);
-  static const IconData chevron_left = const FlutterIconData.octicons(61719);
-  static const IconData chevron_right = const FlutterIconData.octicons(61720);
-  static const IconData chevron_up = const FlutterIconData.octicons(61721);
-  static const IconData circle_slash = const FlutterIconData.octicons(61722);
-  static const IconData circuit_board = const FlutterIconData.octicons(61723);
-  static const IconData clippy = const FlutterIconData.octicons(61724);
-  static const IconData clock = const FlutterIconData.octicons(61725);
-  static const IconData cloud_download = const FlutterIconData.octicons(61726);
-  static const IconData cloud_upload = const FlutterIconData.octicons(61727);
-  static const IconData code = const FlutterIconData.octicons(61728);
-  static const IconData comment = const FlutterIconData.octicons(61729);
-  static const IconData comment_discussion =
-      const FlutterIconData.octicons(61730);
-  static const IconData credit_card = const FlutterIconData.octicons(61731);
-  static const IconData dash = const FlutterIconData.octicons(61732);
-  static const IconData dashboard = const FlutterIconData.octicons(61733);
-  static const IconData database = const FlutterIconData.octicons(61734);
-  static const IconData desktop_download =
-      const FlutterIconData.octicons(61735);
-  static const IconData device_camera = const FlutterIconData.octicons(61736);
-  static const IconData device_camera_video =
-      const FlutterIconData.octicons(61737);
-  static const IconData device_desktop = const FlutterIconData.octicons(61738);
-  static const IconData device_mobile = const FlutterIconData.octicons(61739);
-  static const IconData diff = const FlutterIconData.octicons(61740);
-  static const IconData diff_added = const FlutterIconData.octicons(61741);
-  static const IconData diff_ignored = const FlutterIconData.octicons(61742);
-  static const IconData diff_modified = const FlutterIconData.octicons(61743);
-  static const IconData diff_removed = const FlutterIconData.octicons(61744);
-  static const IconData diff_renamed = const FlutterIconData.octicons(61745);
-  static const IconData ellipsis = const FlutterIconData.octicons(61746);
-  static const IconData eye = const FlutterIconData.octicons(61747);
-  static const IconData file = const FlutterIconData.octicons(61748);
-  static const IconData file_binary = const FlutterIconData.octicons(61749);
-  static const IconData file_code = const FlutterIconData.octicons(61750);
-  static const IconData file_directory = const FlutterIconData.octicons(61751);
-  static const IconData file_media = const FlutterIconData.octicons(61752);
-  static const IconData file_pdf = const FlutterIconData.octicons(61753);
-  static const IconData file_submodule = const FlutterIconData.octicons(61754);
+
+  static const IconData alert = FlutterIconData.octicons(61696);
+  static const IconData archive = FlutterIconData.octicons(61697);
+  static const IconData arrow_down = FlutterIconData.octicons(61698);
+  static const IconData arrow_left = FlutterIconData.octicons(61699);
+  static const IconData arrow_right = FlutterIconData.octicons(61700);
+  static const IconData arrow_small_down = FlutterIconData.octicons(61701);
+  static const IconData arrow_small_left = FlutterIconData.octicons(61702);
+  static const IconData arrow_small_right = FlutterIconData.octicons(61703);
+  static const IconData arrow_small_up = FlutterIconData.octicons(61704);
+  static const IconData arrow_up = FlutterIconData.octicons(61705);
+  static const IconData beaker = FlutterIconData.octicons(61706);
+  static const IconData bell = FlutterIconData.octicons(61707);
+  static const IconData bold = FlutterIconData.octicons(61708);
+  static const IconData book = FlutterIconData.octicons(61709);
+  static const IconData bookmark = FlutterIconData.octicons(61710);
+  static const IconData briefcase = FlutterIconData.octicons(61711);
+  static const IconData broadcast = FlutterIconData.octicons(61712);
+  static const IconData browser = FlutterIconData.octicons(61713);
+  static const IconData bug = FlutterIconData.octicons(61714);
+  static const IconData calendar = FlutterIconData.octicons(61715);
+  static const IconData check = FlutterIconData.octicons(61716);
+  static const IconData checklist = FlutterIconData.octicons(61717);
+  static const IconData chevron_down = FlutterIconData.octicons(61718);
+  static const IconData chevron_left = FlutterIconData.octicons(61719);
+  static const IconData chevron_right = FlutterIconData.octicons(61720);
+  static const IconData chevron_up = FlutterIconData.octicons(61721);
+  static const IconData circle_slash = FlutterIconData.octicons(61722);
+  static const IconData circuit_board = FlutterIconData.octicons(61723);
+  static const IconData clippy = FlutterIconData.octicons(61724);
+  static const IconData clock = FlutterIconData.octicons(61725);
+  static const IconData cloud_download = FlutterIconData.octicons(61726);
+  static const IconData cloud_upload = FlutterIconData.octicons(61727);
+  static const IconData code = FlutterIconData.octicons(61728);
+  static const IconData comment = FlutterIconData.octicons(61729);
+  static const IconData comment_discussion = FlutterIconData.octicons(61730);
+  static const IconData credit_card = FlutterIconData.octicons(61731);
+  static const IconData dash = FlutterIconData.octicons(61732);
+  static const IconData dashboard = FlutterIconData.octicons(61733);
+  static const IconData database = FlutterIconData.octicons(61734);
+  static const IconData desktop_download = FlutterIconData.octicons(61735);
+  static const IconData device_camera = FlutterIconData.octicons(61736);
+  static const IconData device_camera_video = FlutterIconData.octicons(61737);
+  static const IconData device_desktop = FlutterIconData.octicons(61738);
+  static const IconData device_mobile = FlutterIconData.octicons(61739);
+  static const IconData diff = FlutterIconData.octicons(61740);
+  static const IconData diff_added = FlutterIconData.octicons(61741);
+  static const IconData diff_ignored = FlutterIconData.octicons(61742);
+  static const IconData diff_modified = FlutterIconData.octicons(61743);
+  static const IconData diff_removed = FlutterIconData.octicons(61744);
+  static const IconData diff_renamed = FlutterIconData.octicons(61745);
+  static const IconData ellipsis = FlutterIconData.octicons(61746);
+  static const IconData eye = FlutterIconData.octicons(61747);
+  static const IconData file = FlutterIconData.octicons(61748);
+  static const IconData file_binary = FlutterIconData.octicons(61749);
+  static const IconData file_code = FlutterIconData.octicons(61750);
+  static const IconData file_directory = FlutterIconData.octicons(61751);
+  static const IconData file_media = FlutterIconData.octicons(61752);
+  static const IconData file_pdf = FlutterIconData.octicons(61753);
+  static const IconData file_submodule = FlutterIconData.octicons(61754);
   static const IconData file_symlink_directory =
-      const FlutterIconData.octicons(61755);
-  static const IconData file_symlink_file =
-      const FlutterIconData.octicons(61756);
-  static const IconData file_zip = const FlutterIconData.octicons(61757);
-  static const IconData flame = const FlutterIconData.octicons(61758);
-  static const IconData fold = const FlutterIconData.octicons(61759);
-  static const IconData gear = const FlutterIconData.octicons(61760);
-  static const IconData gift = const FlutterIconData.octicons(61761);
-  static const IconData gist = const FlutterIconData.octicons(61762);
-  static const IconData gist_secret = const FlutterIconData.octicons(61763);
-  static const IconData git_branch = const FlutterIconData.octicons(61764);
-  static const IconData git_commit = const FlutterIconData.octicons(61765);
-  static const IconData git_compare = const FlutterIconData.octicons(61766);
-  static const IconData git_merge = const FlutterIconData.octicons(61767);
-  static const IconData git_pull_request =
-      const FlutterIconData.octicons(61768);
-  static const IconData globe = const FlutterIconData.octicons(61769);
-  static const IconData grabber = const FlutterIconData.octicons(61770);
-  static const IconData graph = const FlutterIconData.octicons(61771);
-  static const IconData heart = const FlutterIconData.octicons(61772);
-  static const IconData history = const FlutterIconData.octicons(61773);
-  static const IconData home = const FlutterIconData.octicons(61774);
-  static const IconData horizontal_rule = const FlutterIconData.octicons(61775);
-  static const IconData hubot = const FlutterIconData.octicons(61776);
-  static const IconData inbox = const FlutterIconData.octicons(61777);
-  static const IconData info = const FlutterIconData.octicons(61778);
-  static const IconData issue_closed = const FlutterIconData.octicons(61779);
-  static const IconData issue_opened = const FlutterIconData.octicons(61780);
-  static const IconData issue_reopened = const FlutterIconData.octicons(61781);
-  static const IconData italic = const FlutterIconData.octicons(61782);
-  static const IconData jersey = const FlutterIconData.octicons(61783);
-  static const IconData kebab_horizontal =
-      const FlutterIconData.octicons(61784);
-  static const IconData kebab_vertical = const FlutterIconData.octicons(61785);
-  static const IconData key = const FlutterIconData.octicons(61786);
-  static const IconData keyboard = const FlutterIconData.octicons(61787);
-  static const IconData law = const FlutterIconData.octicons(61788);
-  static const IconData light_bulb = const FlutterIconData.octicons(61789);
-  static const IconData link = const FlutterIconData.octicons(61790);
-  static const IconData link_external = const FlutterIconData.octicons(61791);
-  static const IconData list_ordered = const FlutterIconData.octicons(61792);
-  static const IconData list_unordered = const FlutterIconData.octicons(61793);
-  static const IconData location = const FlutterIconData.octicons(61794);
-  static const IconData lock = const FlutterIconData.octicons(61795);
-  static const IconData logo_gist = const FlutterIconData.octicons(61796);
-  static const IconData logo_github = const FlutterIconData.octicons(61797);
-  static const IconData mail = const FlutterIconData.octicons(61798);
-  static const IconData mail_read = const FlutterIconData.octicons(61799);
-  static const IconData mark_github = const FlutterIconData.octicons(61800);
-  static const IconData markdown = const FlutterIconData.octicons(61801);
-  static const IconData megaphone = const FlutterIconData.octicons(61802);
-  static const IconData mention = const FlutterIconData.octicons(61803);
-  static const IconData milestone = const FlutterIconData.octicons(61804);
-  static const IconData mirror = const FlutterIconData.octicons(61805);
-  static const IconData mortar_board = const FlutterIconData.octicons(61806);
-  static const IconData mute = const FlutterIconData.octicons(61807);
-  static const IconData no_newline = const FlutterIconData.octicons(61808);
-  static const IconData note = const FlutterIconData.octicons(61809);
-  static const IconData octoface = const FlutterIconData.octicons(61810);
-  static const IconData organization = const FlutterIconData.octicons(61811);
-  static const IconData package = const FlutterIconData.octicons(61812);
-  static const IconData paintcan = const FlutterIconData.octicons(61813);
-  static const IconData pencil = const FlutterIconData.octicons(61814);
-  static const IconData person = const FlutterIconData.octicons(61815);
-  static const IconData pin = const FlutterIconData.octicons(61816);
-  static const IconData plug = const FlutterIconData.octicons(61817);
-  static const IconData plus = const FlutterIconData.octicons(61818);
-  static const IconData plus_small = const FlutterIconData.octicons(61819);
-  static const IconData primitive_dot = const FlutterIconData.octicons(61820);
-  static const IconData primitive_square =
-      const FlutterIconData.octicons(61821);
-  static const IconData project = const FlutterIconData.octicons(61822);
-  static const IconData pulse = const FlutterIconData.octicons(61823);
-  static const IconData question = const FlutterIconData.octicons(61824);
-  static const IconData quote = const FlutterIconData.octicons(61825);
-  static const IconData radio_tower = const FlutterIconData.octicons(61826);
-  static const IconData reply = const FlutterIconData.octicons(61827);
-  static const IconData repo = const FlutterIconData.octicons(61828);
-  static const IconData repo_clone = const FlutterIconData.octicons(61829);
-  static const IconData repo_force_push = const FlutterIconData.octicons(61830);
-  static const IconData repo_forked = const FlutterIconData.octicons(61831);
-  static const IconData repo_pull = const FlutterIconData.octicons(61832);
-  static const IconData repo_push = const FlutterIconData.octicons(61833);
-  static const IconData report = const FlutterIconData.octicons(61834);
-  static const IconData rocket = const FlutterIconData.octicons(61835);
-  static const IconData rss = const FlutterIconData.octicons(61836);
-  static const IconData ruby = const FlutterIconData.octicons(61837);
-  static const IconData screen_full = const FlutterIconData.octicons(61838);
-  static const IconData screen_normal = const FlutterIconData.octicons(61839);
-  static const IconData search = const FlutterIconData.octicons(61840);
-  static const IconData server = const FlutterIconData.octicons(61841);
-  static const IconData settings = const FlutterIconData.octicons(61842);
-  static const IconData shield = const FlutterIconData.octicons(61843);
-  static const IconData sign_in = const FlutterIconData.octicons(61844);
-  static const IconData sign_out = const FlutterIconData.octicons(61845);
-  static const IconData smiley = const FlutterIconData.octicons(61846);
-  static const IconData squirrel = const FlutterIconData.octicons(61847);
-  static const IconData star = const FlutterIconData.octicons(61848);
-  static const IconData stop = const FlutterIconData.octicons(61849);
-  static const IconData sync = const FlutterIconData.octicons(61850);
-  static const IconData tag = const FlutterIconData.octicons(61851);
-  static const IconData tasklist = const FlutterIconData.octicons(61852);
-  static const IconData telescope = const FlutterIconData.octicons(61853);
-  static const IconData terminal = const FlutterIconData.octicons(61854);
-  static const IconData text_size = const FlutterIconData.octicons(61855);
-  static const IconData three_bars = const FlutterIconData.octicons(61856);
-  static const IconData thumbsdown = const FlutterIconData.octicons(61857);
-  static const IconData thumbsup = const FlutterIconData.octicons(61858);
-  static const IconData tools = const FlutterIconData.octicons(61859);
-  static const IconData trashcan = const FlutterIconData.octicons(61860);
-  static const IconData triangle_down = const FlutterIconData.octicons(61861);
-  static const IconData triangle_left = const FlutterIconData.octicons(61862);
-  static const IconData triangle_right = const FlutterIconData.octicons(61863);
-  static const IconData triangle_up = const FlutterIconData.octicons(61864);
-  static const IconData unfold = const FlutterIconData.octicons(61865);
-  static const IconData unmute = const FlutterIconData.octicons(61866);
-  static const IconData unverified = const FlutterIconData.octicons(61867);
-  static const IconData verified = const FlutterIconData.octicons(61868);
-  static const IconData versions = const FlutterIconData.octicons(61869);
-  static const IconData watch = const FlutterIconData.octicons(61870);
-  static const IconData x = const FlutterIconData.octicons(61871);
-  static const IconData zap = const FlutterIconData.octicons(61872);
+      FlutterIconData.octicons(61755);
+  static const IconData file_symlink_file = FlutterIconData.octicons(61756);
+  static const IconData file_zip = FlutterIconData.octicons(61757);
+  static const IconData flame = FlutterIconData.octicons(61758);
+  static const IconData fold = FlutterIconData.octicons(61759);
+  static const IconData gear = FlutterIconData.octicons(61760);
+  static const IconData gift = FlutterIconData.octicons(61761);
+  static const IconData gist = FlutterIconData.octicons(61762);
+  static const IconData gist_secret = FlutterIconData.octicons(61763);
+  static const IconData git_branch = FlutterIconData.octicons(61764);
+  static const IconData git_commit = FlutterIconData.octicons(61765);
+  static const IconData git_compare = FlutterIconData.octicons(61766);
+  static const IconData git_merge = FlutterIconData.octicons(61767);
+  static const IconData git_pull_request = FlutterIconData.octicons(61768);
+  static const IconData globe = FlutterIconData.octicons(61769);
+  static const IconData grabber = FlutterIconData.octicons(61770);
+  static const IconData graph = FlutterIconData.octicons(61771);
+  static const IconData heart = FlutterIconData.octicons(61772);
+  static const IconData history = FlutterIconData.octicons(61773);
+  static const IconData home = FlutterIconData.octicons(61774);
+  static const IconData horizontal_rule = FlutterIconData.octicons(61775);
+  static const IconData hubot = FlutterIconData.octicons(61776);
+  static const IconData inbox = FlutterIconData.octicons(61777);
+  static const IconData info = FlutterIconData.octicons(61778);
+  static const IconData issue_closed = FlutterIconData.octicons(61779);
+  static const IconData issue_opened = FlutterIconData.octicons(61780);
+  static const IconData issue_reopened = FlutterIconData.octicons(61781);
+  static const IconData italic = FlutterIconData.octicons(61782);
+  static const IconData jersey = FlutterIconData.octicons(61783);
+  static const IconData kebab_horizontal = FlutterIconData.octicons(61784);
+  static const IconData kebab_vertical = FlutterIconData.octicons(61785);
+  static const IconData key = FlutterIconData.octicons(61786);
+  static const IconData keyboard = FlutterIconData.octicons(61787);
+  static const IconData law = FlutterIconData.octicons(61788);
+  static const IconData light_bulb = FlutterIconData.octicons(61789);
+  static const IconData link = FlutterIconData.octicons(61790);
+  static const IconData link_external = FlutterIconData.octicons(61791);
+  static const IconData list_ordered = FlutterIconData.octicons(61792);
+  static const IconData list_unordered = FlutterIconData.octicons(61793);
+  static const IconData location = FlutterIconData.octicons(61794);
+  static const IconData lock = FlutterIconData.octicons(61795);
+  static const IconData logo_gist = FlutterIconData.octicons(61796);
+  static const IconData logo_github = FlutterIconData.octicons(61797);
+  static const IconData mail = FlutterIconData.octicons(61798);
+  static const IconData mail_read = FlutterIconData.octicons(61799);
+  static const IconData mark_github = FlutterIconData.octicons(61800);
+  static const IconData markdown = FlutterIconData.octicons(61801);
+  static const IconData megaphone = FlutterIconData.octicons(61802);
+  static const IconData mention = FlutterIconData.octicons(61803);
+  static const IconData milestone = FlutterIconData.octicons(61804);
+  static const IconData mirror = FlutterIconData.octicons(61805);
+  static const IconData mortar_board = FlutterIconData.octicons(61806);
+  static const IconData mute = FlutterIconData.octicons(61807);
+  static const IconData no_newline = FlutterIconData.octicons(61808);
+  static const IconData note = FlutterIconData.octicons(61809);
+  static const IconData octoface = FlutterIconData.octicons(61810);
+  static const IconData organization = FlutterIconData.octicons(61811);
+  static const IconData package = FlutterIconData.octicons(61812);
+  static const IconData paintcan = FlutterIconData.octicons(61813);
+  static const IconData pencil = FlutterIconData.octicons(61814);
+  static const IconData person = FlutterIconData.octicons(61815);
+  static const IconData pin = FlutterIconData.octicons(61816);
+  static const IconData plug = FlutterIconData.octicons(61817);
+  static const IconData plus = FlutterIconData.octicons(61818);
+  static const IconData plus_small = FlutterIconData.octicons(61819);
+  static const IconData primitive_dot = FlutterIconData.octicons(61820);
+  static const IconData primitive_square = FlutterIconData.octicons(61821);
+  static const IconData project = FlutterIconData.octicons(61822);
+  static const IconData pulse = FlutterIconData.octicons(61823);
+  static const IconData question = FlutterIconData.octicons(61824);
+  static const IconData quote = FlutterIconData.octicons(61825);
+  static const IconData radio_tower = FlutterIconData.octicons(61826);
+  static const IconData reply = FlutterIconData.octicons(61827);
+  static const IconData repo = FlutterIconData.octicons(61828);
+  static const IconData repo_clone = FlutterIconData.octicons(61829);
+  static const IconData repo_force_push = FlutterIconData.octicons(61830);
+  static const IconData repo_forked = FlutterIconData.octicons(61831);
+  static const IconData repo_pull = FlutterIconData.octicons(61832);
+  static const IconData repo_push = FlutterIconData.octicons(61833);
+  static const IconData report = FlutterIconData.octicons(61834);
+  static const IconData rocket = FlutterIconData.octicons(61835);
+  static const IconData rss = FlutterIconData.octicons(61836);
+  static const IconData ruby = FlutterIconData.octicons(61837);
+  static const IconData screen_full = FlutterIconData.octicons(61838);
+  static const IconData screen_normal = FlutterIconData.octicons(61839);
+  static const IconData search = FlutterIconData.octicons(61840);
+  static const IconData server = FlutterIconData.octicons(61841);
+  static const IconData settings = FlutterIconData.octicons(61842);
+  static const IconData shield = FlutterIconData.octicons(61843);
+  static const IconData sign_in = FlutterIconData.octicons(61844);
+  static const IconData sign_out = FlutterIconData.octicons(61845);
+  static const IconData smiley = FlutterIconData.octicons(61846);
+  static const IconData squirrel = FlutterIconData.octicons(61847);
+  static const IconData star = FlutterIconData.octicons(61848);
+  static const IconData stop = FlutterIconData.octicons(61849);
+  static const IconData sync = FlutterIconData.octicons(61850);
+  static const IconData tag = FlutterIconData.octicons(61851);
+  static const IconData tasklist = FlutterIconData.octicons(61852);
+  static const IconData telescope = FlutterIconData.octicons(61853);
+  static const IconData terminal = FlutterIconData.octicons(61854);
+  static const IconData text_size = FlutterIconData.octicons(61855);
+  static const IconData three_bars = FlutterIconData.octicons(61856);
+  static const IconData thumbsdown = FlutterIconData.octicons(61857);
+  static const IconData thumbsup = FlutterIconData.octicons(61858);
+  static const IconData tools = FlutterIconData.octicons(61859);
+  static const IconData trashcan = FlutterIconData.octicons(61860);
+  static const IconData triangle_down = FlutterIconData.octicons(61861);
+  static const IconData triangle_left = FlutterIconData.octicons(61862);
+  static const IconData triangle_right = FlutterIconData.octicons(61863);
+  static const IconData triangle_up = FlutterIconData.octicons(61864);
+  static const IconData unfold = FlutterIconData.octicons(61865);
+  static const IconData unmute = FlutterIconData.octicons(61866);
+  static const IconData unverified = FlutterIconData.octicons(61867);
+  static const IconData verified = FlutterIconData.octicons(61868);
+  static const IconData versions = FlutterIconData.octicons(61869);
+  static const IconData watch = FlutterIconData.octicons(61870);
+  static const IconData x = FlutterIconData.octicons(61871);
+  static const IconData zap = FlutterIconData.octicons(61872);
 }
