@@ -6,280 +6,214 @@ import 'flutter_icon_data.dart';
 /// https://simplelineicons.github.io/
 class SimpleLineIcons {
   SimpleLineIcons._();
-  static const IconData user = const FlutterIconData.simpleLineIcons(57349);
-  static const IconData people = const FlutterIconData.simpleLineIcons(57345);
-  static const IconData user_female =
-      const FlutterIconData.simpleLineIcons(57344);
-  static const IconData user_follow =
-      const FlutterIconData.simpleLineIcons(57346);
-  static const IconData user_following =
-      const FlutterIconData.simpleLineIcons(57347);
-  static const IconData user_unfollow =
-      const FlutterIconData.simpleLineIcons(57348);
-  static const IconData login = const FlutterIconData.simpleLineIcons(57446);
-  static const IconData logout = const FlutterIconData.simpleLineIcons(57445);
-  static const IconData emotsmile =
-      const FlutterIconData.simpleLineIcons(57377);
-  static const IconData phone = const FlutterIconData.simpleLineIcons(58880);
-  static const IconData call_end = const FlutterIconData.simpleLineIcons(57416);
-  static const IconData call_in = const FlutterIconData.simpleLineIcons(57415);
-  static const IconData call_out = const FlutterIconData.simpleLineIcons(57414);
-  static const IconData map = const FlutterIconData.simpleLineIcons(57395);
-  static const IconData location_pin =
-      const FlutterIconData.simpleLineIcons(57494);
-  static const IconData direction =
-      const FlutterIconData.simpleLineIcons(57410);
-  static const IconData directions =
-      const FlutterIconData.simpleLineIcons(57409);
-  static const IconData compass = const FlutterIconData.simpleLineIcons(57413);
-  static const IconData layers = const FlutterIconData.simpleLineIcons(57396);
-  static const IconData menu = const FlutterIconData.simpleLineIcons(58881);
-  static const IconData list = const FlutterIconData.simpleLineIcons(57447);
+
+  static const IconData user = FlutterIconData.simpleLineIcons(57349);
+  static const IconData people = FlutterIconData.simpleLineIcons(57345);
+  static const IconData user_female = FlutterIconData.simpleLineIcons(57344);
+  static const IconData user_follow = FlutterIconData.simpleLineIcons(57346);
+  static const IconData user_following = FlutterIconData.simpleLineIcons(57347);
+  static const IconData user_unfollow = FlutterIconData.simpleLineIcons(57348);
+  static const IconData login = FlutterIconData.simpleLineIcons(57446);
+  static const IconData logout = FlutterIconData.simpleLineIcons(57445);
+  static const IconData emotsmile = FlutterIconData.simpleLineIcons(57377);
+  static const IconData phone = FlutterIconData.simpleLineIcons(58880);
+  static const IconData call_end = FlutterIconData.simpleLineIcons(57416);
+  static const IconData call_in = FlutterIconData.simpleLineIcons(57415);
+  static const IconData call_out = FlutterIconData.simpleLineIcons(57414);
+  static const IconData map = FlutterIconData.simpleLineIcons(57395);
+  static const IconData location_pin = FlutterIconData.simpleLineIcons(57494);
+  static const IconData direction = FlutterIconData.simpleLineIcons(57410);
+  static const IconData directions = FlutterIconData.simpleLineIcons(57409);
+  static const IconData compass = FlutterIconData.simpleLineIcons(57413);
+  static const IconData layers = FlutterIconData.simpleLineIcons(57396);
+  static const IconData menu = FlutterIconData.simpleLineIcons(58881);
+  static const IconData list = FlutterIconData.simpleLineIcons(57447);
   static const IconData options_vertical =
-      const FlutterIconData.simpleLineIcons(58882);
-  static const IconData options = const FlutterIconData.simpleLineIcons(58883);
-  static const IconData arrow_down =
-      const FlutterIconData.simpleLineIcons(58884);
-  static const IconData arrow_left =
-      const FlutterIconData.simpleLineIcons(58885);
-  static const IconData arrow_right =
-      const FlutterIconData.simpleLineIcons(58886);
-  static const IconData arrow_up = const FlutterIconData.simpleLineIcons(58887);
+      FlutterIconData.simpleLineIcons(58882);
+  static const IconData options = FlutterIconData.simpleLineIcons(58883);
+  static const IconData arrow_down = FlutterIconData.simpleLineIcons(58884);
+  static const IconData arrow_left = FlutterIconData.simpleLineIcons(58885);
+  static const IconData arrow_right = FlutterIconData.simpleLineIcons(58886);
+  static const IconData arrow_up = FlutterIconData.simpleLineIcons(58887);
   static const IconData arrow_up_circle =
-      const FlutterIconData.simpleLineIcons(57464);
+      FlutterIconData.simpleLineIcons(57464);
   static const IconData arrow_left_circle =
-      const FlutterIconData.simpleLineIcons(57466);
+      FlutterIconData.simpleLineIcons(57466);
   static const IconData arrow_right_circle =
-      const FlutterIconData.simpleLineIcons(57465);
+      FlutterIconData.simpleLineIcons(57465);
   static const IconData arrow_down_circle =
-      const FlutterIconData.simpleLineIcons(57467);
-  static const IconData check = const FlutterIconData.simpleLineIcons(57472);
-  static const IconData clock = const FlutterIconData.simpleLineIcons(57473);
-  static const IconData plus = const FlutterIconData.simpleLineIcons(57493);
-  static const IconData minus = const FlutterIconData.simpleLineIcons(58901);
-  static const IconData close = const FlutterIconData.simpleLineIcons(57474);
-  static const IconData event = const FlutterIconData.simpleLineIcons(58905);
-  static const IconData exclamation =
-      const FlutterIconData.simpleLineIcons(58903);
-  static const IconData organization =
-      const FlutterIconData.simpleLineIcons(58902);
-  static const IconData trophy = const FlutterIconData.simpleLineIcons(57350);
+      FlutterIconData.simpleLineIcons(57467);
+  static const IconData check = FlutterIconData.simpleLineIcons(57472);
+  static const IconData clock = FlutterIconData.simpleLineIcons(57473);
+  static const IconData plus = FlutterIconData.simpleLineIcons(57493);
+  static const IconData minus = FlutterIconData.simpleLineIcons(58901);
+  static const IconData close = FlutterIconData.simpleLineIcons(57474);
+  static const IconData event = FlutterIconData.simpleLineIcons(58905);
+  static const IconData exclamation = FlutterIconData.simpleLineIcons(58903);
+  static const IconData organization = FlutterIconData.simpleLineIcons(58902);
+  static const IconData trophy = FlutterIconData.simpleLineIcons(57350);
   static const IconData screen_smartphone =
-      const FlutterIconData.simpleLineIcons(57360);
-  static const IconData screen_desktop =
-      const FlutterIconData.simpleLineIcons(57361);
-  static const IconData plane = const FlutterIconData.simpleLineIcons(57362);
-  static const IconData notebook = const FlutterIconData.simpleLineIcons(57363);
-  static const IconData mustache = const FlutterIconData.simpleLineIcons(57364);
-  static const IconData mouse = const FlutterIconData.simpleLineIcons(57365);
-  static const IconData magnet = const FlutterIconData.simpleLineIcons(57366);
-  static const IconData energy = const FlutterIconData.simpleLineIcons(57376);
-  static const IconData disc = const FlutterIconData.simpleLineIcons(57378);
-  static const IconData cursor = const FlutterIconData.simpleLineIcons(57454);
-  static const IconData cursor_move =
-      const FlutterIconData.simpleLineIcons(57379);
-  static const IconData crop = const FlutterIconData.simpleLineIcons(57380);
-  static const IconData chemistry =
-      const FlutterIconData.simpleLineIcons(57382);
-  static const IconData speedometer =
-      const FlutterIconData.simpleLineIcons(57351);
-  static const IconData shield = const FlutterIconData.simpleLineIcons(57358);
-  static const IconData screen_tablet =
-      const FlutterIconData.simpleLineIcons(57359);
-  static const IconData magic_wand =
-      const FlutterIconData.simpleLineIcons(57367);
-  static const IconData hourglass =
-      const FlutterIconData.simpleLineIcons(57368);
-  static const IconData graduation =
-      const FlutterIconData.simpleLineIcons(57369);
-  static const IconData ghost = const FlutterIconData.simpleLineIcons(57370);
+      FlutterIconData.simpleLineIcons(57360);
+  static const IconData screen_desktop = FlutterIconData.simpleLineIcons(57361);
+  static const IconData plane = FlutterIconData.simpleLineIcons(57362);
+  static const IconData notebook = FlutterIconData.simpleLineIcons(57363);
+  static const IconData mustache = FlutterIconData.simpleLineIcons(57364);
+  static const IconData mouse = FlutterIconData.simpleLineIcons(57365);
+  static const IconData magnet = FlutterIconData.simpleLineIcons(57366);
+  static const IconData energy = FlutterIconData.simpleLineIcons(57376);
+  static const IconData disc = FlutterIconData.simpleLineIcons(57378);
+  static const IconData cursor = FlutterIconData.simpleLineIcons(57454);
+  static const IconData cursor_move = FlutterIconData.simpleLineIcons(57379);
+  static const IconData crop = FlutterIconData.simpleLineIcons(57380);
+  static const IconData chemistry = FlutterIconData.simpleLineIcons(57382);
+  static const IconData speedometer = FlutterIconData.simpleLineIcons(57351);
+  static const IconData shield = FlutterIconData.simpleLineIcons(57358);
+  static const IconData screen_tablet = FlutterIconData.simpleLineIcons(57359);
+  static const IconData magic_wand = FlutterIconData.simpleLineIcons(57367);
+  static const IconData hourglass = FlutterIconData.simpleLineIcons(57368);
+  static const IconData graduation = FlutterIconData.simpleLineIcons(57369);
+  static const IconData ghost = FlutterIconData.simpleLineIcons(57370);
   static const IconData game_controller =
-      const FlutterIconData.simpleLineIcons(57371);
-  static const IconData fire = const FlutterIconData.simpleLineIcons(57372);
-  static const IconData eyeglass = const FlutterIconData.simpleLineIcons(57373);
-  static const IconData envelope_open =
-      const FlutterIconData.simpleLineIcons(57374);
+      FlutterIconData.simpleLineIcons(57371);
+  static const IconData fire = FlutterIconData.simpleLineIcons(57372);
+  static const IconData eyeglass = FlutterIconData.simpleLineIcons(57373);
+  static const IconData envelope_open = FlutterIconData.simpleLineIcons(57374);
   static const IconData envelope_letter =
-      const FlutterIconData.simpleLineIcons(57375);
-  static const IconData bell = const FlutterIconData.simpleLineIcons(57383);
-  static const IconData badge = const FlutterIconData.simpleLineIcons(57384);
-  static const IconData anchor = const FlutterIconData.simpleLineIcons(57385);
-  static const IconData wallet = const FlutterIconData.simpleLineIcons(57386);
-  static const IconData vector = const FlutterIconData.simpleLineIcons(57387);
-  static const IconData speech = const FlutterIconData.simpleLineIcons(57388);
-  static const IconData puzzle = const FlutterIconData.simpleLineIcons(57389);
-  static const IconData printer = const FlutterIconData.simpleLineIcons(57390);
-  static const IconData present = const FlutterIconData.simpleLineIcons(57391);
-  static const IconData playlist = const FlutterIconData.simpleLineIcons(57392);
-  static const IconData pin = const FlutterIconData.simpleLineIcons(57393);
-  static const IconData picture = const FlutterIconData.simpleLineIcons(57394);
-  static const IconData handbag = const FlutterIconData.simpleLineIcons(57397);
-  static const IconData globe_alt =
-      const FlutterIconData.simpleLineIcons(57398);
-  static const IconData globe = const FlutterIconData.simpleLineIcons(57399);
-  static const IconData folder_alt =
-      const FlutterIconData.simpleLineIcons(57401);
-  static const IconData folder = const FlutterIconData.simpleLineIcons(57481);
-  static const IconData film = const FlutterIconData.simpleLineIcons(57402);
-  static const IconData feed = const FlutterIconData.simpleLineIcons(57403);
-  static const IconData drop = const FlutterIconData.simpleLineIcons(57406);
-  static const IconData drawer = const FlutterIconData.simpleLineIcons(57407);
-  static const IconData docs = const FlutterIconData.simpleLineIcons(57408);
-  static const IconData doc = const FlutterIconData.simpleLineIcons(57477);
-  static const IconData diamond = const FlutterIconData.simpleLineIcons(57411);
-  static const IconData cup = const FlutterIconData.simpleLineIcons(57412);
-  static const IconData calculator =
-      const FlutterIconData.simpleLineIcons(57417);
-  static const IconData bubbles = const FlutterIconData.simpleLineIcons(57418);
-  static const IconData briefcase =
-      const FlutterIconData.simpleLineIcons(57419);
-  static const IconData book_open =
-      const FlutterIconData.simpleLineIcons(57420);
-  static const IconData basket_loaded =
-      const FlutterIconData.simpleLineIcons(57421);
-  static const IconData basket = const FlutterIconData.simpleLineIcons(57422);
-  static const IconData bag = const FlutterIconData.simpleLineIcons(57423);
-  static const IconData action_undo =
-      const FlutterIconData.simpleLineIcons(57424);
-  static const IconData action_redo =
-      const FlutterIconData.simpleLineIcons(57425);
-  static const IconData wrench = const FlutterIconData.simpleLineIcons(57426);
-  static const IconData umbrella = const FlutterIconData.simpleLineIcons(57427);
-  static const IconData trash = const FlutterIconData.simpleLineIcons(57428);
-  static const IconData tag = const FlutterIconData.simpleLineIcons(57429);
-  static const IconData support = const FlutterIconData.simpleLineIcons(57430);
-  static const IconData frame = const FlutterIconData.simpleLineIcons(57400);
+      FlutterIconData.simpleLineIcons(57375);
+  static const IconData bell = FlutterIconData.simpleLineIcons(57383);
+  static const IconData badge = FlutterIconData.simpleLineIcons(57384);
+  static const IconData anchor = FlutterIconData.simpleLineIcons(57385);
+  static const IconData wallet = FlutterIconData.simpleLineIcons(57386);
+  static const IconData vector = FlutterIconData.simpleLineIcons(57387);
+  static const IconData speech = FlutterIconData.simpleLineIcons(57388);
+  static const IconData puzzle = FlutterIconData.simpleLineIcons(57389);
+  static const IconData printer = FlutterIconData.simpleLineIcons(57390);
+  static const IconData present = FlutterIconData.simpleLineIcons(57391);
+  static const IconData playlist = FlutterIconData.simpleLineIcons(57392);
+  static const IconData pin = FlutterIconData.simpleLineIcons(57393);
+  static const IconData picture = FlutterIconData.simpleLineIcons(57394);
+  static const IconData handbag = FlutterIconData.simpleLineIcons(57397);
+  static const IconData globe_alt = FlutterIconData.simpleLineIcons(57398);
+  static const IconData globe = FlutterIconData.simpleLineIcons(57399);
+  static const IconData folder_alt = FlutterIconData.simpleLineIcons(57401);
+  static const IconData folder = FlutterIconData.simpleLineIcons(57481);
+  static const IconData film = FlutterIconData.simpleLineIcons(57402);
+  static const IconData feed = FlutterIconData.simpleLineIcons(57403);
+  static const IconData drop = FlutterIconData.simpleLineIcons(57406);
+  static const IconData drawer = FlutterIconData.simpleLineIcons(57407);
+  static const IconData docs = FlutterIconData.simpleLineIcons(57408);
+  static const IconData doc = FlutterIconData.simpleLineIcons(57477);
+  static const IconData diamond = FlutterIconData.simpleLineIcons(57411);
+  static const IconData cup = FlutterIconData.simpleLineIcons(57412);
+  static const IconData calculator = FlutterIconData.simpleLineIcons(57417);
+  static const IconData bubbles = FlutterIconData.simpleLineIcons(57418);
+  static const IconData briefcase = FlutterIconData.simpleLineIcons(57419);
+  static const IconData book_open = FlutterIconData.simpleLineIcons(57420);
+  static const IconData basket_loaded = FlutterIconData.simpleLineIcons(57421);
+  static const IconData basket = FlutterIconData.simpleLineIcons(57422);
+  static const IconData bag = FlutterIconData.simpleLineIcons(57423);
+  static const IconData action_undo = FlutterIconData.simpleLineIcons(57424);
+  static const IconData action_redo = FlutterIconData.simpleLineIcons(57425);
+  static const IconData wrench = FlutterIconData.simpleLineIcons(57426);
+  static const IconData umbrella = FlutterIconData.simpleLineIcons(57427);
+  static const IconData trash = FlutterIconData.simpleLineIcons(57428);
+  static const IconData tag = FlutterIconData.simpleLineIcons(57429);
+  static const IconData support = FlutterIconData.simpleLineIcons(57430);
+  static const IconData frame = FlutterIconData.simpleLineIcons(57400);
   static const IconData size_fullscreen =
-      const FlutterIconData.simpleLineIcons(57431);
-  static const IconData size_actual =
-      const FlutterIconData.simpleLineIcons(57432);
-  static const IconData shuffle = const FlutterIconData.simpleLineIcons(57433);
-  static const IconData share_alt =
-      const FlutterIconData.simpleLineIcons(57434);
-  static const IconData share = const FlutterIconData.simpleLineIcons(57435);
-  static const IconData rocket = const FlutterIconData.simpleLineIcons(57436);
-  static const IconData question = const FlutterIconData.simpleLineIcons(57437);
-  static const IconData pie_chart =
-      const FlutterIconData.simpleLineIcons(57438);
-  static const IconData pencil = const FlutterIconData.simpleLineIcons(57439);
-  static const IconData note = const FlutterIconData.simpleLineIcons(57440);
-  static const IconData loop = const FlutterIconData.simpleLineIcons(57444);
-  static const IconData home = const FlutterIconData.simpleLineIcons(57449);
-  static const IconData grid = const FlutterIconData.simpleLineIcons(57450);
-  static const IconData graph = const FlutterIconData.simpleLineIcons(57451);
-  static const IconData microphone =
-      const FlutterIconData.simpleLineIcons(57443);
-  static const IconData music_tone_alt =
-      const FlutterIconData.simpleLineIcons(57441);
-  static const IconData music_tone =
-      const FlutterIconData.simpleLineIcons(57442);
-  static const IconData earphones_alt =
-      const FlutterIconData.simpleLineIcons(57404);
-  static const IconData earphones =
-      const FlutterIconData.simpleLineIcons(57405);
-  static const IconData equalizer =
-      const FlutterIconData.simpleLineIcons(57452);
-  static const IconData like = const FlutterIconData.simpleLineIcons(57448);
-  static const IconData dislike = const FlutterIconData.simpleLineIcons(57453);
-  static const IconData control_start =
-      const FlutterIconData.simpleLineIcons(57455);
-  static const IconData control_rewind =
-      const FlutterIconData.simpleLineIcons(57456);
-  static const IconData control_play =
-      const FlutterIconData.simpleLineIcons(57457);
-  static const IconData control_pause =
-      const FlutterIconData.simpleLineIcons(57458);
+      FlutterIconData.simpleLineIcons(57431);
+  static const IconData size_actual = FlutterIconData.simpleLineIcons(57432);
+  static const IconData shuffle = FlutterIconData.simpleLineIcons(57433);
+  static const IconData share_alt = FlutterIconData.simpleLineIcons(57434);
+  static const IconData share = FlutterIconData.simpleLineIcons(57435);
+  static const IconData rocket = FlutterIconData.simpleLineIcons(57436);
+  static const IconData question = FlutterIconData.simpleLineIcons(57437);
+  static const IconData pie_chart = FlutterIconData.simpleLineIcons(57438);
+  static const IconData pencil = FlutterIconData.simpleLineIcons(57439);
+  static const IconData note = FlutterIconData.simpleLineIcons(57440);
+  static const IconData loop = FlutterIconData.simpleLineIcons(57444);
+  static const IconData home = FlutterIconData.simpleLineIcons(57449);
+  static const IconData grid = FlutterIconData.simpleLineIcons(57450);
+  static const IconData graph = FlutterIconData.simpleLineIcons(57451);
+  static const IconData microphone = FlutterIconData.simpleLineIcons(57443);
+  static const IconData music_tone_alt = FlutterIconData.simpleLineIcons(57441);
+  static const IconData music_tone = FlutterIconData.simpleLineIcons(57442);
+  static const IconData earphones_alt = FlutterIconData.simpleLineIcons(57404);
+  static const IconData earphones = FlutterIconData.simpleLineIcons(57405);
+  static const IconData equalizer = FlutterIconData.simpleLineIcons(57452);
+  static const IconData like = FlutterIconData.simpleLineIcons(57448);
+  static const IconData dislike = FlutterIconData.simpleLineIcons(57453);
+  static const IconData control_start = FlutterIconData.simpleLineIcons(57455);
+  static const IconData control_rewind = FlutterIconData.simpleLineIcons(57456);
+  static const IconData control_play = FlutterIconData.simpleLineIcons(57457);
+  static const IconData control_pause = FlutterIconData.simpleLineIcons(57458);
   static const IconData control_forward =
-      const FlutterIconData.simpleLineIcons(57459);
-  static const IconData control_end =
-      const FlutterIconData.simpleLineIcons(57460);
-  static const IconData volume_1 = const FlutterIconData.simpleLineIcons(57503);
-  static const IconData volume_2 = const FlutterIconData.simpleLineIcons(57504);
-  static const IconData volume_off =
-      const FlutterIconData.simpleLineIcons(57505);
-  static const IconData calendar = const FlutterIconData.simpleLineIcons(57461);
-  static const IconData bulb = const FlutterIconData.simpleLineIcons(57462);
-  static const IconData chart = const FlutterIconData.simpleLineIcons(57463);
-  static const IconData ban = const FlutterIconData.simpleLineIcons(57468);
-  static const IconData bubble = const FlutterIconData.simpleLineIcons(57469);
-  static const IconData camrecorder =
-      const FlutterIconData.simpleLineIcons(57470);
-  static const IconData camera = const FlutterIconData.simpleLineIcons(57471);
-  static const IconData cloud_download =
-      const FlutterIconData.simpleLineIcons(57475);
-  static const IconData cloud_upload =
-      const FlutterIconData.simpleLineIcons(57476);
-  static const IconData envelope = const FlutterIconData.simpleLineIcons(57478);
-  static const IconData eye = const FlutterIconData.simpleLineIcons(57479);
-  static const IconData flag = const FlutterIconData.simpleLineIcons(57480);
-  static const IconData heart = const FlutterIconData.simpleLineIcons(57482);
-  static const IconData info = const FlutterIconData.simpleLineIcons(57483);
-  static const IconData key = const FlutterIconData.simpleLineIcons(57484);
-  static const IconData link = const FlutterIconData.simpleLineIcons(57485);
-  static const IconData lock = const FlutterIconData.simpleLineIcons(57486);
-  static const IconData lock_open =
-      const FlutterIconData.simpleLineIcons(57487);
-  static const IconData magnifier =
-      const FlutterIconData.simpleLineIcons(57488);
-  static const IconData magnifier_add =
-      const FlutterIconData.simpleLineIcons(57489);
+      FlutterIconData.simpleLineIcons(57459);
+  static const IconData control_end = FlutterIconData.simpleLineIcons(57460);
+  static const IconData volume_1 = FlutterIconData.simpleLineIcons(57503);
+  static const IconData volume_2 = FlutterIconData.simpleLineIcons(57504);
+  static const IconData volume_off = FlutterIconData.simpleLineIcons(57505);
+  static const IconData calendar = FlutterIconData.simpleLineIcons(57461);
+  static const IconData bulb = FlutterIconData.simpleLineIcons(57462);
+  static const IconData chart = FlutterIconData.simpleLineIcons(57463);
+  static const IconData ban = FlutterIconData.simpleLineIcons(57468);
+  static const IconData bubble = FlutterIconData.simpleLineIcons(57469);
+  static const IconData camrecorder = FlutterIconData.simpleLineIcons(57470);
+  static const IconData camera = FlutterIconData.simpleLineIcons(57471);
+  static const IconData cloud_download = FlutterIconData.simpleLineIcons(57475);
+  static const IconData cloud_upload = FlutterIconData.simpleLineIcons(57476);
+  static const IconData envelope = FlutterIconData.simpleLineIcons(57478);
+  static const IconData eye = FlutterIconData.simpleLineIcons(57479);
+  static const IconData flag = FlutterIconData.simpleLineIcons(57480);
+  static const IconData heart = FlutterIconData.simpleLineIcons(57482);
+  static const IconData info = FlutterIconData.simpleLineIcons(57483);
+  static const IconData key = FlutterIconData.simpleLineIcons(57484);
+  static const IconData link = FlutterIconData.simpleLineIcons(57485);
+  static const IconData lock = FlutterIconData.simpleLineIcons(57486);
+  static const IconData lock_open = FlutterIconData.simpleLineIcons(57487);
+  static const IconData magnifier = FlutterIconData.simpleLineIcons(57488);
+  static const IconData magnifier_add = FlutterIconData.simpleLineIcons(57489);
   static const IconData magnifier_remove =
-      const FlutterIconData.simpleLineIcons(57490);
-  static const IconData paper_clip =
-      const FlutterIconData.simpleLineIcons(57491);
-  static const IconData paper_plane =
-      const FlutterIconData.simpleLineIcons(57492);
-  static const IconData power = const FlutterIconData.simpleLineIcons(57495);
-  static const IconData refresh = const FlutterIconData.simpleLineIcons(57496);
-  static const IconData reload = const FlutterIconData.simpleLineIcons(57497);
-  static const IconData settings = const FlutterIconData.simpleLineIcons(57498);
-  static const IconData star = const FlutterIconData.simpleLineIcons(57499);
-  static const IconData symbol_female =
-      const FlutterIconData.simpleLineIcons(57500);
-  static const IconData symbol_male =
-      const FlutterIconData.simpleLineIcons(57501);
-  static const IconData target = const FlutterIconData.simpleLineIcons(57502);
-  static const IconData credit_card =
-      const FlutterIconData.simpleLineIcons(57381);
-  static const IconData paypal = const FlutterIconData.simpleLineIcons(58888);
-  static const IconData social_tumblr =
-      const FlutterIconData.simpleLineIcons(57354);
-  static const IconData social_twitter =
-      const FlutterIconData.simpleLineIcons(57353);
+      FlutterIconData.simpleLineIcons(57490);
+  static const IconData paper_clip = FlutterIconData.simpleLineIcons(57491);
+  static const IconData paper_plane = FlutterIconData.simpleLineIcons(57492);
+  static const IconData power = FlutterIconData.simpleLineIcons(57495);
+  static const IconData refresh = FlutterIconData.simpleLineIcons(57496);
+  static const IconData reload = FlutterIconData.simpleLineIcons(57497);
+  static const IconData settings = FlutterIconData.simpleLineIcons(57498);
+  static const IconData star = FlutterIconData.simpleLineIcons(57499);
+  static const IconData symbol_female = FlutterIconData.simpleLineIcons(57500);
+  static const IconData symbol_male = FlutterIconData.simpleLineIcons(57501);
+  static const IconData target = FlutterIconData.simpleLineIcons(57502);
+  static const IconData credit_card = FlutterIconData.simpleLineIcons(57381);
+  static const IconData paypal = FlutterIconData.simpleLineIcons(58888);
+  static const IconData social_tumblr = FlutterIconData.simpleLineIcons(57354);
+  static const IconData social_twitter = FlutterIconData.simpleLineIcons(57353);
   static const IconData social_facebook =
-      const FlutterIconData.simpleLineIcons(57355);
+      FlutterIconData.simpleLineIcons(57355);
   static const IconData social_instagram =
-      const FlutterIconData.simpleLineIcons(58889);
+      FlutterIconData.simpleLineIcons(58889);
   static const IconData social_linkedin =
-      const FlutterIconData.simpleLineIcons(58890);
+      FlutterIconData.simpleLineIcons(58890);
   static const IconData social_pinterest =
-      const FlutterIconData.simpleLineIcons(58891);
-  static const IconData social_github =
-      const FlutterIconData.simpleLineIcons(58892);
-  static const IconData social_google =
-      const FlutterIconData.simpleLineIcons(58893);
-  static const IconData social_reddit =
-      const FlutterIconData.simpleLineIcons(58894);
-  static const IconData social_skype =
-      const FlutterIconData.simpleLineIcons(58895);
+      FlutterIconData.simpleLineIcons(58891);
+  static const IconData social_github = FlutterIconData.simpleLineIcons(58892);
+  static const IconData social_google = FlutterIconData.simpleLineIcons(58893);
+  static const IconData social_reddit = FlutterIconData.simpleLineIcons(58894);
+  static const IconData social_skype = FlutterIconData.simpleLineIcons(58895);
   static const IconData social_dribbble =
-      const FlutterIconData.simpleLineIcons(57357);
-  static const IconData social_behance =
-      const FlutterIconData.simpleLineIcons(58896);
+      FlutterIconData.simpleLineIcons(57357);
+  static const IconData social_behance = FlutterIconData.simpleLineIcons(58896);
   static const IconData social_foursqare =
-      const FlutterIconData.simpleLineIcons(58897);
+      FlutterIconData.simpleLineIcons(58897);
   static const IconData social_soundcloud =
-      const FlutterIconData.simpleLineIcons(58898);
-  static const IconData social_spotify =
-      const FlutterIconData.simpleLineIcons(58899);
+      FlutterIconData.simpleLineIcons(58898);
+  static const IconData social_spotify = FlutterIconData.simpleLineIcons(58899);
   static const IconData social_stumbleupon =
-      const FlutterIconData.simpleLineIcons(58900);
-  static const IconData social_youtube =
-      const FlutterIconData.simpleLineIcons(57352);
-  static const IconData social_dropbox =
-      const FlutterIconData.simpleLineIcons(57356);
+      FlutterIconData.simpleLineIcons(58900);
+  static const IconData social_youtube = FlutterIconData.simpleLineIcons(57352);
+  static const IconData social_dropbox = FlutterIconData.simpleLineIcons(57356);
   static const IconData social_vkontakte =
-      const FlutterIconData.simpleLineIcons(58904);
-  static const IconData social_steam =
-      const FlutterIconData.simpleLineIcons(58912);
+      FlutterIconData.simpleLineIcons(58904);
+  static const IconData social_steam = FlutterIconData.simpleLineIcons(58912);
 }
