@@ -6,998 +6,727 @@ import 'flutter_icon_data.dart';
 /// https://erikflowers.github.io/weather-icons/
 class WeatherIcons {
   WeatherIcons._();
-  static const IconData wi_day_sunny =
-      const FlutterIconData.weatherIcons(61453);
-  static const IconData wi_day_cloudy =
-      const FlutterIconData.weatherIcons(61442);
+
+  static const IconData wi_day_sunny = FlutterIconData.weatherIcons(61453);
+  static const IconData wi_day_cloudy = FlutterIconData.weatherIcons(61442);
   static const IconData wi_day_cloudy_gusts =
-      const FlutterIconData.weatherIcons(61440);
+      FlutterIconData.weatherIcons(61440);
   static const IconData wi_day_cloudy_windy =
-      const FlutterIconData.weatherIcons(61441);
-  static const IconData wi_day_fog = const FlutterIconData.weatherIcons(61443);
-  static const IconData wi_day_hail = const FlutterIconData.weatherIcons(61444);
-  static const IconData wi_day_haze = const FlutterIconData.weatherIcons(61622);
-  static const IconData wi_day_lightning =
-      const FlutterIconData.weatherIcons(61445);
-  static const IconData wi_day_rain = const FlutterIconData.weatherIcons(61448);
-  static const IconData wi_day_rain_mix =
-      const FlutterIconData.weatherIcons(61446);
-  static const IconData wi_day_rain_wind =
-      const FlutterIconData.weatherIcons(61447);
-  static const IconData wi_day_showers =
-      const FlutterIconData.weatherIcons(61449);
-  static const IconData wi_day_sleet =
-      const FlutterIconData.weatherIcons(61618);
+      FlutterIconData.weatherIcons(61441);
+  static const IconData wi_day_fog = FlutterIconData.weatherIcons(61443);
+  static const IconData wi_day_hail = FlutterIconData.weatherIcons(61444);
+  static const IconData wi_day_haze = FlutterIconData.weatherIcons(61622);
+  static const IconData wi_day_lightning = FlutterIconData.weatherIcons(61445);
+  static const IconData wi_day_rain = FlutterIconData.weatherIcons(61448);
+  static const IconData wi_day_rain_mix = FlutterIconData.weatherIcons(61446);
+  static const IconData wi_day_rain_wind = FlutterIconData.weatherIcons(61447);
+  static const IconData wi_day_showers = FlutterIconData.weatherIcons(61449);
+  static const IconData wi_day_sleet = FlutterIconData.weatherIcons(61618);
   static const IconData wi_day_sleet_storm =
-      const FlutterIconData.weatherIcons(61544);
-  static const IconData wi_day_snow = const FlutterIconData.weatherIcons(61450);
+      FlutterIconData.weatherIcons(61544);
+  static const IconData wi_day_snow = FlutterIconData.weatherIcons(61450);
   static const IconData wi_day_snow_thunderstorm =
-      const FlutterIconData.weatherIcons(61547);
-  static const IconData wi_day_snow_wind =
-      const FlutterIconData.weatherIcons(61541);
-  static const IconData wi_day_sprinkle =
-      const FlutterIconData.weatherIcons(61451);
+      FlutterIconData.weatherIcons(61547);
+  static const IconData wi_day_snow_wind = FlutterIconData.weatherIcons(61541);
+  static const IconData wi_day_sprinkle = FlutterIconData.weatherIcons(61451);
   static const IconData wi_day_storm_showers =
-      const FlutterIconData.weatherIcons(61454);
+      FlutterIconData.weatherIcons(61454);
   static const IconData wi_day_sunny_overcast =
-      const FlutterIconData.weatherIcons(61452);
+      FlutterIconData.weatherIcons(61452);
   static const IconData wi_day_thunderstorm =
-      const FlutterIconData.weatherIcons(61456);
-  static const IconData wi_day_windy =
-      const FlutterIconData.weatherIcons(61573);
-  static const IconData wi_solar_eclipse =
-      const FlutterIconData.weatherIcons(61550);
-  static const IconData wi_hot = const FlutterIconData.weatherIcons(61554);
+      FlutterIconData.weatherIcons(61456);
+  static const IconData wi_day_windy = FlutterIconData.weatherIcons(61573);
+  static const IconData wi_solar_eclipse = FlutterIconData.weatherIcons(61550);
+  static const IconData wi_hot = FlutterIconData.weatherIcons(61554);
   static const IconData wi_day_cloudy_high =
-      const FlutterIconData.weatherIcons(61565);
-  static const IconData wi_day_light_wind =
-      const FlutterIconData.weatherIcons(61636);
-  static const IconData wi_night_clear =
-      const FlutterIconData.weatherIcons(61486);
+      FlutterIconData.weatherIcons(61565);
+  static const IconData wi_day_light_wind = FlutterIconData.weatherIcons(61636);
+  static const IconData wi_night_clear = FlutterIconData.weatherIcons(61486);
   static const IconData wi_night_alt_cloudy =
-      const FlutterIconData.weatherIcons(61574);
+      FlutterIconData.weatherIcons(61574);
   static const IconData wi_night_alt_cloudy_gusts =
-      const FlutterIconData.weatherIcons(61474);
+      FlutterIconData.weatherIcons(61474);
   static const IconData wi_night_alt_cloudy_windy =
-      const FlutterIconData.weatherIcons(61475);
-  static const IconData wi_night_alt_hail =
-      const FlutterIconData.weatherIcons(61476);
+      FlutterIconData.weatherIcons(61475);
+  static const IconData wi_night_alt_hail = FlutterIconData.weatherIcons(61476);
   static const IconData wi_night_alt_lightning =
-      const FlutterIconData.weatherIcons(61477);
-  static const IconData wi_night_alt_rain =
-      const FlutterIconData.weatherIcons(61480);
+      FlutterIconData.weatherIcons(61477);
+  static const IconData wi_night_alt_rain = FlutterIconData.weatherIcons(61480);
   static const IconData wi_night_alt_rain_mix =
-      const FlutterIconData.weatherIcons(61478);
+      FlutterIconData.weatherIcons(61478);
   static const IconData wi_night_alt_rain_wind =
-      const FlutterIconData.weatherIcons(61479);
+      FlutterIconData.weatherIcons(61479);
   static const IconData wi_night_alt_showers =
-      const FlutterIconData.weatherIcons(61481);
+      FlutterIconData.weatherIcons(61481);
   static const IconData wi_night_alt_sleet =
-      const FlutterIconData.weatherIcons(61620);
+      FlutterIconData.weatherIcons(61620);
   static const IconData wi_night_alt_sleet_storm =
-      const FlutterIconData.weatherIcons(61546);
-  static const IconData wi_night_alt_snow =
-      const FlutterIconData.weatherIcons(61482);
+      FlutterIconData.weatherIcons(61546);
+  static const IconData wi_night_alt_snow = FlutterIconData.weatherIcons(61482);
   static const IconData wi_night_alt_snow_thunderstorm =
-      const FlutterIconData.weatherIcons(61549);
+      FlutterIconData.weatherIcons(61549);
   static const IconData wi_night_alt_snow_wind =
-      const FlutterIconData.weatherIcons(61543);
+      FlutterIconData.weatherIcons(61543);
   static const IconData wi_night_alt_sprinkle =
-      const FlutterIconData.weatherIcons(61483);
+      FlutterIconData.weatherIcons(61483);
   static const IconData wi_night_alt_storm_showers =
-      const FlutterIconData.weatherIcons(61484);
+      FlutterIconData.weatherIcons(61484);
   static const IconData wi_night_alt_thunderstorm =
-      const FlutterIconData.weatherIcons(61485);
-  static const IconData wi_night_cloudy =
-      const FlutterIconData.weatherIcons(61489);
+      FlutterIconData.weatherIcons(61485);
+  static const IconData wi_night_cloudy = FlutterIconData.weatherIcons(61489);
   static const IconData wi_night_cloudy_gusts =
-      const FlutterIconData.weatherIcons(61487);
+      FlutterIconData.weatherIcons(61487);
   static const IconData wi_night_cloudy_windy =
-      const FlutterIconData.weatherIcons(61488);
-  static const IconData wi_night_fog =
-      const FlutterIconData.weatherIcons(61514);
-  static const IconData wi_night_hail =
-      const FlutterIconData.weatherIcons(61490);
+      FlutterIconData.weatherIcons(61488);
+  static const IconData wi_night_fog = FlutterIconData.weatherIcons(61514);
+  static const IconData wi_night_hail = FlutterIconData.weatherIcons(61490);
   static const IconData wi_night_lightning =
-      const FlutterIconData.weatherIcons(61491);
+      FlutterIconData.weatherIcons(61491);
   static const IconData wi_night_partly_cloudy =
-      const FlutterIconData.weatherIcons(61571);
-  static const IconData wi_night_rain =
-      const FlutterIconData.weatherIcons(61494);
-  static const IconData wi_night_rain_mix =
-      const FlutterIconData.weatherIcons(61492);
+      FlutterIconData.weatherIcons(61571);
+  static const IconData wi_night_rain = FlutterIconData.weatherIcons(61494);
+  static const IconData wi_night_rain_mix = FlutterIconData.weatherIcons(61492);
   static const IconData wi_night_rain_wind =
-      const FlutterIconData.weatherIcons(61493);
-  static const IconData wi_night_showers =
-      const FlutterIconData.weatherIcons(61495);
-  static const IconData wi_night_sleet =
-      const FlutterIconData.weatherIcons(61619);
+      FlutterIconData.weatherIcons(61493);
+  static const IconData wi_night_showers = FlutterIconData.weatherIcons(61495);
+  static const IconData wi_night_sleet = FlutterIconData.weatherIcons(61619);
   static const IconData wi_night_sleet_storm =
-      const FlutterIconData.weatherIcons(61545);
-  static const IconData wi_night_snow =
-      const FlutterIconData.weatherIcons(61496);
+      FlutterIconData.weatherIcons(61545);
+  static const IconData wi_night_snow = FlutterIconData.weatherIcons(61496);
   static const IconData wi_night_snow_thunderstorm =
-      const FlutterIconData.weatherIcons(61548);
+      FlutterIconData.weatherIcons(61548);
   static const IconData wi_night_snow_wind =
-      const FlutterIconData.weatherIcons(61542);
-  static const IconData wi_night_sprinkle =
-      const FlutterIconData.weatherIcons(61497);
+      FlutterIconData.weatherIcons(61542);
+  static const IconData wi_night_sprinkle = FlutterIconData.weatherIcons(61497);
   static const IconData wi_night_storm_showers =
-      const FlutterIconData.weatherIcons(61498);
+      FlutterIconData.weatherIcons(61498);
   static const IconData wi_night_thunderstorm =
-      const FlutterIconData.weatherIcons(61499);
-  static const IconData wi_lunar_eclipse =
-      const FlutterIconData.weatherIcons(61552);
-  static const IconData wi_stars = const FlutterIconData.weatherIcons(61559);
-  static const IconData wi_storm_showers =
-      const FlutterIconData.weatherIcons(61469);
-  static const IconData wi_thunderstorm =
-      const FlutterIconData.weatherIcons(61470);
+      FlutterIconData.weatherIcons(61499);
+  static const IconData wi_lunar_eclipse = FlutterIconData.weatherIcons(61552);
+  static const IconData wi_stars = FlutterIconData.weatherIcons(61559);
+  static const IconData wi_storm_showers = FlutterIconData.weatherIcons(61469);
+  static const IconData wi_thunderstorm = FlutterIconData.weatherIcons(61470);
   static const IconData wi_night_alt_cloudy_high =
-      const FlutterIconData.weatherIcons(61566);
+      FlutterIconData.weatherIcons(61566);
   static const IconData wi_night_cloudy_high =
-      const FlutterIconData.weatherIcons(61568);
+      FlutterIconData.weatherIcons(61568);
   static const IconData wi_night_alt_partly_cloudy =
-      const FlutterIconData.weatherIcons(61569);
-  static const IconData wi_cloud = const FlutterIconData.weatherIcons(61505);
-  static const IconData wi_cloudy = const FlutterIconData.weatherIcons(61459);
-  static const IconData wi_cloudy_gusts =
-      const FlutterIconData.weatherIcons(61457);
-  static const IconData wi_cloudy_windy =
-      const FlutterIconData.weatherIcons(61458);
-  static const IconData wi_fog = const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_hail = const FlutterIconData.weatherIcons(61461);
-  static const IconData wi_rain = const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_rain_mix = const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_rain_wind =
-      const FlutterIconData.weatherIcons(61464);
-  static const IconData wi_showers = const FlutterIconData.weatherIcons(61466);
-  static const IconData wi_sleet = const FlutterIconData.weatherIcons(61621);
-  static const IconData wi_snow = const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_sprinkle = const FlutterIconData.weatherIcons(61468);
-  static const IconData wi_snow_wind =
-      const FlutterIconData.weatherIcons(61540);
-  static const IconData wi_smog = const FlutterIconData.weatherIcons(61556);
-  static const IconData wi_smoke = const FlutterIconData.weatherIcons(61538);
-  static const IconData wi_lightning =
-      const FlutterIconData.weatherIcons(61462);
-  static const IconData wi_raindrops =
-      const FlutterIconData.weatherIcons(61518);
-  static const IconData wi_raindrop = const FlutterIconData.weatherIcons(61560);
-  static const IconData wi_dust = const FlutterIconData.weatherIcons(61539);
-  static const IconData wi_snowflake_cold =
-      const FlutterIconData.weatherIcons(61558);
-  static const IconData wi_windy = const FlutterIconData.weatherIcons(61473);
-  static const IconData wi_strong_wind =
-      const FlutterIconData.weatherIcons(61520);
-  static const IconData wi_sandstorm =
-      const FlutterIconData.weatherIcons(61570);
-  static const IconData wi_earthquake =
-      const FlutterIconData.weatherIcons(61638);
-  static const IconData wi_fire = const FlutterIconData.weatherIcons(61639);
-  static const IconData wi_flood = const FlutterIconData.weatherIcons(61564);
-  static const IconData wi_meteor = const FlutterIconData.weatherIcons(61553);
-  static const IconData wi_tsunami = const FlutterIconData.weatherIcons(61637);
-  static const IconData wi_volcano = const FlutterIconData.weatherIcons(61640);
-  static const IconData wi_hurricane =
-      const FlutterIconData.weatherIcons(61555);
-  static const IconData wi_tornado = const FlutterIconData.weatherIcons(61526);
+      FlutterIconData.weatherIcons(61569);
+  static const IconData wi_cloud = FlutterIconData.weatherIcons(61505);
+  static const IconData wi_cloudy = FlutterIconData.weatherIcons(61459);
+  static const IconData wi_cloudy_gusts = FlutterIconData.weatherIcons(61457);
+  static const IconData wi_cloudy_windy = FlutterIconData.weatherIcons(61458);
+  static const IconData wi_fog = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_hail = FlutterIconData.weatherIcons(61461);
+  static const IconData wi_rain = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_rain_mix = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_rain_wind = FlutterIconData.weatherIcons(61464);
+  static const IconData wi_showers = FlutterIconData.weatherIcons(61466);
+  static const IconData wi_sleet = FlutterIconData.weatherIcons(61621);
+  static const IconData wi_snow = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_sprinkle = FlutterIconData.weatherIcons(61468);
+  static const IconData wi_snow_wind = FlutterIconData.weatherIcons(61540);
+  static const IconData wi_smog = FlutterIconData.weatherIcons(61556);
+  static const IconData wi_smoke = FlutterIconData.weatherIcons(61538);
+  static const IconData wi_lightning = FlutterIconData.weatherIcons(61462);
+  static const IconData wi_raindrops = FlutterIconData.weatherIcons(61518);
+  static const IconData wi_raindrop = FlutterIconData.weatherIcons(61560);
+  static const IconData wi_dust = FlutterIconData.weatherIcons(61539);
+  static const IconData wi_snowflake_cold = FlutterIconData.weatherIcons(61558);
+  static const IconData wi_windy = FlutterIconData.weatherIcons(61473);
+  static const IconData wi_strong_wind = FlutterIconData.weatherIcons(61520);
+  static const IconData wi_sandstorm = FlutterIconData.weatherIcons(61570);
+  static const IconData wi_earthquake = FlutterIconData.weatherIcons(61638);
+  static const IconData wi_fire = FlutterIconData.weatherIcons(61639);
+  static const IconData wi_flood = FlutterIconData.weatherIcons(61564);
+  static const IconData wi_meteor = FlutterIconData.weatherIcons(61553);
+  static const IconData wi_tsunami = FlutterIconData.weatherIcons(61637);
+  static const IconData wi_volcano = FlutterIconData.weatherIcons(61640);
+  static const IconData wi_hurricane = FlutterIconData.weatherIcons(61555);
+  static const IconData wi_tornado = FlutterIconData.weatherIcons(61526);
   static const IconData wi_small_craft_advisory =
-      const FlutterIconData.weatherIcons(61644);
-  static const IconData wi_gale_warning =
-      const FlutterIconData.weatherIcons(61645);
-  static const IconData wi_storm_warning =
-      const FlutterIconData.weatherIcons(61646);
+      FlutterIconData.weatherIcons(61644);
+  static const IconData wi_gale_warning = FlutterIconData.weatherIcons(61645);
+  static const IconData wi_storm_warning = FlutterIconData.weatherIcons(61646);
   static const IconData wi_hurricane_warning =
-      const FlutterIconData.weatherIcons(61647);
-  static const IconData wi_wind_direction =
-      const FlutterIconData.weatherIcons(61617);
-  static const IconData wi_alien = const FlutterIconData.weatherIcons(61557);
-  static const IconData wi_celsius = const FlutterIconData.weatherIcons(61500);
-  static const IconData wi_fahrenheit =
-      const FlutterIconData.weatherIcons(61509);
-  static const IconData wi_degrees = const FlutterIconData.weatherIcons(61506);
-  static const IconData wi_thermometer =
-      const FlutterIconData.weatherIcons(61525);
+      FlutterIconData.weatherIcons(61647);
+  static const IconData wi_wind_direction = FlutterIconData.weatherIcons(61617);
+  static const IconData wi_alien = FlutterIconData.weatherIcons(61557);
+  static const IconData wi_celsius = FlutterIconData.weatherIcons(61500);
+  static const IconData wi_fahrenheit = FlutterIconData.weatherIcons(61509);
+  static const IconData wi_degrees = FlutterIconData.weatherIcons(61506);
+  static const IconData wi_thermometer = FlutterIconData.weatherIcons(61525);
   static const IconData wi_thermometer_exterior =
-      const FlutterIconData.weatherIcons(61523);
+      FlutterIconData.weatherIcons(61523);
   static const IconData wi_thermometer_internal =
-      const FlutterIconData.weatherIcons(61524);
-  static const IconData wi_cloud_down =
-      const FlutterIconData.weatherIcons(61501);
-  static const IconData wi_cloud_up = const FlutterIconData.weatherIcons(61504);
-  static const IconData wi_cloud_refresh =
-      const FlutterIconData.weatherIcons(61502);
-  static const IconData wi_horizon = const FlutterIconData.weatherIcons(61511);
-  static const IconData wi_horizon_alt =
-      const FlutterIconData.weatherIcons(61510);
-  static const IconData wi_sunrise = const FlutterIconData.weatherIcons(61521);
-  static const IconData wi_sunset = const FlutterIconData.weatherIcons(61522);
-  static const IconData wi_moonrise = const FlutterIconData.weatherIcons(61641);
-  static const IconData wi_moonset = const FlutterIconData.weatherIcons(61642);
-  static const IconData wi_refresh = const FlutterIconData.weatherIcons(61516);
-  static const IconData wi_refresh_alt =
-      const FlutterIconData.weatherIcons(61515);
-  static const IconData wi_umbrella = const FlutterIconData.weatherIcons(61572);
-  static const IconData wi_barometer =
-      const FlutterIconData.weatherIcons(61561);
-  static const IconData wi_humidity = const FlutterIconData.weatherIcons(61562);
-  static const IconData wi_na = const FlutterIconData.weatherIcons(61563);
-  static const IconData wi_train = const FlutterIconData.weatherIcons(61643);
-  static const IconData wi_moon_new = const FlutterIconData.weatherIcons(61589);
+      FlutterIconData.weatherIcons(61524);
+  static const IconData wi_cloud_down = FlutterIconData.weatherIcons(61501);
+  static const IconData wi_cloud_up = FlutterIconData.weatherIcons(61504);
+  static const IconData wi_cloud_refresh = FlutterIconData.weatherIcons(61502);
+  static const IconData wi_horizon = FlutterIconData.weatherIcons(61511);
+  static const IconData wi_horizon_alt = FlutterIconData.weatherIcons(61510);
+  static const IconData wi_sunrise = FlutterIconData.weatherIcons(61521);
+  static const IconData wi_sunset = FlutterIconData.weatherIcons(61522);
+  static const IconData wi_moonrise = FlutterIconData.weatherIcons(61641);
+  static const IconData wi_moonset = FlutterIconData.weatherIcons(61642);
+  static const IconData wi_refresh = FlutterIconData.weatherIcons(61516);
+  static const IconData wi_refresh_alt = FlutterIconData.weatherIcons(61515);
+  static const IconData wi_umbrella = FlutterIconData.weatherIcons(61572);
+  static const IconData wi_barometer = FlutterIconData.weatherIcons(61561);
+  static const IconData wi_humidity = FlutterIconData.weatherIcons(61562);
+  static const IconData wi_na = FlutterIconData.weatherIcons(61563);
+  static const IconData wi_train = FlutterIconData.weatherIcons(61643);
+  static const IconData wi_moon_new = FlutterIconData.weatherIcons(61589);
   static const IconData wi_moon_waxing_crescent_1 =
-      const FlutterIconData.weatherIcons(61590);
+      FlutterIconData.weatherIcons(61590);
   static const IconData wi_moon_waxing_crescent_2 =
-      const FlutterIconData.weatherIcons(61591);
+      FlutterIconData.weatherIcons(61591);
   static const IconData wi_moon_waxing_crescent_3 =
-      const FlutterIconData.weatherIcons(61592);
+      FlutterIconData.weatherIcons(61592);
   static const IconData wi_moon_waxing_crescent_4 =
-      const FlutterIconData.weatherIcons(61593);
+      FlutterIconData.weatherIcons(61593);
   static const IconData wi_moon_waxing_crescent_5 =
-      const FlutterIconData.weatherIcons(61594);
+      FlutterIconData.weatherIcons(61594);
   static const IconData wi_moon_waxing_crescent_6 =
-      const FlutterIconData.weatherIcons(61595);
+      FlutterIconData.weatherIcons(61595);
   static const IconData wi_moon_first_quarter =
-      const FlutterIconData.weatherIcons(61596);
+      FlutterIconData.weatherIcons(61596);
   static const IconData wi_moon_waxing_gibbous_1 =
-      const FlutterIconData.weatherIcons(61597);
+      FlutterIconData.weatherIcons(61597);
   static const IconData wi_moon_waxing_gibbous_2 =
-      const FlutterIconData.weatherIcons(61598);
+      FlutterIconData.weatherIcons(61598);
   static const IconData wi_moon_waxing_gibbous_3 =
-      const FlutterIconData.weatherIcons(61599);
+      FlutterIconData.weatherIcons(61599);
   static const IconData wi_moon_waxing_gibbous_4 =
-      const FlutterIconData.weatherIcons(61600);
+      FlutterIconData.weatherIcons(61600);
   static const IconData wi_moon_waxing_gibbous_5 =
-      const FlutterIconData.weatherIcons(61601);
+      FlutterIconData.weatherIcons(61601);
   static const IconData wi_moon_waxing_gibbous_6 =
-      const FlutterIconData.weatherIcons(61602);
-  static const IconData wi_moon_full =
-      const FlutterIconData.weatherIcons(61603);
+      FlutterIconData.weatherIcons(61602);
+  static const IconData wi_moon_full = FlutterIconData.weatherIcons(61603);
   static const IconData wi_moon_waning_gibbous_1 =
-      const FlutterIconData.weatherIcons(61604);
+      FlutterIconData.weatherIcons(61604);
   static const IconData wi_moon_waning_gibbous_2 =
-      const FlutterIconData.weatherIcons(61605);
+      FlutterIconData.weatherIcons(61605);
   static const IconData wi_moon_waning_gibbous_3 =
-      const FlutterIconData.weatherIcons(61606);
+      FlutterIconData.weatherIcons(61606);
   static const IconData wi_moon_waning_gibbous_4 =
-      const FlutterIconData.weatherIcons(61607);
+      FlutterIconData.weatherIcons(61607);
   static const IconData wi_moon_waning_gibbous_5 =
-      const FlutterIconData.weatherIcons(61608);
+      FlutterIconData.weatherIcons(61608);
   static const IconData wi_moon_waning_gibbous_6 =
-      const FlutterIconData.weatherIcons(61609);
+      FlutterIconData.weatherIcons(61609);
   static const IconData wi_moon_third_quarter =
-      const FlutterIconData.weatherIcons(61610);
+      FlutterIconData.weatherIcons(61610);
   static const IconData wi_moon_waning_crescent_1 =
-      const FlutterIconData.weatherIcons(61611);
+      FlutterIconData.weatherIcons(61611);
   static const IconData wi_moon_waning_crescent_2 =
-      const FlutterIconData.weatherIcons(61612);
+      FlutterIconData.weatherIcons(61612);
   static const IconData wi_moon_waning_crescent_3 =
-      const FlutterIconData.weatherIcons(61613);
+      FlutterIconData.weatherIcons(61613);
   static const IconData wi_moon_waning_crescent_4 =
-      const FlutterIconData.weatherIcons(61614);
+      FlutterIconData.weatherIcons(61614);
   static const IconData wi_moon_waning_crescent_5 =
-      const FlutterIconData.weatherIcons(61615);
+      FlutterIconData.weatherIcons(61615);
   static const IconData wi_moon_waning_crescent_6 =
-      const FlutterIconData.weatherIcons(61616);
-  static const IconData wi_moon_alt_new =
-      const FlutterIconData.weatherIcons(61675);
+      FlutterIconData.weatherIcons(61616);
+  static const IconData wi_moon_alt_new = FlutterIconData.weatherIcons(61675);
   static const IconData wi_moon_alt_waxing_crescent_1 =
-      const FlutterIconData.weatherIcons(61648);
+      FlutterIconData.weatherIcons(61648);
   static const IconData wi_moon_alt_waxing_crescent_2 =
-      const FlutterIconData.weatherIcons(61649);
+      FlutterIconData.weatherIcons(61649);
   static const IconData wi_moon_alt_waxing_crescent_3 =
-      const FlutterIconData.weatherIcons(61650);
+      FlutterIconData.weatherIcons(61650);
   static const IconData wi_moon_alt_waxing_crescent_4 =
-      const FlutterIconData.weatherIcons(61651);
+      FlutterIconData.weatherIcons(61651);
   static const IconData wi_moon_alt_waxing_crescent_5 =
-      const FlutterIconData.weatherIcons(61652);
+      FlutterIconData.weatherIcons(61652);
   static const IconData wi_moon_alt_waxing_crescent_6 =
-      const FlutterIconData.weatherIcons(61653);
+      FlutterIconData.weatherIcons(61653);
   static const IconData wi_moon_alt_first_quarter =
-      const FlutterIconData.weatherIcons(61654);
+      FlutterIconData.weatherIcons(61654);
   static const IconData wi_moon_alt_waxing_gibbous_1 =
-      const FlutterIconData.weatherIcons(61655);
+      FlutterIconData.weatherIcons(61655);
   static const IconData wi_moon_alt_waxing_gibbous_2 =
-      const FlutterIconData.weatherIcons(61656);
+      FlutterIconData.weatherIcons(61656);
   static const IconData wi_moon_alt_waxing_gibbous_3 =
-      const FlutterIconData.weatherIcons(61657);
+      FlutterIconData.weatherIcons(61657);
   static const IconData wi_moon_alt_waxing_gibbous_4 =
-      const FlutterIconData.weatherIcons(61658);
+      FlutterIconData.weatherIcons(61658);
   static const IconData wi_moon_alt_waxing_gibbous_5 =
-      const FlutterIconData.weatherIcons(61659);
+      FlutterIconData.weatherIcons(61659);
   static const IconData wi_moon_alt_waxing_gibbous_6 =
-      const FlutterIconData.weatherIcons(61660);
-  static const IconData wi_moon_alt_full =
-      const FlutterIconData.weatherIcons(61661);
+      FlutterIconData.weatherIcons(61660);
+  static const IconData wi_moon_alt_full = FlutterIconData.weatherIcons(61661);
   static const IconData wi_moon_alt_waning_gibbous_1 =
-      const FlutterIconData.weatherIcons(61662);
+      FlutterIconData.weatherIcons(61662);
   static const IconData wi_moon_alt_waning_gibbous_2 =
-      const FlutterIconData.weatherIcons(61663);
+      FlutterIconData.weatherIcons(61663);
   static const IconData wi_moon_alt_waning_gibbous_3 =
-      const FlutterIconData.weatherIcons(61664);
+      FlutterIconData.weatherIcons(61664);
   static const IconData wi_moon_alt_waning_gibbous_4 =
-      const FlutterIconData.weatherIcons(61665);
+      FlutterIconData.weatherIcons(61665);
   static const IconData wi_moon_alt_waning_gibbous_5 =
-      const FlutterIconData.weatherIcons(61666);
+      FlutterIconData.weatherIcons(61666);
   static const IconData wi_moon_alt_waning_gibbous_6 =
-      const FlutterIconData.weatherIcons(61667);
+      FlutterIconData.weatherIcons(61667);
   static const IconData wi_moon_alt_third_quarter =
-      const FlutterIconData.weatherIcons(61668);
+      FlutterIconData.weatherIcons(61668);
   static const IconData wi_moon_alt_waning_crescent_1 =
-      const FlutterIconData.weatherIcons(61669);
+      FlutterIconData.weatherIcons(61669);
   static const IconData wi_moon_alt_waning_crescent_2 =
-      const FlutterIconData.weatherIcons(61670);
+      FlutterIconData.weatherIcons(61670);
   static const IconData wi_moon_alt_waning_crescent_3 =
-      const FlutterIconData.weatherIcons(61671);
+      FlutterIconData.weatherIcons(61671);
   static const IconData wi_moon_alt_waning_crescent_4 =
-      const FlutterIconData.weatherIcons(61672);
+      FlutterIconData.weatherIcons(61672);
   static const IconData wi_moon_alt_waning_crescent_5 =
-      const FlutterIconData.weatherIcons(61673);
+      FlutterIconData.weatherIcons(61673);
   static const IconData wi_moon_alt_waning_crescent_6 =
-      const FlutterIconData.weatherIcons(61674);
-  static const IconData wi_moon_0 = const FlutterIconData.weatherIcons(61589);
-  static const IconData wi_moon_1 = const FlutterIconData.weatherIcons(61590);
-  static const IconData wi_moon_2 = const FlutterIconData.weatherIcons(61591);
-  static const IconData wi_moon_3 = const FlutterIconData.weatherIcons(61592);
-  static const IconData wi_moon_4 = const FlutterIconData.weatherIcons(61593);
-  static const IconData wi_moon_5 = const FlutterIconData.weatherIcons(61594);
-  static const IconData wi_moon_6 = const FlutterIconData.weatherIcons(61595);
-  static const IconData wi_moon_7 = const FlutterIconData.weatherIcons(61596);
-  static const IconData wi_moon_8 = const FlutterIconData.weatherIcons(61597);
-  static const IconData wi_moon_9 = const FlutterIconData.weatherIcons(61598);
-  static const IconData wi_moon_10 = const FlutterIconData.weatherIcons(61599);
-  static const IconData wi_moon_11 = const FlutterIconData.weatherIcons(61600);
-  static const IconData wi_moon_12 = const FlutterIconData.weatherIcons(61601);
-  static const IconData wi_moon_13 = const FlutterIconData.weatherIcons(61602);
-  static const IconData wi_moon_14 = const FlutterIconData.weatherIcons(61603);
-  static const IconData wi_moon_15 = const FlutterIconData.weatherIcons(61604);
-  static const IconData wi_moon_16 = const FlutterIconData.weatherIcons(61605);
-  static const IconData wi_moon_17 = const FlutterIconData.weatherIcons(61606);
-  static const IconData wi_moon_18 = const FlutterIconData.weatherIcons(61607);
-  static const IconData wi_moon_19 = const FlutterIconData.weatherIcons(61608);
-  static const IconData wi_moon_20 = const FlutterIconData.weatherIcons(61609);
-  static const IconData wi_moon_21 = const FlutterIconData.weatherIcons(61610);
-  static const IconData wi_moon_22 = const FlutterIconData.weatherIcons(61611);
-  static const IconData wi_moon_23 = const FlutterIconData.weatherIcons(61612);
-  static const IconData wi_moon_24 = const FlutterIconData.weatherIcons(61613);
-  static const IconData wi_moon_25 = const FlutterIconData.weatherIcons(61614);
-  static const IconData wi_moon_26 = const FlutterIconData.weatherIcons(61615);
-  static const IconData wi_moon_27 = const FlutterIconData.weatherIcons(61616);
-  static const IconData wi_time_1 = const FlutterIconData.weatherIcons(61578);
-  static const IconData wi_time_2 = const FlutterIconData.weatherIcons(61579);
-  static const IconData wi_time_3 = const FlutterIconData.weatherIcons(61580);
-  static const IconData wi_time_4 = const FlutterIconData.weatherIcons(61581);
-  static const IconData wi_time_5 = const FlutterIconData.weatherIcons(61582);
-  static const IconData wi_time_6 = const FlutterIconData.weatherIcons(61583);
-  static const IconData wi_time_7 = const FlutterIconData.weatherIcons(61584);
-  static const IconData wi_time_8 = const FlutterIconData.weatherIcons(61585);
-  static const IconData wi_time_9 = const FlutterIconData.weatherIcons(61586);
-  static const IconData wi_time_10 = const FlutterIconData.weatherIcons(61587);
-  static const IconData wi_time_11 = const FlutterIconData.weatherIcons(61588);
-  static const IconData wi_time_12 = const FlutterIconData.weatherIcons(61577);
-  static const IconData wi_direction_up =
-      const FlutterIconData.weatherIcons(61528);
+      FlutterIconData.weatherIcons(61674);
+  static const IconData wi_moon_0 = FlutterIconData.weatherIcons(61589);
+  static const IconData wi_moon_1 = FlutterIconData.weatherIcons(61590);
+  static const IconData wi_moon_2 = FlutterIconData.weatherIcons(61591);
+  static const IconData wi_moon_3 = FlutterIconData.weatherIcons(61592);
+  static const IconData wi_moon_4 = FlutterIconData.weatherIcons(61593);
+  static const IconData wi_moon_5 = FlutterIconData.weatherIcons(61594);
+  static const IconData wi_moon_6 = FlutterIconData.weatherIcons(61595);
+  static const IconData wi_moon_7 = FlutterIconData.weatherIcons(61596);
+  static const IconData wi_moon_8 = FlutterIconData.weatherIcons(61597);
+  static const IconData wi_moon_9 = FlutterIconData.weatherIcons(61598);
+  static const IconData wi_moon_10 = FlutterIconData.weatherIcons(61599);
+  static const IconData wi_moon_11 = FlutterIconData.weatherIcons(61600);
+  static const IconData wi_moon_12 = FlutterIconData.weatherIcons(61601);
+  static const IconData wi_moon_13 = FlutterIconData.weatherIcons(61602);
+  static const IconData wi_moon_14 = FlutterIconData.weatherIcons(61603);
+  static const IconData wi_moon_15 = FlutterIconData.weatherIcons(61604);
+  static const IconData wi_moon_16 = FlutterIconData.weatherIcons(61605);
+  static const IconData wi_moon_17 = FlutterIconData.weatherIcons(61606);
+  static const IconData wi_moon_18 = FlutterIconData.weatherIcons(61607);
+  static const IconData wi_moon_19 = FlutterIconData.weatherIcons(61608);
+  static const IconData wi_moon_20 = FlutterIconData.weatherIcons(61609);
+  static const IconData wi_moon_21 = FlutterIconData.weatherIcons(61610);
+  static const IconData wi_moon_22 = FlutterIconData.weatherIcons(61611);
+  static const IconData wi_moon_23 = FlutterIconData.weatherIcons(61612);
+  static const IconData wi_moon_24 = FlutterIconData.weatherIcons(61613);
+  static const IconData wi_moon_25 = FlutterIconData.weatherIcons(61614);
+  static const IconData wi_moon_26 = FlutterIconData.weatherIcons(61615);
+  static const IconData wi_moon_27 = FlutterIconData.weatherIcons(61616);
+  static const IconData wi_time_1 = FlutterIconData.weatherIcons(61578);
+  static const IconData wi_time_2 = FlutterIconData.weatherIcons(61579);
+  static const IconData wi_time_3 = FlutterIconData.weatherIcons(61580);
+  static const IconData wi_time_4 = FlutterIconData.weatherIcons(61581);
+  static const IconData wi_time_5 = FlutterIconData.weatherIcons(61582);
+  static const IconData wi_time_6 = FlutterIconData.weatherIcons(61583);
+  static const IconData wi_time_7 = FlutterIconData.weatherIcons(61584);
+  static const IconData wi_time_8 = FlutterIconData.weatherIcons(61585);
+  static const IconData wi_time_9 = FlutterIconData.weatherIcons(61586);
+  static const IconData wi_time_10 = FlutterIconData.weatherIcons(61587);
+  static const IconData wi_time_11 = FlutterIconData.weatherIcons(61588);
+  static const IconData wi_time_12 = FlutterIconData.weatherIcons(61577);
+  static const IconData wi_direction_up = FlutterIconData.weatherIcons(61528);
   static const IconData wi_direction_up_right =
-      const FlutterIconData.weatherIcons(61527);
+      FlutterIconData.weatherIcons(61527);
   static const IconData wi_direction_right =
-      const FlutterIconData.weatherIcons(61517);
+      FlutterIconData.weatherIcons(61517);
   static const IconData wi_direction_down_right =
-      const FlutterIconData.weatherIcons(61576);
-  static const IconData wi_direction_down =
-      const FlutterIconData.weatherIcons(61508);
+      FlutterIconData.weatherIcons(61576);
+  static const IconData wi_direction_down = FlutterIconData.weatherIcons(61508);
   static const IconData wi_direction_down_left =
-      const FlutterIconData.weatherIcons(61507);
-  static const IconData wi_direction_left =
-      const FlutterIconData.weatherIcons(61512);
+      FlutterIconData.weatherIcons(61507);
+  static const IconData wi_direction_left = FlutterIconData.weatherIcons(61512);
   static const IconData wi_direction_up_left =
-      const FlutterIconData.weatherIcons(61575);
+      FlutterIconData.weatherIcons(61575);
   static const IconData wi_wind_beaufort_0 =
-      const FlutterIconData.weatherIcons(61623);
+      FlutterIconData.weatherIcons(61623);
   static const IconData wi_wind_beaufort_1 =
-      const FlutterIconData.weatherIcons(61624);
+      FlutterIconData.weatherIcons(61624);
   static const IconData wi_wind_beaufort_2 =
-      const FlutterIconData.weatherIcons(61625);
+      FlutterIconData.weatherIcons(61625);
   static const IconData wi_wind_beaufort_3 =
-      const FlutterIconData.weatherIcons(61626);
+      FlutterIconData.weatherIcons(61626);
   static const IconData wi_wind_beaufort_4 =
-      const FlutterIconData.weatherIcons(61627);
+      FlutterIconData.weatherIcons(61627);
   static const IconData wi_wind_beaufort_5 =
-      const FlutterIconData.weatherIcons(61628);
+      FlutterIconData.weatherIcons(61628);
   static const IconData wi_wind_beaufort_6 =
-      const FlutterIconData.weatherIcons(61629);
+      FlutterIconData.weatherIcons(61629);
   static const IconData wi_wind_beaufort_7 =
-      const FlutterIconData.weatherIcons(61630);
+      FlutterIconData.weatherIcons(61630);
   static const IconData wi_wind_beaufort_8 =
-      const FlutterIconData.weatherIcons(61631);
+      FlutterIconData.weatherIcons(61631);
   static const IconData wi_wind_beaufort_9 =
-      const FlutterIconData.weatherIcons(61632);
+      FlutterIconData.weatherIcons(61632);
   static const IconData wi_wind_beaufort_10 =
-      const FlutterIconData.weatherIcons(61633);
+      FlutterIconData.weatherIcons(61633);
   static const IconData wi_wind_beaufort_11 =
-      const FlutterIconData.weatherIcons(61634);
+      FlutterIconData.weatherIcons(61634);
   static const IconData wi_wind_beaufort_12 =
-      const FlutterIconData.weatherIcons(61635);
-  static const IconData wi_yahoo_0 = const FlutterIconData.weatherIcons(61526);
-  static const IconData wi_yahoo_1 = const FlutterIconData.weatherIcons(61454);
-  static const IconData wi_yahoo_2 = const FlutterIconData.weatherIcons(61555);
-  static const IconData wi_yahoo_3 = const FlutterIconData.weatherIcons(61470);
-  static const IconData wi_yahoo_4 = const FlutterIconData.weatherIcons(61470);
-  static const IconData wi_yahoo_5 = const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_yahoo_6 = const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_yahoo_7 = const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_yahoo_8 = const FlutterIconData.weatherIcons(61461);
-  static const IconData wi_yahoo_9 = const FlutterIconData.weatherIcons(61466);
-  static const IconData wi_yahoo_10 = const FlutterIconData.weatherIcons(61461);
-  static const IconData wi_yahoo_11 = const FlutterIconData.weatherIcons(61466);
-  static const IconData wi_yahoo_12 = const FlutterIconData.weatherIcons(61466);
-  static const IconData wi_yahoo_13 = const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_yahoo_14 = const FlutterIconData.weatherIcons(61450);
-  static const IconData wi_yahoo_15 = const FlutterIconData.weatherIcons(61540);
-  static const IconData wi_yahoo_16 = const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_yahoo_17 = const FlutterIconData.weatherIcons(61461);
-  static const IconData wi_yahoo_18 = const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_yahoo_19 = const FlutterIconData.weatherIcons(61539);
-  static const IconData wi_yahoo_20 = const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_yahoo_21 = const FlutterIconData.weatherIcons(61473);
-  static const IconData wi_yahoo_22 = const FlutterIconData.weatherIcons(61538);
-  static const IconData wi_yahoo_23 = const FlutterIconData.weatherIcons(61520);
-  static const IconData wi_yahoo_24 = const FlutterIconData.weatherIcons(61520);
-  static const IconData wi_yahoo_25 = const FlutterIconData.weatherIcons(61558);
-  static const IconData wi_yahoo_26 = const FlutterIconData.weatherIcons(61459);
-  static const IconData wi_yahoo_27 = const FlutterIconData.weatherIcons(61489);
-  static const IconData wi_yahoo_28 = const FlutterIconData.weatherIcons(61442);
-  static const IconData wi_yahoo_29 = const FlutterIconData.weatherIcons(61489);
-  static const IconData wi_yahoo_30 = const FlutterIconData.weatherIcons(61442);
-  static const IconData wi_yahoo_31 = const FlutterIconData.weatherIcons(61486);
-  static const IconData wi_yahoo_32 = const FlutterIconData.weatherIcons(61453);
-  static const IconData wi_yahoo_33 = const FlutterIconData.weatherIcons(61571);
-  static const IconData wi_yahoo_34 = const FlutterIconData.weatherIcons(61452);
-  static const IconData wi_yahoo_35 = const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_yahoo_36 = const FlutterIconData.weatherIcons(61554);
-  static const IconData wi_yahoo_37 = const FlutterIconData.weatherIcons(61454);
-  static const IconData wi_yahoo_38 = const FlutterIconData.weatherIcons(61454);
-  static const IconData wi_yahoo_39 = const FlutterIconData.weatherIcons(61454);
-  static const IconData wi_yahoo_40 = const FlutterIconData.weatherIcons(61466);
-  static const IconData wi_yahoo_41 = const FlutterIconData.weatherIcons(61540);
-  static const IconData wi_yahoo_42 = const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_yahoo_43 = const FlutterIconData.weatherIcons(61540);
-  static const IconData wi_yahoo_44 = const FlutterIconData.weatherIcons(61452);
-  static const IconData wi_yahoo_45 = const FlutterIconData.weatherIcons(61454);
-  static const IconData wi_yahoo_46 = const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_yahoo_47 = const FlutterIconData.weatherIcons(61454);
-  static const IconData wi_yahoo_3200 =
-      const FlutterIconData.weatherIcons(61559);
+      FlutterIconData.weatherIcons(61635);
+  static const IconData wi_yahoo_0 = FlutterIconData.weatherIcons(61526);
+  static const IconData wi_yahoo_1 = FlutterIconData.weatherIcons(61454);
+  static const IconData wi_yahoo_2 = FlutterIconData.weatherIcons(61555);
+  static const IconData wi_yahoo_3 = FlutterIconData.weatherIcons(61470);
+  static const IconData wi_yahoo_4 = FlutterIconData.weatherIcons(61470);
+  static const IconData wi_yahoo_5 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_yahoo_6 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_yahoo_7 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_yahoo_8 = FlutterIconData.weatherIcons(61461);
+  static const IconData wi_yahoo_9 = FlutterIconData.weatherIcons(61466);
+  static const IconData wi_yahoo_10 = FlutterIconData.weatherIcons(61461);
+  static const IconData wi_yahoo_11 = FlutterIconData.weatherIcons(61466);
+  static const IconData wi_yahoo_12 = FlutterIconData.weatherIcons(61466);
+  static const IconData wi_yahoo_13 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_yahoo_14 = FlutterIconData.weatherIcons(61450);
+  static const IconData wi_yahoo_15 = FlutterIconData.weatherIcons(61540);
+  static const IconData wi_yahoo_16 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_yahoo_17 = FlutterIconData.weatherIcons(61461);
+  static const IconData wi_yahoo_18 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_yahoo_19 = FlutterIconData.weatherIcons(61539);
+  static const IconData wi_yahoo_20 = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_yahoo_21 = FlutterIconData.weatherIcons(61473);
+  static const IconData wi_yahoo_22 = FlutterIconData.weatherIcons(61538);
+  static const IconData wi_yahoo_23 = FlutterIconData.weatherIcons(61520);
+  static const IconData wi_yahoo_24 = FlutterIconData.weatherIcons(61520);
+  static const IconData wi_yahoo_25 = FlutterIconData.weatherIcons(61558);
+  static const IconData wi_yahoo_26 = FlutterIconData.weatherIcons(61459);
+  static const IconData wi_yahoo_27 = FlutterIconData.weatherIcons(61489);
+  static const IconData wi_yahoo_28 = FlutterIconData.weatherIcons(61442);
+  static const IconData wi_yahoo_29 = FlutterIconData.weatherIcons(61489);
+  static const IconData wi_yahoo_30 = FlutterIconData.weatherIcons(61442);
+  static const IconData wi_yahoo_31 = FlutterIconData.weatherIcons(61486);
+  static const IconData wi_yahoo_32 = FlutterIconData.weatherIcons(61453);
+  static const IconData wi_yahoo_33 = FlutterIconData.weatherIcons(61571);
+  static const IconData wi_yahoo_34 = FlutterIconData.weatherIcons(61452);
+  static const IconData wi_yahoo_35 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_yahoo_36 = FlutterIconData.weatherIcons(61554);
+  static const IconData wi_yahoo_37 = FlutterIconData.weatherIcons(61454);
+  static const IconData wi_yahoo_38 = FlutterIconData.weatherIcons(61454);
+  static const IconData wi_yahoo_39 = FlutterIconData.weatherIcons(61454);
+  static const IconData wi_yahoo_40 = FlutterIconData.weatherIcons(61466);
+  static const IconData wi_yahoo_41 = FlutterIconData.weatherIcons(61540);
+  static const IconData wi_yahoo_42 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_yahoo_43 = FlutterIconData.weatherIcons(61540);
+  static const IconData wi_yahoo_44 = FlutterIconData.weatherIcons(61452);
+  static const IconData wi_yahoo_45 = FlutterIconData.weatherIcons(61454);
+  static const IconData wi_yahoo_46 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_yahoo_47 = FlutterIconData.weatherIcons(61454);
+  static const IconData wi_yahoo_3200 = FlutterIconData.weatherIcons(61559);
   static const IconData wi_forecast_io_clear_day =
-      const FlutterIconData.weatherIcons(61453);
+      FlutterIconData.weatherIcons(61453);
   static const IconData wi_forecast_io_clear_night =
-      const FlutterIconData.weatherIcons(61486);
+      FlutterIconData.weatherIcons(61486);
   static const IconData wi_forecast_io_rain =
-      const FlutterIconData.weatherIcons(61465);
+      FlutterIconData.weatherIcons(61465);
   static const IconData wi_forecast_io_snow =
-      const FlutterIconData.weatherIcons(61467);
+      FlutterIconData.weatherIcons(61467);
   static const IconData wi_forecast_io_sleet =
-      const FlutterIconData.weatherIcons(61621);
+      FlutterIconData.weatherIcons(61621);
   static const IconData wi_forecast_io_wind =
-      const FlutterIconData.weatherIcons(61520);
+      FlutterIconData.weatherIcons(61520);
   static const IconData wi_forecast_io_fog =
-      const FlutterIconData.weatherIcons(61460);
+      FlutterIconData.weatherIcons(61460);
   static const IconData wi_forecast_io_cloudy =
-      const FlutterIconData.weatherIcons(61459);
+      FlutterIconData.weatherIcons(61459);
   static const IconData wi_forecast_io_partly_cloudy_day =
-      const FlutterIconData.weatherIcons(61442);
+      FlutterIconData.weatherIcons(61442);
   static const IconData wi_forecast_io_partly_cloudy_night =
-      const FlutterIconData.weatherIcons(61489);
+      FlutterIconData.weatherIcons(61489);
   static const IconData wi_forecast_io_hail =
-      const FlutterIconData.weatherIcons(61461);
+      FlutterIconData.weatherIcons(61461);
   static const IconData wi_forecast_io_thunderstorm =
-      const FlutterIconData.weatherIcons(61470);
+      FlutterIconData.weatherIcons(61470);
   static const IconData wi_forecast_io_tornado =
-      const FlutterIconData.weatherIcons(61526);
-  static const IconData wi_wmo4680_0 =
-      const FlutterIconData.weatherIcons(61525);
-  static const IconData wi_wmo4680_00 =
-      const FlutterIconData.weatherIcons(61525);
-  static const IconData wi_wmo4680_1 =
-      const FlutterIconData.weatherIcons(61459);
-  static const IconData wi_wmo4680_01 =
-      const FlutterIconData.weatherIcons(61459);
-  static const IconData wi_wmo4680_2 =
-      const FlutterIconData.weatherIcons(61525);
-  static const IconData wi_wmo4680_02 =
-      const FlutterIconData.weatherIcons(61525);
-  static const IconData wi_wmo4680_3 =
-      const FlutterIconData.weatherIcons(61459);
-  static const IconData wi_wmo4680_03 =
-      const FlutterIconData.weatherIcons(61459);
-  static const IconData wi_wmo4680_4 =
-      const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_wmo4680_04 =
-      const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_wmo4680_5 =
-      const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_wmo4680_05 =
-      const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_wmo4680_10 =
-      const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_wmo4680_11 =
-      const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_wmo4680_12 =
-      const FlutterIconData.weatherIcons(61462);
-  static const IconData wi_wmo4680_18 =
-      const FlutterIconData.weatherIcons(61520);
-  static const IconData wi_wmo4680_20 =
-      const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_wmo4680_21 =
-      const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_wmo4680_22 =
-      const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_wmo4680_23 =
-      const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_wmo4680_24 =
-      const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_wmo4680_25 =
-      const FlutterIconData.weatherIcons(61461);
-  static const IconData wi_wmo4680_26 =
-      const FlutterIconData.weatherIcons(61470);
-  static const IconData wi_wmo4680_27 =
-      const FlutterIconData.weatherIcons(61539);
-  static const IconData wi_wmo4680_28 =
-      const FlutterIconData.weatherIcons(61539);
-  static const IconData wi_wmo4680_29 =
-      const FlutterIconData.weatherIcons(61539);
-  static const IconData wi_wmo4680_30 =
-      const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_wmo4680_31 =
-      const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_wmo4680_32 =
-      const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_wmo4680_33 =
-      const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_wmo4680_34 =
-      const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_wmo4680_35 =
-      const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_wmo4680_40 =
-      const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_wmo4680_41 =
-      const FlutterIconData.weatherIcons(61468);
-  static const IconData wi_wmo4680_42 =
-      const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_wmo4680_43 =
-      const FlutterIconData.weatherIcons(61468);
-  static const IconData wi_wmo4680_44 =
-      const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_wmo4680_45 =
-      const FlutterIconData.weatherIcons(61461);
-  static const IconData wi_wmo4680_46 =
-      const FlutterIconData.weatherIcons(61461);
-  static const IconData wi_wmo4680_47 =
-      const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_wmo4680_48 =
-      const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_wmo4680_50 =
-      const FlutterIconData.weatherIcons(61468);
-  static const IconData wi_wmo4680_51 =
-      const FlutterIconData.weatherIcons(61468);
-  static const IconData wi_wmo4680_52 =
-      const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_wmo4680_53 =
-      const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_wmo4680_54 =
-      const FlutterIconData.weatherIcons(61558);
-  static const IconData wi_wmo4680_55 =
-      const FlutterIconData.weatherIcons(61558);
-  static const IconData wi_wmo4680_56 =
-      const FlutterIconData.weatherIcons(61558);
-  static const IconData wi_wmo4680_57 =
-      const FlutterIconData.weatherIcons(61468);
-  static const IconData wi_wmo4680_58 =
-      const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_wmo4680_60 =
-      const FlutterIconData.weatherIcons(61468);
-  static const IconData wi_wmo4680_61 =
-      const FlutterIconData.weatherIcons(61468);
-  static const IconData wi_wmo4680_62 =
-      const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_wmo4680_63 =
-      const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_wmo4680_64 =
-      const FlutterIconData.weatherIcons(61461);
-  static const IconData wi_wmo4680_65 =
-      const FlutterIconData.weatherIcons(61461);
-  static const IconData wi_wmo4680_66 =
-      const FlutterIconData.weatherIcons(61461);
-  static const IconData wi_wmo4680_67 =
-      const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_wmo4680_68 =
-      const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_wmo4680_70 =
-      const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_wmo4680_71 =
-      const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_wmo4680_72 =
-      const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_wmo4680_73 =
-      const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_wmo4680_74 =
-      const FlutterIconData.weatherIcons(61558);
-  static const IconData wi_wmo4680_75 =
-      const FlutterIconData.weatherIcons(61558);
-  static const IconData wi_wmo4680_76 =
-      const FlutterIconData.weatherIcons(61558);
-  static const IconData wi_wmo4680_77 =
-      const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_wmo4680_78 =
-      const FlutterIconData.weatherIcons(61558);
-  static const IconData wi_wmo4680_80 =
-      const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_wmo4680_81 =
-      const FlutterIconData.weatherIcons(61468);
-  static const IconData wi_wmo4680_82 =
-      const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_wmo4680_83 =
-      const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_wmo4680_84 =
-      const FlutterIconData.weatherIcons(61469);
-  static const IconData wi_wmo4680_85 =
-      const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_wmo4680_86 =
-      const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_wmo4680_87 =
-      const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_wmo4680_89 =
-      const FlutterIconData.weatherIcons(61461);
-  static const IconData wi_wmo4680_90 =
-      const FlutterIconData.weatherIcons(61462);
-  static const IconData wi_wmo4680_91 =
-      const FlutterIconData.weatherIcons(61469);
-  static const IconData wi_wmo4680_92 =
-      const FlutterIconData.weatherIcons(61470);
-  static const IconData wi_wmo4680_93 =
-      const FlutterIconData.weatherIcons(61470);
-  static const IconData wi_wmo4680_94 =
-      const FlutterIconData.weatherIcons(61462);
-  static const IconData wi_wmo4680_95 =
-      const FlutterIconData.weatherIcons(61470);
-  static const IconData wi_wmo4680_96 =
-      const FlutterIconData.weatherIcons(61470);
-  static const IconData wi_wmo4680_99 =
-      const FlutterIconData.weatherIcons(61526);
-  static const IconData wi_owm_200 = const FlutterIconData.weatherIcons(61470);
-  static const IconData wi_owm_201 = const FlutterIconData.weatherIcons(61470);
-  static const IconData wi_owm_202 = const FlutterIconData.weatherIcons(61470);
-  static const IconData wi_owm_210 = const FlutterIconData.weatherIcons(61462);
-  static const IconData wi_owm_211 = const FlutterIconData.weatherIcons(61462);
-  static const IconData wi_owm_212 = const FlutterIconData.weatherIcons(61462);
-  static const IconData wi_owm_221 = const FlutterIconData.weatherIcons(61462);
-  static const IconData wi_owm_230 = const FlutterIconData.weatherIcons(61470);
-  static const IconData wi_owm_231 = const FlutterIconData.weatherIcons(61470);
-  static const IconData wi_owm_232 = const FlutterIconData.weatherIcons(61470);
-  static const IconData wi_owm_300 = const FlutterIconData.weatherIcons(61468);
-  static const IconData wi_owm_301 = const FlutterIconData.weatherIcons(61468);
-  static const IconData wi_owm_302 = const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_owm_310 = const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_owm_311 = const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_owm_312 = const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_owm_313 = const FlutterIconData.weatherIcons(61466);
-  static const IconData wi_owm_314 = const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_owm_321 = const FlutterIconData.weatherIcons(61468);
-  static const IconData wi_owm_500 = const FlutterIconData.weatherIcons(61468);
-  static const IconData wi_owm_501 = const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_owm_502 = const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_owm_503 = const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_owm_504 = const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_owm_511 = const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_owm_520 = const FlutterIconData.weatherIcons(61466);
-  static const IconData wi_owm_521 = const FlutterIconData.weatherIcons(61466);
-  static const IconData wi_owm_522 = const FlutterIconData.weatherIcons(61466);
-  static const IconData wi_owm_531 = const FlutterIconData.weatherIcons(61469);
-  static const IconData wi_owm_600 = const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_owm_601 = const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_owm_602 = const FlutterIconData.weatherIcons(61621);
-  static const IconData wi_owm_611 = const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_owm_612 = const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_owm_615 = const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_owm_616 = const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_owm_620 = const FlutterIconData.weatherIcons(61463);
-  static const IconData wi_owm_621 = const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_owm_622 = const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_owm_701 = const FlutterIconData.weatherIcons(61466);
-  static const IconData wi_owm_711 = const FlutterIconData.weatherIcons(61538);
-  static const IconData wi_owm_721 = const FlutterIconData.weatherIcons(61622);
-  static const IconData wi_owm_731 = const FlutterIconData.weatherIcons(61539);
-  static const IconData wi_owm_741 = const FlutterIconData.weatherIcons(61460);
-  static const IconData wi_owm_761 = const FlutterIconData.weatherIcons(61539);
-  static const IconData wi_owm_762 = const FlutterIconData.weatherIcons(61539);
-  static const IconData wi_owm_771 = const FlutterIconData.weatherIcons(61457);
-  static const IconData wi_owm_781 = const FlutterIconData.weatherIcons(61526);
-  static const IconData wi_owm_800 = const FlutterIconData.weatherIcons(61453);
-  static const IconData wi_owm_801 = const FlutterIconData.weatherIcons(61457);
-  static const IconData wi_owm_802 = const FlutterIconData.weatherIcons(61457);
-  static const IconData wi_owm_803 = const FlutterIconData.weatherIcons(61458);
-  static const IconData wi_owm_804 = const FlutterIconData.weatherIcons(61459);
-  static const IconData wi_owm_900 = const FlutterIconData.weatherIcons(61526);
-  static const IconData wi_owm_901 = const FlutterIconData.weatherIcons(61469);
-  static const IconData wi_owm_902 = const FlutterIconData.weatherIcons(61555);
-  static const IconData wi_owm_903 = const FlutterIconData.weatherIcons(61558);
-  static const IconData wi_owm_904 = const FlutterIconData.weatherIcons(61554);
-  static const IconData wi_owm_905 = const FlutterIconData.weatherIcons(61473);
-  static const IconData wi_owm_906 = const FlutterIconData.weatherIcons(61461);
-  static const IconData wi_owm_957 = const FlutterIconData.weatherIcons(61520);
-  static const IconData wi_owm_day_200 =
-      const FlutterIconData.weatherIcons(61456);
-  static const IconData wi_owm_day_201 =
-      const FlutterIconData.weatherIcons(61456);
-  static const IconData wi_owm_day_202 =
-      const FlutterIconData.weatherIcons(61456);
-  static const IconData wi_owm_day_210 =
-      const FlutterIconData.weatherIcons(61445);
-  static const IconData wi_owm_day_211 =
-      const FlutterIconData.weatherIcons(61445);
-  static const IconData wi_owm_day_212 =
-      const FlutterIconData.weatherIcons(61445);
-  static const IconData wi_owm_day_221 =
-      const FlutterIconData.weatherIcons(61445);
-  static const IconData wi_owm_day_230 =
-      const FlutterIconData.weatherIcons(61456);
-  static const IconData wi_owm_day_231 =
-      const FlutterIconData.weatherIcons(61456);
-  static const IconData wi_owm_day_232 =
-      const FlutterIconData.weatherIcons(61456);
-  static const IconData wi_owm_day_300 =
-      const FlutterIconData.weatherIcons(61451);
-  static const IconData wi_owm_day_301 =
-      const FlutterIconData.weatherIcons(61451);
-  static const IconData wi_owm_day_302 =
-      const FlutterIconData.weatherIcons(61448);
-  static const IconData wi_owm_day_310 =
-      const FlutterIconData.weatherIcons(61448);
-  static const IconData wi_owm_day_311 =
-      const FlutterIconData.weatherIcons(61448);
-  static const IconData wi_owm_day_312 =
-      const FlutterIconData.weatherIcons(61448);
-  static const IconData wi_owm_day_313 =
-      const FlutterIconData.weatherIcons(61448);
-  static const IconData wi_owm_day_314 =
-      const FlutterIconData.weatherIcons(61448);
-  static const IconData wi_owm_day_321 =
-      const FlutterIconData.weatherIcons(61451);
-  static const IconData wi_owm_day_500 =
-      const FlutterIconData.weatherIcons(61451);
-  static const IconData wi_owm_day_501 =
-      const FlutterIconData.weatherIcons(61448);
-  static const IconData wi_owm_day_502 =
-      const FlutterIconData.weatherIcons(61448);
-  static const IconData wi_owm_day_503 =
-      const FlutterIconData.weatherIcons(61448);
-  static const IconData wi_owm_day_504 =
-      const FlutterIconData.weatherIcons(61448);
-  static const IconData wi_owm_day_511 =
-      const FlutterIconData.weatherIcons(61446);
-  static const IconData wi_owm_day_520 =
-      const FlutterIconData.weatherIcons(61449);
-  static const IconData wi_owm_day_521 =
-      const FlutterIconData.weatherIcons(61449);
-  static const IconData wi_owm_day_522 =
-      const FlutterIconData.weatherIcons(61449);
-  static const IconData wi_owm_day_531 =
-      const FlutterIconData.weatherIcons(61454);
-  static const IconData wi_owm_day_600 =
-      const FlutterIconData.weatherIcons(61450);
-  static const IconData wi_owm_day_601 =
-      const FlutterIconData.weatherIcons(61618);
-  static const IconData wi_owm_day_602 =
-      const FlutterIconData.weatherIcons(61450);
-  static const IconData wi_owm_day_611 =
-      const FlutterIconData.weatherIcons(61446);
-  static const IconData wi_owm_day_612 =
-      const FlutterIconData.weatherIcons(61446);
-  static const IconData wi_owm_day_615 =
-      const FlutterIconData.weatherIcons(61446);
-  static const IconData wi_owm_day_616 =
-      const FlutterIconData.weatherIcons(61446);
-  static const IconData wi_owm_day_620 =
-      const FlutterIconData.weatherIcons(61446);
-  static const IconData wi_owm_day_621 =
-      const FlutterIconData.weatherIcons(61450);
-  static const IconData wi_owm_day_622 =
-      const FlutterIconData.weatherIcons(61450);
-  static const IconData wi_owm_day_701 =
-      const FlutterIconData.weatherIcons(61449);
-  static const IconData wi_owm_day_711 =
-      const FlutterIconData.weatherIcons(61538);
-  static const IconData wi_owm_day_721 =
-      const FlutterIconData.weatherIcons(61622);
-  static const IconData wi_owm_day_731 =
-      const FlutterIconData.weatherIcons(61539);
-  static const IconData wi_owm_day_741 =
-      const FlutterIconData.weatherIcons(61443);
-  static const IconData wi_owm_day_761 =
-      const FlutterIconData.weatherIcons(61539);
-  static const IconData wi_owm_day_762 =
-      const FlutterIconData.weatherIcons(61539);
-  static const IconData wi_owm_day_771 =
-      const FlutterIconData.weatherIcons(61440);
-  static const IconData wi_owm_day_781 =
-      const FlutterIconData.weatherIcons(61526);
-  static const IconData wi_owm_day_800 =
-      const FlutterIconData.weatherIcons(61453);
-  static const IconData wi_owm_day_801 =
-      const FlutterIconData.weatherIcons(61440);
-  static const IconData wi_owm_day_802 =
-      const FlutterIconData.weatherIcons(61440);
-  static const IconData wi_owm_day_803 =
-      const FlutterIconData.weatherIcons(61440);
-  static const IconData wi_owm_day_804 =
-      const FlutterIconData.weatherIcons(61452);
-  static const IconData wi_owm_day_900 =
-      const FlutterIconData.weatherIcons(61526);
-  static const IconData wi_owm_day_901 =
-      const FlutterIconData.weatherIcons(61454);
-  static const IconData wi_owm_day_902 =
-      const FlutterIconData.weatherIcons(61555);
-  static const IconData wi_owm_day_903 =
-      const FlutterIconData.weatherIcons(61558);
-  static const IconData wi_owm_day_904 =
-      const FlutterIconData.weatherIcons(61554);
-  static const IconData wi_owm_day_905 =
-      const FlutterIconData.weatherIcons(61636);
-  static const IconData wi_owm_day_906 =
-      const FlutterIconData.weatherIcons(61444);
-  static const IconData wi_owm_day_957 =
-      const FlutterIconData.weatherIcons(61520);
-  static const IconData wi_owm_night_200 =
-      const FlutterIconData.weatherIcons(61485);
-  static const IconData wi_owm_night_201 =
-      const FlutterIconData.weatherIcons(61485);
-  static const IconData wi_owm_night_202 =
-      const FlutterIconData.weatherIcons(61485);
-  static const IconData wi_owm_night_210 =
-      const FlutterIconData.weatherIcons(61477);
-  static const IconData wi_owm_night_211 =
-      const FlutterIconData.weatherIcons(61477);
-  static const IconData wi_owm_night_212 =
-      const FlutterIconData.weatherIcons(61477);
-  static const IconData wi_owm_night_221 =
-      const FlutterIconData.weatherIcons(61477);
-  static const IconData wi_owm_night_230 =
-      const FlutterIconData.weatherIcons(61485);
-  static const IconData wi_owm_night_231 =
-      const FlutterIconData.weatherIcons(61485);
-  static const IconData wi_owm_night_232 =
-      const FlutterIconData.weatherIcons(61485);
-  static const IconData wi_owm_night_300 =
-      const FlutterIconData.weatherIcons(61483);
-  static const IconData wi_owm_night_301 =
-      const FlutterIconData.weatherIcons(61483);
-  static const IconData wi_owm_night_302 =
-      const FlutterIconData.weatherIcons(61480);
-  static const IconData wi_owm_night_310 =
-      const FlutterIconData.weatherIcons(61480);
-  static const IconData wi_owm_night_311 =
-      const FlutterIconData.weatherIcons(61480);
-  static const IconData wi_owm_night_312 =
-      const FlutterIconData.weatherIcons(61480);
-  static const IconData wi_owm_night_313 =
-      const FlutterIconData.weatherIcons(61480);
-  static const IconData wi_owm_night_314 =
-      const FlutterIconData.weatherIcons(61480);
-  static const IconData wi_owm_night_321 =
-      const FlutterIconData.weatherIcons(61483);
-  static const IconData wi_owm_night_500 =
-      const FlutterIconData.weatherIcons(61483);
-  static const IconData wi_owm_night_501 =
-      const FlutterIconData.weatherIcons(61480);
-  static const IconData wi_owm_night_502 =
-      const FlutterIconData.weatherIcons(61480);
-  static const IconData wi_owm_night_503 =
-      const FlutterIconData.weatherIcons(61480);
-  static const IconData wi_owm_night_504 =
-      const FlutterIconData.weatherIcons(61480);
-  static const IconData wi_owm_night_511 =
-      const FlutterIconData.weatherIcons(61478);
-  static const IconData wi_owm_night_520 =
-      const FlutterIconData.weatherIcons(61481);
-  static const IconData wi_owm_night_521 =
-      const FlutterIconData.weatherIcons(61481);
-  static const IconData wi_owm_night_522 =
-      const FlutterIconData.weatherIcons(61481);
-  static const IconData wi_owm_night_531 =
-      const FlutterIconData.weatherIcons(61484);
-  static const IconData wi_owm_night_600 =
-      const FlutterIconData.weatherIcons(61482);
-  static const IconData wi_owm_night_601 =
-      const FlutterIconData.weatherIcons(61620);
-  static const IconData wi_owm_night_602 =
-      const FlutterIconData.weatherIcons(61482);
-  static const IconData wi_owm_night_611 =
-      const FlutterIconData.weatherIcons(61478);
-  static const IconData wi_owm_night_612 =
-      const FlutterIconData.weatherIcons(61478);
-  static const IconData wi_owm_night_615 =
-      const FlutterIconData.weatherIcons(61478);
-  static const IconData wi_owm_night_616 =
-      const FlutterIconData.weatherIcons(61478);
-  static const IconData wi_owm_night_620 =
-      const FlutterIconData.weatherIcons(61478);
-  static const IconData wi_owm_night_621 =
-      const FlutterIconData.weatherIcons(61482);
-  static const IconData wi_owm_night_622 =
-      const FlutterIconData.weatherIcons(61482);
-  static const IconData wi_owm_night_701 =
-      const FlutterIconData.weatherIcons(61481);
-  static const IconData wi_owm_night_711 =
-      const FlutterIconData.weatherIcons(61538);
-  static const IconData wi_owm_night_721 =
-      const FlutterIconData.weatherIcons(61622);
-  static const IconData wi_owm_night_731 =
-      const FlutterIconData.weatherIcons(61539);
-  static const IconData wi_owm_night_741 =
-      const FlutterIconData.weatherIcons(61514);
-  static const IconData wi_owm_night_761 =
-      const FlutterIconData.weatherIcons(61539);
-  static const IconData wi_owm_night_762 =
-      const FlutterIconData.weatherIcons(61539);
-  static const IconData wi_owm_night_771 =
-      const FlutterIconData.weatherIcons(61474);
-  static const IconData wi_owm_night_781 =
-      const FlutterIconData.weatherIcons(61526);
-  static const IconData wi_owm_night_800 =
-      const FlutterIconData.weatherIcons(61486);
-  static const IconData wi_owm_night_801 =
-      const FlutterIconData.weatherIcons(61474);
-  static const IconData wi_owm_night_802 =
-      const FlutterIconData.weatherIcons(61474);
-  static const IconData wi_owm_night_803 =
-      const FlutterIconData.weatherIcons(61474);
-  static const IconData wi_owm_night_804 =
-      const FlutterIconData.weatherIcons(61574);
-  static const IconData wi_owm_night_900 =
-      const FlutterIconData.weatherIcons(61526);
-  static const IconData wi_owm_night_901 =
-      const FlutterIconData.weatherIcons(61498);
-  static const IconData wi_owm_night_902 =
-      const FlutterIconData.weatherIcons(61555);
-  static const IconData wi_owm_night_903 =
-      const FlutterIconData.weatherIcons(61558);
-  static const IconData wi_owm_night_904 =
-      const FlutterIconData.weatherIcons(61554);
-  static const IconData wi_owm_night_905 =
-      const FlutterIconData.weatherIcons(61473);
-  static const IconData wi_owm_night_906 =
-      const FlutterIconData.weatherIcons(61476);
-  static const IconData wi_owm_night_957 =
-      const FlutterIconData.weatherIcons(61520);
+      FlutterIconData.weatherIcons(61526);
+  static const IconData wi_wmo4680_0 = FlutterIconData.weatherIcons(61525);
+  static const IconData wi_wmo4680_00 = FlutterIconData.weatherIcons(61525);
+  static const IconData wi_wmo4680_1 = FlutterIconData.weatherIcons(61459);
+  static const IconData wi_wmo4680_01 = FlutterIconData.weatherIcons(61459);
+  static const IconData wi_wmo4680_2 = FlutterIconData.weatherIcons(61525);
+  static const IconData wi_wmo4680_02 = FlutterIconData.weatherIcons(61525);
+  static const IconData wi_wmo4680_3 = FlutterIconData.weatherIcons(61459);
+  static const IconData wi_wmo4680_03 = FlutterIconData.weatherIcons(61459);
+  static const IconData wi_wmo4680_4 = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_wmo4680_04 = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_wmo4680_5 = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_wmo4680_05 = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_wmo4680_10 = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_wmo4680_11 = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_wmo4680_12 = FlutterIconData.weatherIcons(61462);
+  static const IconData wi_wmo4680_18 = FlutterIconData.weatherIcons(61520);
+  static const IconData wi_wmo4680_20 = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_wmo4680_21 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_wmo4680_22 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_wmo4680_23 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_wmo4680_24 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_wmo4680_25 = FlutterIconData.weatherIcons(61461);
+  static const IconData wi_wmo4680_26 = FlutterIconData.weatherIcons(61470);
+  static const IconData wi_wmo4680_27 = FlutterIconData.weatherIcons(61539);
+  static const IconData wi_wmo4680_28 = FlutterIconData.weatherIcons(61539);
+  static const IconData wi_wmo4680_29 = FlutterIconData.weatherIcons(61539);
+  static const IconData wi_wmo4680_30 = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_wmo4680_31 = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_wmo4680_32 = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_wmo4680_33 = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_wmo4680_34 = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_wmo4680_35 = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_wmo4680_40 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_wmo4680_41 = FlutterIconData.weatherIcons(61468);
+  static const IconData wi_wmo4680_42 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_wmo4680_43 = FlutterIconData.weatherIcons(61468);
+  static const IconData wi_wmo4680_44 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_wmo4680_45 = FlutterIconData.weatherIcons(61461);
+  static const IconData wi_wmo4680_46 = FlutterIconData.weatherIcons(61461);
+  static const IconData wi_wmo4680_47 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_wmo4680_48 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_wmo4680_50 = FlutterIconData.weatherIcons(61468);
+  static const IconData wi_wmo4680_51 = FlutterIconData.weatherIcons(61468);
+  static const IconData wi_wmo4680_52 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_wmo4680_53 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_wmo4680_54 = FlutterIconData.weatherIcons(61558);
+  static const IconData wi_wmo4680_55 = FlutterIconData.weatherIcons(61558);
+  static const IconData wi_wmo4680_56 = FlutterIconData.weatherIcons(61558);
+  static const IconData wi_wmo4680_57 = FlutterIconData.weatherIcons(61468);
+  static const IconData wi_wmo4680_58 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_wmo4680_60 = FlutterIconData.weatherIcons(61468);
+  static const IconData wi_wmo4680_61 = FlutterIconData.weatherIcons(61468);
+  static const IconData wi_wmo4680_62 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_wmo4680_63 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_wmo4680_64 = FlutterIconData.weatherIcons(61461);
+  static const IconData wi_wmo4680_65 = FlutterIconData.weatherIcons(61461);
+  static const IconData wi_wmo4680_66 = FlutterIconData.weatherIcons(61461);
+  static const IconData wi_wmo4680_67 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_wmo4680_68 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_wmo4680_70 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_wmo4680_71 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_wmo4680_72 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_wmo4680_73 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_wmo4680_74 = FlutterIconData.weatherIcons(61558);
+  static const IconData wi_wmo4680_75 = FlutterIconData.weatherIcons(61558);
+  static const IconData wi_wmo4680_76 = FlutterIconData.weatherIcons(61558);
+  static const IconData wi_wmo4680_77 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_wmo4680_78 = FlutterIconData.weatherIcons(61558);
+  static const IconData wi_wmo4680_80 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_wmo4680_81 = FlutterIconData.weatherIcons(61468);
+  static const IconData wi_wmo4680_82 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_wmo4680_83 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_wmo4680_84 = FlutterIconData.weatherIcons(61469);
+  static const IconData wi_wmo4680_85 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_wmo4680_86 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_wmo4680_87 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_wmo4680_89 = FlutterIconData.weatherIcons(61461);
+  static const IconData wi_wmo4680_90 = FlutterIconData.weatherIcons(61462);
+  static const IconData wi_wmo4680_91 = FlutterIconData.weatherIcons(61469);
+  static const IconData wi_wmo4680_92 = FlutterIconData.weatherIcons(61470);
+  static const IconData wi_wmo4680_93 = FlutterIconData.weatherIcons(61470);
+  static const IconData wi_wmo4680_94 = FlutterIconData.weatherIcons(61462);
+  static const IconData wi_wmo4680_95 = FlutterIconData.weatherIcons(61470);
+  static const IconData wi_wmo4680_96 = FlutterIconData.weatherIcons(61470);
+  static const IconData wi_wmo4680_99 = FlutterIconData.weatherIcons(61526);
+  static const IconData wi_owm_200 = FlutterIconData.weatherIcons(61470);
+  static const IconData wi_owm_201 = FlutterIconData.weatherIcons(61470);
+  static const IconData wi_owm_202 = FlutterIconData.weatherIcons(61470);
+  static const IconData wi_owm_210 = FlutterIconData.weatherIcons(61462);
+  static const IconData wi_owm_211 = FlutterIconData.weatherIcons(61462);
+  static const IconData wi_owm_212 = FlutterIconData.weatherIcons(61462);
+  static const IconData wi_owm_221 = FlutterIconData.weatherIcons(61462);
+  static const IconData wi_owm_230 = FlutterIconData.weatherIcons(61470);
+  static const IconData wi_owm_231 = FlutterIconData.weatherIcons(61470);
+  static const IconData wi_owm_232 = FlutterIconData.weatherIcons(61470);
+  static const IconData wi_owm_300 = FlutterIconData.weatherIcons(61468);
+  static const IconData wi_owm_301 = FlutterIconData.weatherIcons(61468);
+  static const IconData wi_owm_302 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_owm_310 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_owm_311 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_owm_312 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_owm_313 = FlutterIconData.weatherIcons(61466);
+  static const IconData wi_owm_314 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_owm_321 = FlutterIconData.weatherIcons(61468);
+  static const IconData wi_owm_500 = FlutterIconData.weatherIcons(61468);
+  static const IconData wi_owm_501 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_owm_502 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_owm_503 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_owm_504 = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_owm_511 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_owm_520 = FlutterIconData.weatherIcons(61466);
+  static const IconData wi_owm_521 = FlutterIconData.weatherIcons(61466);
+  static const IconData wi_owm_522 = FlutterIconData.weatherIcons(61466);
+  static const IconData wi_owm_531 = FlutterIconData.weatherIcons(61469);
+  static const IconData wi_owm_600 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_owm_601 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_owm_602 = FlutterIconData.weatherIcons(61621);
+  static const IconData wi_owm_611 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_owm_612 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_owm_615 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_owm_616 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_owm_620 = FlutterIconData.weatherIcons(61463);
+  static const IconData wi_owm_621 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_owm_622 = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_owm_701 = FlutterIconData.weatherIcons(61466);
+  static const IconData wi_owm_711 = FlutterIconData.weatherIcons(61538);
+  static const IconData wi_owm_721 = FlutterIconData.weatherIcons(61622);
+  static const IconData wi_owm_731 = FlutterIconData.weatherIcons(61539);
+  static const IconData wi_owm_741 = FlutterIconData.weatherIcons(61460);
+  static const IconData wi_owm_761 = FlutterIconData.weatherIcons(61539);
+  static const IconData wi_owm_762 = FlutterIconData.weatherIcons(61539);
+  static const IconData wi_owm_771 = FlutterIconData.weatherIcons(61457);
+  static const IconData wi_owm_781 = FlutterIconData.weatherIcons(61526);
+  static const IconData wi_owm_800 = FlutterIconData.weatherIcons(61453);
+  static const IconData wi_owm_801 = FlutterIconData.weatherIcons(61457);
+  static const IconData wi_owm_802 = FlutterIconData.weatherIcons(61457);
+  static const IconData wi_owm_803 = FlutterIconData.weatherIcons(61458);
+  static const IconData wi_owm_804 = FlutterIconData.weatherIcons(61459);
+  static const IconData wi_owm_900 = FlutterIconData.weatherIcons(61526);
+  static const IconData wi_owm_901 = FlutterIconData.weatherIcons(61469);
+  static const IconData wi_owm_902 = FlutterIconData.weatherIcons(61555);
+  static const IconData wi_owm_903 = FlutterIconData.weatherIcons(61558);
+  static const IconData wi_owm_904 = FlutterIconData.weatherIcons(61554);
+  static const IconData wi_owm_905 = FlutterIconData.weatherIcons(61473);
+  static const IconData wi_owm_906 = FlutterIconData.weatherIcons(61461);
+  static const IconData wi_owm_957 = FlutterIconData.weatherIcons(61520);
+  static const IconData wi_owm_day_200 = FlutterIconData.weatherIcons(61456);
+  static const IconData wi_owm_day_201 = FlutterIconData.weatherIcons(61456);
+  static const IconData wi_owm_day_202 = FlutterIconData.weatherIcons(61456);
+  static const IconData wi_owm_day_210 = FlutterIconData.weatherIcons(61445);
+  static const IconData wi_owm_day_211 = FlutterIconData.weatherIcons(61445);
+  static const IconData wi_owm_day_212 = FlutterIconData.weatherIcons(61445);
+  static const IconData wi_owm_day_221 = FlutterIconData.weatherIcons(61445);
+  static const IconData wi_owm_day_230 = FlutterIconData.weatherIcons(61456);
+  static const IconData wi_owm_day_231 = FlutterIconData.weatherIcons(61456);
+  static const IconData wi_owm_day_232 = FlutterIconData.weatherIcons(61456);
+  static const IconData wi_owm_day_300 = FlutterIconData.weatherIcons(61451);
+  static const IconData wi_owm_day_301 = FlutterIconData.weatherIcons(61451);
+  static const IconData wi_owm_day_302 = FlutterIconData.weatherIcons(61448);
+  static const IconData wi_owm_day_310 = FlutterIconData.weatherIcons(61448);
+  static const IconData wi_owm_day_311 = FlutterIconData.weatherIcons(61448);
+  static const IconData wi_owm_day_312 = FlutterIconData.weatherIcons(61448);
+  static const IconData wi_owm_day_313 = FlutterIconData.weatherIcons(61448);
+  static const IconData wi_owm_day_314 = FlutterIconData.weatherIcons(61448);
+  static const IconData wi_owm_day_321 = FlutterIconData.weatherIcons(61451);
+  static const IconData wi_owm_day_500 = FlutterIconData.weatherIcons(61451);
+  static const IconData wi_owm_day_501 = FlutterIconData.weatherIcons(61448);
+  static const IconData wi_owm_day_502 = FlutterIconData.weatherIcons(61448);
+  static const IconData wi_owm_day_503 = FlutterIconData.weatherIcons(61448);
+  static const IconData wi_owm_day_504 = FlutterIconData.weatherIcons(61448);
+  static const IconData wi_owm_day_511 = FlutterIconData.weatherIcons(61446);
+  static const IconData wi_owm_day_520 = FlutterIconData.weatherIcons(61449);
+  static const IconData wi_owm_day_521 = FlutterIconData.weatherIcons(61449);
+  static const IconData wi_owm_day_522 = FlutterIconData.weatherIcons(61449);
+  static const IconData wi_owm_day_531 = FlutterIconData.weatherIcons(61454);
+  static const IconData wi_owm_day_600 = FlutterIconData.weatherIcons(61450);
+  static const IconData wi_owm_day_601 = FlutterIconData.weatherIcons(61618);
+  static const IconData wi_owm_day_602 = FlutterIconData.weatherIcons(61450);
+  static const IconData wi_owm_day_611 = FlutterIconData.weatherIcons(61446);
+  static const IconData wi_owm_day_612 = FlutterIconData.weatherIcons(61446);
+  static const IconData wi_owm_day_615 = FlutterIconData.weatherIcons(61446);
+  static const IconData wi_owm_day_616 = FlutterIconData.weatherIcons(61446);
+  static const IconData wi_owm_day_620 = FlutterIconData.weatherIcons(61446);
+  static const IconData wi_owm_day_621 = FlutterIconData.weatherIcons(61450);
+  static const IconData wi_owm_day_622 = FlutterIconData.weatherIcons(61450);
+  static const IconData wi_owm_day_701 = FlutterIconData.weatherIcons(61449);
+  static const IconData wi_owm_day_711 = FlutterIconData.weatherIcons(61538);
+  static const IconData wi_owm_day_721 = FlutterIconData.weatherIcons(61622);
+  static const IconData wi_owm_day_731 = FlutterIconData.weatherIcons(61539);
+  static const IconData wi_owm_day_741 = FlutterIconData.weatherIcons(61443);
+  static const IconData wi_owm_day_761 = FlutterIconData.weatherIcons(61539);
+  static const IconData wi_owm_day_762 = FlutterIconData.weatherIcons(61539);
+  static const IconData wi_owm_day_771 = FlutterIconData.weatherIcons(61440);
+  static const IconData wi_owm_day_781 = FlutterIconData.weatherIcons(61526);
+  static const IconData wi_owm_day_800 = FlutterIconData.weatherIcons(61453);
+  static const IconData wi_owm_day_801 = FlutterIconData.weatherIcons(61440);
+  static const IconData wi_owm_day_802 = FlutterIconData.weatherIcons(61440);
+  static const IconData wi_owm_day_803 = FlutterIconData.weatherIcons(61440);
+  static const IconData wi_owm_day_804 = FlutterIconData.weatherIcons(61452);
+  static const IconData wi_owm_day_900 = FlutterIconData.weatherIcons(61526);
+  static const IconData wi_owm_day_901 = FlutterIconData.weatherIcons(61454);
+  static const IconData wi_owm_day_902 = FlutterIconData.weatherIcons(61555);
+  static const IconData wi_owm_day_903 = FlutterIconData.weatherIcons(61558);
+  static const IconData wi_owm_day_904 = FlutterIconData.weatherIcons(61554);
+  static const IconData wi_owm_day_905 = FlutterIconData.weatherIcons(61636);
+  static const IconData wi_owm_day_906 = FlutterIconData.weatherIcons(61444);
+  static const IconData wi_owm_day_957 = FlutterIconData.weatherIcons(61520);
+  static const IconData wi_owm_night_200 = FlutterIconData.weatherIcons(61485);
+  static const IconData wi_owm_night_201 = FlutterIconData.weatherIcons(61485);
+  static const IconData wi_owm_night_202 = FlutterIconData.weatherIcons(61485);
+  static const IconData wi_owm_night_210 = FlutterIconData.weatherIcons(61477);
+  static const IconData wi_owm_night_211 = FlutterIconData.weatherIcons(61477);
+  static const IconData wi_owm_night_212 = FlutterIconData.weatherIcons(61477);
+  static const IconData wi_owm_night_221 = FlutterIconData.weatherIcons(61477);
+  static const IconData wi_owm_night_230 = FlutterIconData.weatherIcons(61485);
+  static const IconData wi_owm_night_231 = FlutterIconData.weatherIcons(61485);
+  static const IconData wi_owm_night_232 = FlutterIconData.weatherIcons(61485);
+  static const IconData wi_owm_night_300 = FlutterIconData.weatherIcons(61483);
+  static const IconData wi_owm_night_301 = FlutterIconData.weatherIcons(61483);
+  static const IconData wi_owm_night_302 = FlutterIconData.weatherIcons(61480);
+  static const IconData wi_owm_night_310 = FlutterIconData.weatherIcons(61480);
+  static const IconData wi_owm_night_311 = FlutterIconData.weatherIcons(61480);
+  static const IconData wi_owm_night_312 = FlutterIconData.weatherIcons(61480);
+  static const IconData wi_owm_night_313 = FlutterIconData.weatherIcons(61480);
+  static const IconData wi_owm_night_314 = FlutterIconData.weatherIcons(61480);
+  static const IconData wi_owm_night_321 = FlutterIconData.weatherIcons(61483);
+  static const IconData wi_owm_night_500 = FlutterIconData.weatherIcons(61483);
+  static const IconData wi_owm_night_501 = FlutterIconData.weatherIcons(61480);
+  static const IconData wi_owm_night_502 = FlutterIconData.weatherIcons(61480);
+  static const IconData wi_owm_night_503 = FlutterIconData.weatherIcons(61480);
+  static const IconData wi_owm_night_504 = FlutterIconData.weatherIcons(61480);
+  static const IconData wi_owm_night_511 = FlutterIconData.weatherIcons(61478);
+  static const IconData wi_owm_night_520 = FlutterIconData.weatherIcons(61481);
+  static const IconData wi_owm_night_521 = FlutterIconData.weatherIcons(61481);
+  static const IconData wi_owm_night_522 = FlutterIconData.weatherIcons(61481);
+  static const IconData wi_owm_night_531 = FlutterIconData.weatherIcons(61484);
+  static const IconData wi_owm_night_600 = FlutterIconData.weatherIcons(61482);
+  static const IconData wi_owm_night_601 = FlutterIconData.weatherIcons(61620);
+  static const IconData wi_owm_night_602 = FlutterIconData.weatherIcons(61482);
+  static const IconData wi_owm_night_611 = FlutterIconData.weatherIcons(61478);
+  static const IconData wi_owm_night_612 = FlutterIconData.weatherIcons(61478);
+  static const IconData wi_owm_night_615 = FlutterIconData.weatherIcons(61478);
+  static const IconData wi_owm_night_616 = FlutterIconData.weatherIcons(61478);
+  static const IconData wi_owm_night_620 = FlutterIconData.weatherIcons(61478);
+  static const IconData wi_owm_night_621 = FlutterIconData.weatherIcons(61482);
+  static const IconData wi_owm_night_622 = FlutterIconData.weatherIcons(61482);
+  static const IconData wi_owm_night_701 = FlutterIconData.weatherIcons(61481);
+  static const IconData wi_owm_night_711 = FlutterIconData.weatherIcons(61538);
+  static const IconData wi_owm_night_721 = FlutterIconData.weatherIcons(61622);
+  static const IconData wi_owm_night_731 = FlutterIconData.weatherIcons(61539);
+  static const IconData wi_owm_night_741 = FlutterIconData.weatherIcons(61514);
+  static const IconData wi_owm_night_761 = FlutterIconData.weatherIcons(61539);
+  static const IconData wi_owm_night_762 = FlutterIconData.weatherIcons(61539);
+  static const IconData wi_owm_night_771 = FlutterIconData.weatherIcons(61474);
+  static const IconData wi_owm_night_781 = FlutterIconData.weatherIcons(61526);
+  static const IconData wi_owm_night_800 = FlutterIconData.weatherIcons(61486);
+  static const IconData wi_owm_night_801 = FlutterIconData.weatherIcons(61474);
+  static const IconData wi_owm_night_802 = FlutterIconData.weatherIcons(61474);
+  static const IconData wi_owm_night_803 = FlutterIconData.weatherIcons(61474);
+  static const IconData wi_owm_night_804 = FlutterIconData.weatherIcons(61574);
+  static const IconData wi_owm_night_900 = FlutterIconData.weatherIcons(61526);
+  static const IconData wi_owm_night_901 = FlutterIconData.weatherIcons(61498);
+  static const IconData wi_owm_night_902 = FlutterIconData.weatherIcons(61555);
+  static const IconData wi_owm_night_903 = FlutterIconData.weatherIcons(61558);
+  static const IconData wi_owm_night_904 = FlutterIconData.weatherIcons(61554);
+  static const IconData wi_owm_night_905 = FlutterIconData.weatherIcons(61473);
+  static const IconData wi_owm_night_906 = FlutterIconData.weatherIcons(61476);
+  static const IconData wi_owm_night_957 = FlutterIconData.weatherIcons(61520);
   static const IconData wi_wu_chanceflurries =
-      const FlutterIconData.weatherIcons(61540);
-  static const IconData wi_wu_chancerain =
-      const FlutterIconData.weatherIcons(61465);
-  static const IconData wi_wu_chancesleat =
-      const FlutterIconData.weatherIcons(61621);
-  static const IconData wi_wu_chancesnow =
-      const FlutterIconData.weatherIcons(61467);
+      FlutterIconData.weatherIcons(61540);
+  static const IconData wi_wu_chancerain = FlutterIconData.weatherIcons(61465);
+  static const IconData wi_wu_chancesleat = FlutterIconData.weatherIcons(61621);
+  static const IconData wi_wu_chancesnow = FlutterIconData.weatherIcons(61467);
   static const IconData wi_wu_chancetstorms =
-      const FlutterIconData.weatherIcons(61470);
-  static const IconData wi_wu_clear = const FlutterIconData.weatherIcons(61453);
-  static const IconData wi_wu_cloudy =
-      const FlutterIconData.weatherIcons(61442);
-  static const IconData wi_wu_flurries =
-      const FlutterIconData.weatherIcons(61540);
-  static const IconData wi_wu_hazy = const FlutterIconData.weatherIcons(61622);
+      FlutterIconData.weatherIcons(61470);
+  static const IconData wi_wu_clear = FlutterIconData.weatherIcons(61453);
+  static const IconData wi_wu_cloudy = FlutterIconData.weatherIcons(61442);
+  static const IconData wi_wu_flurries = FlutterIconData.weatherIcons(61540);
+  static const IconData wi_wu_hazy = FlutterIconData.weatherIcons(61622);
   static const IconData wi_wu_mostlycloudy =
-      const FlutterIconData.weatherIcons(61442);
-  static const IconData wi_wu_mostlysunny =
-      const FlutterIconData.weatherIcons(61453);
+      FlutterIconData.weatherIcons(61442);
+  static const IconData wi_wu_mostlysunny = FlutterIconData.weatherIcons(61453);
   static const IconData wi_wu_partlycloudy =
-      const FlutterIconData.weatherIcons(61442);
-  static const IconData wi_wu_partlysunny =
-      const FlutterIconData.weatherIcons(61453);
-  static const IconData wi_wu_rain = const FlutterIconData.weatherIcons(61466);
-  static const IconData wi_wu_sleat = const FlutterIconData.weatherIcons(61621);
-  static const IconData wi_wu_snow = const FlutterIconData.weatherIcons(61467);
-  static const IconData wi_wu_sunny = const FlutterIconData.weatherIcons(61453);
-  static const IconData wi_wu_tstorms =
-      const FlutterIconData.weatherIcons(61470);
-  static const IconData wi_wu_unknown =
-      const FlutterIconData.weatherIcons(61453);
+      FlutterIconData.weatherIcons(61442);
+  static const IconData wi_wu_partlysunny = FlutterIconData.weatherIcons(61453);
+  static const IconData wi_wu_rain = FlutterIconData.weatherIcons(61466);
+  static const IconData wi_wu_sleat = FlutterIconData.weatherIcons(61621);
+  static const IconData wi_wu_snow = FlutterIconData.weatherIcons(61467);
+  static const IconData wi_wu_sunny = FlutterIconData.weatherIcons(61453);
+  static const IconData wi_wu_tstorms = FlutterIconData.weatherIcons(61470);
+  static const IconData wi_wu_unknown = FlutterIconData.weatherIcons(61453);
 }
