@@ -6,104 +6,105 @@ import 'flutter_icon_data.dart';
 /// http://zocial.smcllns.com/
 class Zocial {
   Zocial._();
-  static const IconData acrobat = const FlutterIconData.zocial(61696);
-  static const IconData amazon = const FlutterIconData.zocial(61697);
-  static const IconData android = const FlutterIconData.zocial(61698);
-  static const IconData angellist = const FlutterIconData.zocial(61699);
-  static const IconData aol = const FlutterIconData.zocial(61700);
-  static const IconData appnet = const FlutterIconData.zocial(61701);
-  static const IconData appstore = const FlutterIconData.zocial(61702);
-  static const IconData bitbucket = const FlutterIconData.zocial(61703);
-  static const IconData bitcoin = const FlutterIconData.zocial(61704);
-  static const IconData blogger = const FlutterIconData.zocial(61705);
-  static const IconData buffer = const FlutterIconData.zocial(61706);
-  static const IconData cal = const FlutterIconData.zocial(61707);
-  static const IconData call = const FlutterIconData.zocial(61708);
-  static const IconData cart = const FlutterIconData.zocial(61709);
-  static const IconData chrome = const FlutterIconData.zocial(61710);
-  static const IconData cloudapp = const FlutterIconData.zocial(61711);
-  static const IconData creativecommons = const FlutterIconData.zocial(61712);
-  static const IconData delicious = const FlutterIconData.zocial(61713);
-  static const IconData digg = const FlutterIconData.zocial(61714);
-  static const IconData disqus = const FlutterIconData.zocial(61715);
-  static const IconData dribbble = const FlutterIconData.zocial(61716);
-  static const IconData dropbox = const FlutterIconData.zocial(61717);
-  static const IconData drupal = const FlutterIconData.zocial(61718);
-  static const IconData dwolla = const FlutterIconData.zocial(61720);
-  static const IconData email = const FlutterIconData.zocial(61721);
-  static const IconData eventasaurus = const FlutterIconData.zocial(61722);
-  static const IconData eventbrite = const FlutterIconData.zocial(61723);
-  static const IconData eventful = const FlutterIconData.zocial(61724);
-  static const IconData evernote = const FlutterIconData.zocial(61725);
-  static const IconData facebook = const FlutterIconData.zocial(61726);
-  static const IconData fivehundredpx = const FlutterIconData.zocial(61727);
-  static const IconData flattr = const FlutterIconData.zocial(61728);
-  static const IconData flickr = const FlutterIconData.zocial(61729);
-  static const IconData forrst = const FlutterIconData.zocial(61730);
-  static const IconData foursquare = const FlutterIconData.zocial(61731);
-  static const IconData github = const FlutterIconData.zocial(61732);
-  static const IconData gmail = const FlutterIconData.zocial(61733);
-  static const IconData google = const FlutterIconData.zocial(61734);
-  static const IconData googleplay = const FlutterIconData.zocial(61735);
-  static const IconData googleplus = const FlutterIconData.zocial(61736);
-  static const IconData gowalla = const FlutterIconData.zocial(61737);
-  static const IconData grooveshark = const FlutterIconData.zocial(61738);
-  static const IconData guest = const FlutterIconData.zocial(61739);
-  static const IconData html5 = const FlutterIconData.zocial(61740);
-  static const IconData ie = const FlutterIconData.zocial(61741);
-  static const IconData instagram = const FlutterIconData.zocial(61742);
-  static const IconData instapaper = const FlutterIconData.zocial(61743);
-  static const IconData intensedebate = const FlutterIconData.zocial(61744);
-  static const IconData itunes = const FlutterIconData.zocial(61745);
-  static const IconData klout = const FlutterIconData.zocial(61746);
-  static const IconData lanyrd = const FlutterIconData.zocial(61747);
-  static const IconData lastfm = const FlutterIconData.zocial(61748);
-  static const IconData lego = const FlutterIconData.zocial(61749);
-  static const IconData linkedin = const FlutterIconData.zocial(61750);
-  static const IconData lkdto = const FlutterIconData.zocial(61751);
-  static const IconData logmein = const FlutterIconData.zocial(61752);
-  static const IconData macstore = const FlutterIconData.zocial(61753);
-  static const IconData meetup = const FlutterIconData.zocial(61754);
-  static const IconData myspace = const FlutterIconData.zocial(61755);
-  static const IconData ninetyninedesigns = const FlutterIconData.zocial(61756);
-  static const IconData openid = const FlutterIconData.zocial(61757);
-  static const IconData opentable = const FlutterIconData.zocial(61758);
-  static const IconData paypal = const FlutterIconData.zocial(61759);
-  static const IconData persona = const FlutterIconData.zocial(61796);
-  static const IconData pinboard = const FlutterIconData.zocial(61760);
-  static const IconData pinterest = const FlutterIconData.zocial(61761);
-  static const IconData plancast = const FlutterIconData.zocial(61762);
-  static const IconData plurk = const FlutterIconData.zocial(61763);
-  static const IconData pocket = const FlutterIconData.zocial(61764);
-  static const IconData podcast = const FlutterIconData.zocial(61765);
-  static const IconData posterous = const FlutterIconData.zocial(61766);
-  static const IconData print = const FlutterIconData.zocial(61767);
-  static const IconData quora = const FlutterIconData.zocial(61768);
-  static const IconData reddit = const FlutterIconData.zocial(61769);
-  static const IconData rss = const FlutterIconData.zocial(61770);
-  static const IconData scribd = const FlutterIconData.zocial(61771);
-  static const IconData skype = const FlutterIconData.zocial(61772);
-  static const IconData smashing = const FlutterIconData.zocial(61773);
-  static const IconData songkick = const FlutterIconData.zocial(61774);
-  static const IconData soundcloud = const FlutterIconData.zocial(61775);
-  static const IconData spotify = const FlutterIconData.zocial(61776);
-  static const IconData stackoverflow = const FlutterIconData.zocial(61777);
-  static const IconData statusnet = const FlutterIconData.zocial(61778);
-  static const IconData steam = const FlutterIconData.zocial(61779);
-  static const IconData stripe = const FlutterIconData.zocial(61780);
-  static const IconData stumbleupon = const FlutterIconData.zocial(61781);
-  static const IconData tumblr = const FlutterIconData.zocial(61782);
-  static const IconData twitter = const FlutterIconData.zocial(61783);
-  static const IconData viadeo = const FlutterIconData.zocial(61784);
-  static const IconData vimeo = const FlutterIconData.zocial(61785);
-  static const IconData vk = const FlutterIconData.zocial(61786);
-  static const IconData weibo = const FlutterIconData.zocial(61787);
-  static const IconData wikipedia = const FlutterIconData.zocial(61788);
-  static const IconData windows = const FlutterIconData.zocial(61789);
-  static const IconData wordpress = const FlutterIconData.zocial(61790);
-  static const IconData xing = const FlutterIconData.zocial(61791);
-  static const IconData yahoo = const FlutterIconData.zocial(61792);
-  static const IconData ycombinator = const FlutterIconData.zocial(61793);
-  static const IconData yelp = const FlutterIconData.zocial(61794);
-  static const IconData youtube = const FlutterIconData.zocial(61795);
+
+  static const IconData acrobat = FlutterIconData.zocial(61696);
+  static const IconData amazon = FlutterIconData.zocial(61697);
+  static const IconData android = FlutterIconData.zocial(61698);
+  static const IconData angellist = FlutterIconData.zocial(61699);
+  static const IconData aol = FlutterIconData.zocial(61700);
+  static const IconData appnet = FlutterIconData.zocial(61701);
+  static const IconData appstore = FlutterIconData.zocial(61702);
+  static const IconData bitbucket = FlutterIconData.zocial(61703);
+  static const IconData bitcoin = FlutterIconData.zocial(61704);
+  static const IconData blogger = FlutterIconData.zocial(61705);
+  static const IconData buffer = FlutterIconData.zocial(61706);
+  static const IconData cal = FlutterIconData.zocial(61707);
+  static const IconData call = FlutterIconData.zocial(61708);
+  static const IconData cart = FlutterIconData.zocial(61709);
+  static const IconData chrome = FlutterIconData.zocial(61710);
+  static const IconData cloudapp = FlutterIconData.zocial(61711);
+  static const IconData creativecommons = FlutterIconData.zocial(61712);
+  static const IconData delicious = FlutterIconData.zocial(61713);
+  static const IconData digg = FlutterIconData.zocial(61714);
+  static const IconData disqus = FlutterIconData.zocial(61715);
+  static const IconData dribbble = FlutterIconData.zocial(61716);
+  static const IconData dropbox = FlutterIconData.zocial(61717);
+  static const IconData drupal = FlutterIconData.zocial(61718);
+  static const IconData dwolla = FlutterIconData.zocial(61720);
+  static const IconData email = FlutterIconData.zocial(61721);
+  static const IconData eventasaurus = FlutterIconData.zocial(61722);
+  static const IconData eventbrite = FlutterIconData.zocial(61723);
+  static const IconData eventful = FlutterIconData.zocial(61724);
+  static const IconData evernote = FlutterIconData.zocial(61725);
+  static const IconData facebook = FlutterIconData.zocial(61726);
+  static const IconData fivehundredpx = FlutterIconData.zocial(61727);
+  static const IconData flattr = FlutterIconData.zocial(61728);
+  static const IconData flickr = FlutterIconData.zocial(61729);
+  static const IconData forrst = FlutterIconData.zocial(61730);
+  static const IconData foursquare = FlutterIconData.zocial(61731);
+  static const IconData github = FlutterIconData.zocial(61732);
+  static const IconData gmail = FlutterIconData.zocial(61733);
+  static const IconData google = FlutterIconData.zocial(61734);
+  static const IconData googleplay = FlutterIconData.zocial(61735);
+  static const IconData googleplus = FlutterIconData.zocial(61736);
+  static const IconData gowalla = FlutterIconData.zocial(61737);
+  static const IconData grooveshark = FlutterIconData.zocial(61738);
+  static const IconData guest = FlutterIconData.zocial(61739);
+  static const IconData html5 = FlutterIconData.zocial(61740);
+  static const IconData ie = FlutterIconData.zocial(61741);
+  static const IconData instagram = FlutterIconData.zocial(61742);
+  static const IconData instapaper = FlutterIconData.zocial(61743);
+  static const IconData intensedebate = FlutterIconData.zocial(61744);
+  static const IconData itunes = FlutterIconData.zocial(61745);
+  static const IconData klout = FlutterIconData.zocial(61746);
+  static const IconData lanyrd = FlutterIconData.zocial(61747);
+  static const IconData lastfm = FlutterIconData.zocial(61748);
+  static const IconData lego = FlutterIconData.zocial(61749);
+  static const IconData linkedin = FlutterIconData.zocial(61750);
+  static const IconData lkdto = FlutterIconData.zocial(61751);
+  static const IconData logmein = FlutterIconData.zocial(61752);
+  static const IconData macstore = FlutterIconData.zocial(61753);
+  static const IconData meetup = FlutterIconData.zocial(61754);
+  static const IconData myspace = FlutterIconData.zocial(61755);
+  static const IconData ninetyninedesigns = FlutterIconData.zocial(61756);
+  static const IconData openid = FlutterIconData.zocial(61757);
+  static const IconData opentable = FlutterIconData.zocial(61758);
+  static const IconData paypal = FlutterIconData.zocial(61759);
+  static const IconData persona = FlutterIconData.zocial(61796);
+  static const IconData pinboard = FlutterIconData.zocial(61760);
+  static const IconData pinterest = FlutterIconData.zocial(61761);
+  static const IconData plancast = FlutterIconData.zocial(61762);
+  static const IconData plurk = FlutterIconData.zocial(61763);
+  static const IconData pocket = FlutterIconData.zocial(61764);
+  static const IconData podcast = FlutterIconData.zocial(61765);
+  static const IconData posterous = FlutterIconData.zocial(61766);
+  static const IconData print = FlutterIconData.zocial(61767);
+  static const IconData quora = FlutterIconData.zocial(61768);
+  static const IconData reddit = FlutterIconData.zocial(61769);
+  static const IconData rss = FlutterIconData.zocial(61770);
+  static const IconData scribd = FlutterIconData.zocial(61771);
+  static const IconData skype = FlutterIconData.zocial(61772);
+  static const IconData smashing = FlutterIconData.zocial(61773);
+  static const IconData songkick = FlutterIconData.zocial(61774);
+  static const IconData soundcloud = FlutterIconData.zocial(61775);
+  static const IconData spotify = FlutterIconData.zocial(61776);
+  static const IconData stackoverflow = FlutterIconData.zocial(61777);
+  static const IconData statusnet = FlutterIconData.zocial(61778);
+  static const IconData steam = FlutterIconData.zocial(61779);
+  static const IconData stripe = FlutterIconData.zocial(61780);
+  static const IconData stumbleupon = FlutterIconData.zocial(61781);
+  static const IconData tumblr = FlutterIconData.zocial(61782);
+  static const IconData twitter = FlutterIconData.zocial(61783);
+  static const IconData viadeo = FlutterIconData.zocial(61784);
+  static const IconData vimeo = FlutterIconData.zocial(61785);
+  static const IconData vk = FlutterIconData.zocial(61786);
+  static const IconData weibo = FlutterIconData.zocial(61787);
+  static const IconData wikipedia = FlutterIconData.zocial(61788);
+  static const IconData windows = FlutterIconData.zocial(61789);
+  static const IconData wordpress = FlutterIconData.zocial(61790);
+  static const IconData xing = FlutterIconData.zocial(61791);
+  static const IconData yahoo = FlutterIconData.zocial(61792);
+  static const IconData ycombinator = FlutterIconData.zocial(61793);
+  static const IconData yelp = FlutterIconData.zocial(61794);
+  static const IconData youtube = FlutterIconData.zocial(61795);
 }
